### 2025-10-06 #9150 @kyleniemeyer
All the preliminary checks pass; however, I notice that the paper does not have any references, which is a bit unusual. I do see mentions of other software packages/projects that should include citations, at minimum.
