### 2025-10-07 d7ffc1b
choice between direct solver and iterativ solver in modal analysis
