### 2025-10-07 c7c5466
condensed stiffness matrix in solveDisplacement
