### 2025-10-08 25924e6
choice between condensed and non-condensed solution method in solveDisplacement (with or without elastic support)
