### 2025-10-09 5f20d7e
use cholesky instead of lu
