### 2025-10-10 7fd4616
mute gmsh at precompile to avoid pipe_write error in IJulia
