+++ examples/multithread-test/test-1.jl
@@ -8,12 +8,9 @@ gmsh.open("cube.geo")
 mat = material("body")
 prob = Problem([mat])
 
-@time K = LowLevelFEM.stiffnessMatrixSolidOld(prob)
-@time K = LowLevelFEM.stiffnessMatrixSolid(prob)
-@time K = LowLevelFEM.stiffnessMatrixSolidParallel(prob)
+@time K = stiffnessMatrix(prob)
 
-@btime K = LowLevelFEM.stiffnessMatrixSolidOld(prob)
-@btime K = LowLevelFEM.stiffnessMatrixSolid(prob)
-@btime K = LowLevelFEM.stiffnessMatrixSolidParallel(prob)
+@btime K = stiffnessMatrix(prob)
+@btime K = stiffnessMatrix(prob, forceOneThread=true)
 
 gmsh.finalize()
