+++ Project.toml
@@ -16,10 +16,10 @@ gmsh_jll = "630162c2-fc9b-58b3-9910-8442a8a132e6"
 [compat]
 Arpack = "0.5"
 JLD2 = "0.6"
-LinearAlgebra = "1.11.0"
+LinearAlgebra = "1.10.0, 1.11.0, 1.12.0"
 Polyester = "0.7.18"
 PrecompileTools = "1.2.1"
-SparseArrays = "1.11.0"
+SparseArrays = "1.10.0, 1.11.0, 1.12.0"
 StaticArrays = "1.9.14"
 gmsh_jll = "4.13"
 julia = "1"
