### 2025-10-10 f4235a9
stiffnessMatrix -> forceOneThead=true by default
for stability
