+++ v1.10.4/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"Operators/#Operators-and-Algebra","page":"Operators","title":"Operators and Algebra","text":"","category":"section"},{"location":"Operators/#Base.:*-Tuple{Number, ScalarField}","page":"Operators","title":"Base.:*","text":"*(b::Number, A::ScalarField)\n\nPerforms multiplication of a ScalarField objects and a Number.\n\nReturn: ScalarField\n\nExamples\n\nC = 2.0 * A\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:*-Tuple{ScalarField, Number}","page":"Operators","title":"Base.:*","text":"*(A::ScalarField, b::Number)\n\nPerforms multiplication of a ScalarField objects and a Number.\n\nReturn: ScalarField\n\nExamples\n\nC = A * 2.0\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:*-Tuple{ScalarField, ScalarField}","page":"Operators","title":"Base.:*","text":"*(A::ScalarField, B::ScalarField)\n\nPerforms element-wise multiplication of two ScalarField objects on the same set of elements.\n\nReturns: ScalarField\n\nExamples\n\nC = A * B\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:*-Tuple{ScalarField, VectorField}","page":"Operators","title":"Base.:*","text":"*(A::ScalarField, B::VectorField)\n\nScales a VectorField by a ScalarField element-wise on matching elements.\n\nReturns: VectorField\n\nExamples\n\nv2 = s .* v  # equivalent to s * v\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:*-Tuple{TensorField, TensorField}","page":"Operators","title":"Base.:*","text":"*(A::TensorField, B::TensorField)\n\nTensor contraction (matrix multiplication) for each element/node: reshapes 9×1 blocks into 3×3, multiplies, then flattens back.\n\nReturns: TensorField\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:*-Tuple{VectorField, ScalarField}","page":"Operators","title":"Base.:*","text":"*(B::VectorField, A::ScalarField)\n\nScales a VectorField by a ScalarField element-wise on matching elements.\n\nReturns: VectorField\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:+-Tuple{ScalarField, ScalarField}","page":"Operators","title":"Base.:+","text":"+(A::ScalarField, B::ScalarField)\n\nPerforms element-wise addition of two ScalarField objects on the same set of elements.\n\nReturns: ScalarField\n\nExamples\n\nC = A + B\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:--Tuple{ScalarField, ScalarField}","page":"Operators","title":"Base.:-","text":"-(A::ScalarField, B::ScalarField)\n\nPerforms element-wise subtraction of two ScalarField objects on the same set of elements.\n\nReturns: ScalarField\n\nExamples\n\nC = A - B\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:/-Tuple{ScalarField, ScalarField}","page":"Operators","title":"Base.:/","text":"/(A::ScalarField, B::ScalarField)\n\nPerforms element-wise division of two ScalarField objects on the same set of elements.\n\nReturns: ScalarField\n\nExamples\n\nC = A / B\n\n\n\n\n\n","category":"method"},{"location":"Operators/#Base.:/-Tuple{VectorFi
[…]
