### 2025-10-11 7423605
corrected paper
