+++ paper/paper.md
@@ -1,49 +1,35 @@
 ---
 title: 'LowLevelFEM.jl: A lightweight finite element toolbox in Julia'
 tags:
-  - Julia
-  - finite element method
-  - structural mechanics
-  - continuum mechanics
-  - open source
-authors:
-  - name: Balázs Pere
-    orcid: 0000-0002-1161-4206
-    affiliation: 1
-affiliations:
-  - name: Department of Applied Mechanics, Széchenyi István University, Győr, Hungary
-    index: 1
-date: 2025-09-07
-bibliography: paper.bib
+
+* Julia
+* finite element method
+* structural mechanics
+* continuum mechanics
+* open source
+  authors:
+* name: Balázs Pere
+  orcid: 0000-0002-1161-4206
+  affiliation: 1
+  affiliations:
+* name: Department of Applied Mechanics, Széchenyi István University, Győr, Hungary
+  index: 1
+  date: 2025-09-07
+  bibliography: paper.bib
+
 ---
 
 # Summary
 
-LowLevelFEM.jl is a finite element method (FEM) toolbox written entirely in the Julia
-programming language. Its design philosophy emphasizes **simplicity, transparency, and
-performance**, making it suitable for both educational and research purposes in mechanics
-and engineering. Unlike large frameworks that rely on domain-specific languages or
-compiled backends, LowLevelFEM provides users with direct access to FEM building blocks
-in Julia, enabling full control over discretization, assembly, and solution steps.
+LowLevelFEM.jl is a finite element method (FEM) [@zienkiewicz2005] toolbox written entirely in the Julia programming language [@bezanson2017]. Its design philosophy emphasizes **simplicity, transparency, and performance**, making it suitable for both educational and research purposes in mechanics and engineering. Unlike large frameworks that rely on domain-specific languages or compiled backends, LowLevelFEM provides users with direct access to FEM building blocks in Julia, enabling full control over discretization, assembly, and solution steps.
 
-The package currently supports two- and three-dimensional solid mechanics problems
-including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimalistic
-design lowers the entry barrier for students while still offering enough flexibility for
-advanced users to **inspect and control algorithms step by step, process intermediate
-results during the solution procedure, and perform operations with scalar, vector, and
-tensor fields**. LowLevelFEM uses **Gmsh** as its pre- and post-processor, ensuring
-compatibility with a widely adopted meshing and visualization tool.
+The package currently supports two- and three-dimensional solid mechanics problems including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimalistic design lowers the entry barrier for students while still offering enough flexibility for advanced users to **inspect and control algorithms step by step, process intermediate results during the solution procedure, and perform operations with scalar, vector, and tensor fields**. LowLevelFEM uses **Gmsh** [@geuzaine2009] as its pre- and post-processor, ensuring compatibility with a widely adopted meshing and visualization tool.
 
-Thanks to Julia’s just-in-time compilation and multiple dispatch, the code remains concise
-while achieving performance comparable to traditional FEM codes in C/C++ or Fortran.
-LowLevelFEM is released under the MIT license and distributed via the Julia General
-registry. Documentation, tutorials, and examples are available at <https://juliahub.com/ui/Packages/General/LowLevelFEM> or
-<https://perebalazs.github.io/LowLevelFEM.jl/stable/>.
[…]
