### 2025-10-11 6e53ca9
Update paper.md
