+++ src/general.jl
@@ -1509,83 +1509,90 @@ openPostProcessor()
 """
 function normalVector(problem::Problem, phName::String)
     gmsh.model.setCurrent(problem.name)
+
     non = problem.non
     ncoord2 = zeros(3 * non)
     numElem = Int[]
-    normalvectors = []
-    
+    normalvectors = Vector{Matrix{Float64}}()
+
+    # Fizikai csoport lekérdezése (pl. felület)
     dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
-    sz = 0
-    for idm in 1:length(dimTags)
-        dimTag = dimTags[idm]
-        edim = dimTag[1]
-        etag = dimTag[2]
-        elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
-        for i in 1:length(elemTypes)
-            sz += length(elemTags[i])
-        end
-    end
-    sizehint!(numElem, sz)
-    sizehint!(normalvectors, sz)
-    
-    for idm in 1:length(dimTags)
-        dimTag = dimTags[idm]
-        edim = dimTag[1]
-        etag = dimTag[2]
+
+    for dimTag in dimTags
+        edim, etag = dimTag
         if edim != 2
             error("normalVector: physical group is not a surface.")
         end
+
         elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
-        nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes(edim, etag, true, false)
-        ncoord2[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
-        ncoord2[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
-        ncoord2[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
+        nodeTags, ncoord, _ = gmsh.model.mesh.getNodes(edim, etag, true, false)
+
+        # Globális csomópontkoordináták betöltése
+        ncoord2[nodeTags*3 .- 2] = ncoord[1:3:end]
+        ncoord2[nodeTags*3 .- 1] = ncoord[2:3:end]
+        ncoord2[nodeTags*3 .- 0] = ncoord[3:3:end]
+
+        # --- Elemciklus ---
         for i in 1:length(elemTypes)
             et = elemTypes[i]
-            elementName, eldim, order, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(et)
+            elementName, eldim, order, numNodes, localNodeCoord, _ =
+                gmsh.model.mesh.getElementProperties(et)
+
+            # Lokális koordináták átalakítása 3D formába
             localNodeCoord2 = zeros(length(localNodeCoord) ÷ 2 * 3)
-            localNodeCoord2[1:3:length(localNodeCoord2)] .= localNodeCoord[1:2:length(localNodeCoord)]
-            localNodeCoord2[2:3:length(localNodeCoord2)] .= localNodeCoord[2:2:length(localNodeCoord)]
-            #comp, fun, ori = gmsh.model.mesh.getBasisFunctions(et, localNodeCoord2, "Lagrange")
-            #h = reshape(fun, :, numNodes)
-            comp3, dfun, ori = gmsh.model.mesh.getBasisFunctions(et, localNodeCoord2, "GradLagrange")
-            ∇h = reshape(dfun, :, numNodes)
+            localNodeCoord2[1:3:end] .= localNodeCoord[1:2:end]
+            localNodeCoord2[2:3:end] .= localNodeCoord[2:2:end]
+
+            # Bázisfüggvények deriváltjai a referenciadoménben
[…]
