### 2025-10-13 d472d2e
Update general.jl
Define scalar, vector and tensor fields in a simpler way.
