### 2025-10-14 d2d83a5
simpler definiton of scalarField, vectorField and tensorField
