+++ src/LowLevelFEM.jl
@@ -25,8 +25,8 @@ include("nonlinear.jl")
 
 @setup_workload begin
     @compile_workload begin
-        gmsh.option.setNumber("General.Terminal", 0)
         gmsh.initialize()
+        gmsh.option.setNumber("General.Terminal", 0)
         
         # nagyon kicsi 3D próba-probléma (pl. egy 1×1×1 kocka)
         gmsh.model.add("cube")
@@ -37,9 +37,13 @@ include("nonlinear.jl")
         gmsh.option.setNumber("Mesh.MeshSizeMax", 1.0)
         gmsh.option.setNumber("Mesh.MeshSizeFromCurvature", 0)
         gmsh.model.occ.synchronize()
+        gmsh.model.mesh.generate(3)
         gmsh.model.addPhysicalGroup(3, [1], 10)
         gmsh.model.setPhysicalName(3, 10, "body")
-        gmsh.model.mesh.generate(3)
+        gmsh.model.addPhysicalGroup(2, [1], 11)
+        gmsh.model.setPhysicalName(2, 11, "left")
+        gmsh.model.addPhysicalGroup(2, [2], 12)
+        gmsh.model.setPhysicalName(2, 12, "right")
         
         mat = material("body")
         prob = Problem([mat])
