### 2025-10-14 99eddd8
gc at the end of stiffnessMatrix
+ total precompile for a simple elastic problem
