+++ Project.toml
@@ -21,7 +21,7 @@ Polyester = "0.7.18"
 PrecompileTools = "1.2.1"
 SparseArrays = "1.10.0, 1.11.0, 1.12.0"
 StaticArrays = "1.9.14"
-gmsh_jll = "4.13"
+gmsh_jll = "4.14"
 julia = "1"
 
 [extras]
