### 2025-10-17 e913b3e
iterative solver
