+++ Project.toml
@@ -1,10 +1,12 @@
 name = "LowLevelFEM"
 uuid = "6171b9fb-adbf-4751-adb9-5faded75de07"
-authors = ["Balázs Pere"]
 version = "1.10.7"
+authors = ["Balázs Pere"]
 
 [deps]
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
+IncompleteLU = "40713840-3770-5561-ab4c-a76e7d0d7895"
+IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
@@ -15,6 +17,8 @@ gmsh_jll = "630162c2-fc9b-58b3-9910-8442a8a132e6"
 
 [compat]
 Arpack = "0.5"
+IncompleteLU = "0.2.1"
+IterativeSolvers = "0.9.4"
 JLD2 = "0.6"
 LinearAlgebra = "1.10.0, 1.11.0, 1.12.0"
 Polyester = "0.7.18"
