### 2025-10-19 fc667a3
better precompile
more options for cg solver
