### 2025-10-20 b839e31
new CI + test
