+++ .github/workflows/CI.yml
@@ -8,35 +8,64 @@ on:
   pull_request:
 
 jobs:
-  build:
+  test:
+    name: Run tests (Julia 1.12)
     runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
 
-      - uses: julia-actions/setup-julia@v2
+      - name: Set up Julia
+        uses: julia-actions/setup-julia@v2
         with:
-          version: '1.11'
+          version: '1.12'
 
-      #- name: Install LaTeX (TeX Live)
-        #run: sudo apt-get update && sudo apt-get install -y texlive-latex-base texlive-latex-extra texlive-fonts-recommended texlive-fonts-extra texlive-xetex
+      - name: Instantiate project
+        run: julia --project -e 'using Pkg; Pkg.instantiate()'
 
-      - name: Install Julia dependencies
-        run: julia --project=docs -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+      - name: Run tests with coverage
+        env:
+          JULIA_NUM_THREADS: 2
+        run: |
+          julia --project -e '
+            using Pkg;
+            Pkg.test(; coverage=true);
+          '
 
-      - name: Build HTML and LaTeX docs
+      # --- Generate coverage report ---
+      - uses: julia-actions/julia-processcoverage@v1
+
+      # --- Upload to Codecov ---
+      - uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: lcov.info
+          fail_ci_if_error: true
+          flags: unittests
+          verbose: true
+
+  docs:
+    name: Build & deploy docs
+    runs-on: ubuntu-latest
+    needs: test
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Julia
+        uses: julia-actions/setup-julia@v2
+        with:
+          version: '1.12'
+
+      - name: Install documentation dependencies
+        run: |
+          julia --project=docs -e '
+            using Pkg;
+            Pkg.develop(PackageSpec(path=pwd()));
+            Pkg.instantiate();
+          '
+
+      - name: Build docs (HTML + LaTeX)
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
         run: julia --project=docs docs/make.jl
-
-      #- name: Compile PDF from LaTeX
-        #run: |
-          #cd docs/build/latex
-          #pdflatex LowLevelFEM.tex
-          #pdflatex LowLevelFEM.tex
-
-      #- name: Copy PDF to build folder
-        #run: cp docs/build/latex/LowLevelFEM.pdf docs/build/LowLevelFEM.pdf
-
-      # A deploydocs() Documenter-en keresztül fut, nem kell külön lépés
