+++ README.md
@@ -1,6 +1,11 @@
 [![Build Status](https://github.com/perebalazs/LowLevelFEM.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/perebalazs/LowLevelFEM.jl/actions/workflows/CI.yml?query=branch%3Amain)
 [![](https://img.shields.io/badge/docs-dev-blue.svg)](https://perebalazs.github.io/LowLevelFEM.jl/dev)
 [![](https://img.shields.io/badge/docs-stable-blue.svg)](https://perebalazs.github.io/LowLevelFEM.jl/stable)
+[![codecov](https://codecov.io/gh/perebalazs/LowLevelFEM.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/perebalazs/LowLevelFEM.jl)
+[![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)](https://opensource.org/licenses/MIT)
+<!-- [![DOI](https://joss.theoj.org/papers/10.21105/joss.XXXXX/status.svg)](https://doi.org/10.21105/joss.XXXXX) -->
+<!-- [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.YYYYYYY.svg)](https://doi.org/10.5281/zenodo.YYYYYYY) -->
+
 
 # LowLevelFEM
 
