### 2025-10-20 3af0f94
update CI.yml
