### 2025-10-20 4bfaf5a
Update CI.yml
