+++ src/extra.jl
@@ -1,0 +1,66 @@
+function showGapThickness(phName; name=phName, visible=false)
+    SS = gmsh.view.add(name)
+    dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+    nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes(-1, -1, true, false)
+    ncoord2 = similar(ncoord)
+    ncoord2[nodeTags*3 .- 2] .= ncoord[1:3:length(ncoord)]
+    ncoord2[nodeTags*3 .- 1] .= ncoord[2:3:length(ncoord)]
+    ncoord2[nodeTags*3 .- 0] .= ncoord[3:3:length(ncoord)]
+    ret2 = []
+    ret3 = []
+    ret4 = []
+    el2 = 0
+    el3 = 0
+    el4 = 0
+    for idm in 1:length(dimTags)
+        dimTag = dimTags[idm]
+        edim = dimTag[1]
+        etag = dimTag[2]
+        elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+        for i in 1:length(elemTypes)
+            et = elemTypes[i]
+            elementName, dim, order, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(et)
+            nn = zeros(numPrimaryNodes * 4)
+            for j in 1:length(elemTags[i])
+                elem = elemTags[i][j]
+                for l in 1:numPrimaryNodes
+                    for k in 1:2
+                        nn[k*numPrimaryNodes-numPrimaryNodes + l] = ncoord2[elemNodeTags[i][j*numNodes-(numNodes-l)]*3-(3-k)]
+                    end
+                    k = 3
+                    nn[k*numPrimaryNodes-numPrimaryNodes + l] = 0
+                    k = 4
+                    k2 = 3
+                    nn[4*numPrimaryNodes-numPrimaryNodes + l] = ncoord2[elemNodeTags[i][j*numNodes-(numNodes-l)]*3-(3-k2)]
+                end
+                if numPrimaryNodes == 3
+                append!(ret3,nn)
+                elseif numPrimaryNodes == 4
+                    append!(ret4,nn)
+                elseif numPrimaryNodes == 2
+                    append!(ret2,nn)
+                end
+            end
+            if numPrimaryNodes == 3
+                el3 += length(elemTags[i])
+            elseif numPrimaryNodes == 4
+                el4 += length(elemTags[i])
+            elseif numPrimaryNodes == 2
+                el2 += length(elemTags[i])
+            end
+        end
+    end
+    if el3 > 0
+        gmsh.view.addListData(SS, "ST", el3, ret3)
+    end
+    if el4 > 0
+        gmsh.view.addListData(SS, "SQ", el4, ret4)
+    end
+    if el2 > 0
+        gmsh.view.addListData(SS, "SL", el2, ret2)
+    end
+    if visible == false
+        gmsh.view.option.setNumber(SS, "Visible", 0)
+    end
+    return SS
+end
