### 2025-10-21 077e5d0
debug tools
