+++ src/extra.jl
@@ -1,3 +1,6 @@
+export systemMatrix
+
+
 function showGapThickness(phName; name=phName, visible=false)
     SS = gmsh.view.add(name)
     dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
@@ -63,4 +66,174 @@ function showGapThickness(phName; name=phName, visible=false)
         gmsh.view.option.setNumber(SS, "Visible", 0)
     end
     return SS
+end
+
+function pressureConstraint(name; p=1im)
+    bc0 = name, p
+    return bc0
+end
+
+function flowRate(name; q=0)
+    ld0 = name, q
+    return ld0
+end
+
+function systemMatrix(problem, αInNodes::ScalarField, velocity)
+    gmsh.model.setCurrent(problem.name)
+    if problem.type ≠ :Reynolds && problem.material[1].type ≠ :JFO
+        error("systemMatrix: only Reynolds equation with JFO cavitation model is implemented.")
+    end
+    if length(problem.material) ≠ 1
+        error("systemMatrix: only one material is permitted.")
+    end
+    elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(problem.dim, -1)
+    lengthOfIJV = sum([(div(length(elemNodeTags[i]), length(elemTags[i])) * problem.dim)^2 * length(elemTags[i]) for i in 1:length(elemTags)])
+    nn = []
+    I = []
+    J = []
+    V = []
+    V2 = []
+    V = convert(Vector{Float64}, V)
+    V2 = convert(Vector{Float64}, V2)
+    sizehint!(I, lengthOfIJV)
+    sizehint!(J, lengthOfIJV)
+    sizehint!(V, lengthOfIJV)
+    sizehint!(V2, lengthOfIJV)
+    ncoord2 = zeros(3 * problem.non)
+    phName = problem.material[1].phName
+    nameGap = problem.geometry.nameGap
+    η = problem.material[1].η
+    κ = problem.material[1].κ
+
+    for ipg in 1:1#length(problem.material)
+        #η = problem.material.η
+        dim = problem.dim
+        pdim = problem.pdim
+        if dim == 2
+            rowsOfB = 2
+        elseif dim == 1
+            rowsOfB = 1
+        else
+            error("systemMatrix: rowsOfB = $rowsOfB")
+        end
+
+        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+        for idm in 1:length(dimTags)
+            dimTag = dimTags[idm]
+            edim = dimTag[1]
+            etag = dimTag[2]
+            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+            tag = getTagForPhysicalName(phName)
+            nodeTags, ncoord = gmsh.model.mesh.getNodesForPhysicalGroup(dim, tag)
+            ncoord2[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
+            ncoord2[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
+            ncoord2[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
+            for i in 1:length(elemTypes)
+                et = elemTypes[i]
+                elementName, dim, order, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(et)
+                intPoints, intWeights = gmsh.model.mesh.getIntegrationPoints(et, "Gauss" * string(2order + 1))
+                numIntPoints = length(intWeights)
+                comp, dfun, ori = gmsh.model.mesh.getBasisFunctions(et, intPoints, "GradLagrange")
+                ∇h = reshape(dfun, :, numIntPoints)
[…]
