### 2025-10-21 1db0990
Update operators.jl
cuberoot...
