+++ src/extra.jl
@@ -78,7 +78,13 @@ function flowRate(name; q=0)
     return ld0
 end
 
-function systemMatrix(problem, αInNodes::ScalarField, velocity)
+function filmThickness(problem::Problem)
+    fh(x, y, z) = gmsh.view.probe(problem.geometry.tagTop, x, y, 0)[1][1] - z
+    return scalarField(problem, problem.material[1].phName, fh)
+end
+
+#function systemMatrix(problem, αInNodes::ScalarField, velocity::Number, height::ScalarField)
+function systemMatrix(problem, velocity::Number, height::ScalarField)
     gmsh.model.setCurrent(problem.name)
     if problem.type ≠ :Reynolds && problem.material[1].type ≠ :JFO
         error("systemMatrix: only Reynolds equation with JFO cavitation model is implemented.")
@@ -86,6 +92,12 @@ function systemMatrix(problem, αInNodes::ScalarField, velocity)
     if length(problem.material) ≠ 1
         error("systemMatrix: only one material is permitted.")
     end
+
+    ex = VectorField(problem, problem.material[1].phName, [1,0,0])
+    gradh = ∇(height)
+    dhdx = gradh * ex
+    dhdx = elementsToNodes(dhdx)
+
     elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(problem.dim, -1)
     lengthOfIJV = sum([(div(length(elemNodeTags[i]), length(elemTags[i])) * problem.dim)^2 * length(elemTags[i]) for i in 1:length(elemTags)])
     nn = []
@@ -158,7 +170,9 @@ function systemMatrix(problem, αInNodes::ScalarField, velocity)
                 end
                 x = zeros(numIntPoints)
                 y = dim == 2 ? zeros(numIntPoints) : 0
-                α = zeros(numIntPoints)
+                #α = zeros(numIntPoints)
+                hh = zeros(numIntPoints)
+                dhhdx = zeros(numIntPoints)
                 for j in 1:length(elemTags[i])
                     elem = elemTags[i][j]
                     for k in 1:numNodes
@@ -172,7 +186,9 @@ function systemMatrix(problem, αInNodes::ScalarField, velocity)
                         if dim == 2
                             y[k] = h[:, k]' * ncoord2[nnet[j, :] * 3 .- 1]
                         end
-                        α[k] = h[:, k]' * αInNodes.a[nnet[j, :]]
+                        #α[k] = h[:, k]' * αInNodes.a[nnet[j, :]]
+                        hh[k] = h[:, k]' * height.a[nnet[j, :]]
+                        dhhdx[k] = h[:, k]' * dhdx.a[nnet[j, :]]
                     end
                     ∂h .*= 0
                     for k in 1:numIntPoints, l in 1:numNodes
@@ -193,29 +209,29 @@ function systemMatrix(problem, αInNodes::ScalarField, velocity)
                     end
                     K1 .*= 0
                     K2 .*= 0
-                    hh = 0
-                    dhdx = 0
+                    #hh = 0
+                    #dhdx = 0
                     for k in 1:numIntPoints
                         #g = α[k] < 1.0 ? 0.0 : 1.0
-                        if dim == 1
-                            hh = gmsh.view.probe(problem.geometry.tagTop, x[k], 0, 0, -1, -1, false, -1)[1][1]
-                            hhB = gmsh.view.probe(problem.geometry.tagBottom, x[k], 0, 0, -1, -1, false, -1)[1][1]
-                            hh -= hhB
-                            dhdx = gmsh.view.probe(problem.geometry.tagTop, x[k], 0, 0,-1,-1,true, -1)[1][1]
-                        elseif dim == 2
[…]
