+++ docs/src/index.md
@@ -16,7 +16,7 @@ LowLevelFEM is a Julia package for finite element analysis with an engineering-f
 - Matrices: stiffness `K`, mass `M` (lumped or consistent), proportional damping `C` (Rayleigh/Caughey), heat conduction/capacity, latent heat, and convection matrices/vectors.
 - Loads and constraints: nodal and distributed loads on physical groups; function-based loads and temperature BCs; elastic supports; initial displacement/velocity/temperature.
 - Thermal–structural coupling: thermal expansion, thermal stresses, and heat generated by elastic deformation.
-- Solvers: static and transient dynamics (central difference and HHT-α from the Newmark family).
+- Solvers: static (direct and iterative solvers with arbitrary preconditioners) and transient dynamics (central difference and HHT-α from the Newmark family).
 - Eigenproblems: modal analysis (frequencies and mode shapes, optionally prestressed) and linear buckling (critical factors and modes).
 - Field operators and results: gradient/divergence/curl; stress/strain and heat flux as element or nodal fields; smoothing at nodes with field jumps; user-defined scalar/vector/tensor FE fields.
 - Visualization and plots: Gmsh-based views for displacements, stresses, strains, heat flux, with animation for dynamics; plot results along user-defined paths; show results on surfaces.
