+++ src/extra.jl
@@ -1,4 +1,5 @@
-export systemMatrix
+export pressureConstraint, flowRate
+export solvePressure
 
 
 function showGapThickness(phName; name=phName, visible=false)
@@ -68,12 +69,12 @@ function showGapThickness(phName; name=phName, visible=false)
     return SS
 end
 
-function pressureConstraint(name; p=1im)
-    bc0 = name, p
+function pressureConstraint(name::String; p=1im)
+    bc0 = name, p, 1im, 1im
     return bc0
 end
 
-function flowRate(name; q=0)
+function flowRate(name::String; q=0)
     ld0 = name, q
     return ld0
 end
@@ -260,13 +261,120 @@ function systemMatrix(problem, velocity::Number)
     GC.gc()
 end
 
+function flowRateVector(problem, loads)
+    gmsh.model.setCurrent(problem.name)
+    dim0 = problem.dim
+    pdim = problem.pdim
+    non = problem.non
+    dof = pdim * non
+    #η = problem.material[1].η
+    fp = zeros(dof)
+    ncoord2 = zeros(3 * problem.non)
+    for n in 1:length(loads)
+        name, ff = loads[n]
+        f = ff
+        dimTags = gmsh.model.getEntitiesForPhysicalName(name)
+        for i ∈ 1:length(dimTags)
+            dimTag = dimTags[i]
+            dim = dimTag[1]
+            tag = dimTag[2]
+            elementTypes, elementTags, elemNodeTags = gmsh.model.mesh.getElements(dim, tag)
+            nodeTags::Vector{Int64}, ncoord, parametricCoord = gmsh.model.mesh.getNodes(dim, tag, true, false)
+            ncoord2[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
+            ncoord2[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
+            ncoord2[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
+            for ii in 1:length(elementTypes)
+                elementName, dim, order, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(elementTypes[ii])
+                nnoe = reshape(elemNodeTags[ii], numNodes, :)'
+                intPoints, intWeights = gmsh.model.mesh.getIntegrationPoints(elementTypes[ii], "Gauss" * string(order+1))
+                numIntPoints = length(intWeights)
+                comp, dfun, ori = gmsh.model.mesh.getBasisFunctions(elementTypes[ii], intPoints, "GradLagrange")
+                ∇h = reshape(dfun, :, numIntPoints)
+                comp, fun, ori = gmsh.model.mesh.getBasisFunctions(elementTypes[ii], intPoints, "Lagrange")
+                h = reshape(fun, :, numIntPoints)
+                nnet = zeros(Int, length(elementTags[i]), numNodes)
+                invJac = zeros(3, 3numIntPoints)
+                H = zeros(pdim * numIntPoints, pdim * numNodes)
+                for j in 1:numIntPoints
+                    for k in 1:numNodes
+                        for l in 1:pdim
+                            H[j*pdim-(pdim-l), k*pdim-(pdim-l)] = h[k, j]
+                        end
+                    end
+                end
+                ∂h = zeros(dim0, numNodes * numIntPoints)
+                f1 = zeros(pdim * numNodes)
+                nn2 = zeros(Int, pdim * numNodes)
+                for l in 1:length(elementTags[ii])
+                    elem = elementTags[ii][l]
+                    for k in 1:numNodes
+                        nnet[l, k] = elemNodeTags[i][(l-1)*numNodes+k]
+                    end
[…]
