### 2025-10-30 ca2a739
bug fix in TensorField constructor
