+++ src/general.jl
@@ -71,7 +71,7 @@ mutable struct Geometry
     dhdx::Any
     hh::Any
     p::Any
-    Geometry() = new()
+    Geometry() = new("", "", 0, 0, nothing, nothing, nothing, nothing)
     function Geometry(nameGap, nameVolume)
         dimTags = gmsh.model.getEntitiesForPhysicalName(nameGap)
         dim = dimTags[1][1]
@@ -330,7 +330,48 @@ struct ScalarField <: AbstractField
     function ScalarField(A0, a0, t0, numElem0, nsteps0, type0, model)
         return new(A0, a0, t0, numElem0, nsteps0, type0, model)
     end
-    function ScalarField(problem, dataField)
+    function ScalarField(problem::Problem, dataField::Vector)
+    gmsh.model.setCurrent(problem.name)
+    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+    type = :scalar
+    nsteps = 1
+    A = Vector{Matrix{Float64}}()
+    numElem = Int[]
+
+    @inbounds for data in dataField
+        phName, f, fx, fy, fz, fxy, fyz, fzx = data
+        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+
+        for (edim, etag) in dimTags
+            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+            @inbounds for (i, et) in enumerate(elemTypes)
+                _, _, _, numNodes, _, _ = gmsh.model.mesh.getElementProperties(et)
+                for (j, elem) in enumerate(elemTags[i])
+                    sc1 = zeros(numNodes, nsteps)
+                    push!(numElem, elem)
+                    if f != :no
+                        for k in 1:numNodes
+                            nodeTag = elemNodeTags[i][(j-1)*numNodes + k]
+                            if f isa Function
+                                x = coords[3*nodeTag - 2]
+                                y = coords[3*nodeTag - 1]
+                                z = coords[3*nodeTag]
+                                sc1[k, 1] = f(x, y, z)
+                            else
+                                sc1[k, 1] = f
+                            end
+                        end
+                    end
+                    push!(A, sc1)
+                end
+            end
+        end
+    end
+
+    return ScalarField(A, [;;], [0.0], numElem, nsteps, type, problem)
+end
+#=
+function ScalarField(problem, dataField)
         if !isa(dataField, Vector)
             error("ScalarField: dataField are not arranged in a vector. Put them in [...]")
         end
@@ -381,6 +422,7 @@ struct ScalarField <: AbstractField
         t = [0.0]
         return new(A, a, t, numElem, nsteps, type, problem)
     end
+    =#
     function ScalarField(problem::Problem, phName::String, data::Union{Number,Function})
         f = field(phName, f=data)
         return ScalarField(problem, [f])
