+++ src/extra.jl
@@ -1,5 +1,5 @@
 export pressureConstraint, flowRate
-export solvePressure
+export solvePressure, pressureInVolume
 
 
 function showGapThickness(phName; name=phName, visible=false)
@@ -90,6 +90,24 @@ function initialize(problem::Problem)
     gmsh.view.remove(problem.geometry.tagTop)
 end
 
+function pressureInVolume(p::ScalarField)
+    if p.model.geometry.nameVolume == ""
+        error("initializePressure: no volume for lubricant has been defined.")
+    end
+    if p.model.geometry.hh == nothing
+        view_h = showDoFResults(p.model.geometry.h)
+        fh(x, y, z) = gmsh.view.probe(view_h, x, y, 0, -1, -1, false, -1)[1][1]
+        p.model.geometry.hh = ScalarField(p.model, [field(p.model.geometry.nameVolume, f=fh)])
+        gmsh.view.remove(view_h)
+    end
+    view_p = showDoFResults(p)
+    fp(x, y, z) = gmsh.view.probe(view_p, x, y, 0, -1, -1, false, -1)[1][1]
+    pp = ScalarField(p.model, [field(p.model.geometry.nameVolume, f=fp)])
+    gmsh.view.remove(view_p)
+    return pp
+end
+
+#=
 function initializePressure(p::ScalarField)
     if p.model.geometry.nameVolume == ""
         error("initializePressure: no volume for lubricant has been defined.")
@@ -105,6 +123,7 @@ function initializePressure(p::ScalarField)
     p.model.geometry.p = ScalarField(p.model, [field(p.model.geometry.nameVolume, f=fp)])
     gmsh.view.remove(view_p)
 end
+=#
 
 #function systemMatrix(problem, αInNodes::ScalarField, velocity::Number, height::ScalarField)
 function systemMatrix(problem, velocity::Number)
@@ -376,6 +395,7 @@ function flowRateVector(problem, loads)
     return ScalarField([], reshape(fp, :,1), [0.0], [], 1, type, problem)
 end
 
+#=
 function solvePressure(problem, load, BC, V; cav=false, periodicSlave="", periodicMaster="")
     if problem.type == :dummy
         return nothing
@@ -534,5 +554,180 @@ function solvePressure(problem, load, BC, V; cav=false, periodicSlave="", period
         #return α, αcav # log.(α), αcav
         pret[fluid] =  p0 .+ κ * g .* log.(α)
         return ScalarField([], reshape(pret, :,1), [0], [], 1, :other, problem), ScalarField([], reshape(αcav, :,1), [0], [], 1, :p, problem)
+    end
+end
+=#
+
+function solvePressure(problem, load, BC, V; cav=false, periodicSlave="", periodicMaster="")
+    if problem.type == :dummy
+        return nothing
+    end
+    if problem.type != :Reynolds
+        error("solvePressure: bad problem type '$type'.")
+    end
+    p0 = problem.material[1].p₀
+    κ = problem.material[1].κ
+    if problem.geometry.h == nothing
+        initialize(problem)
+    end
+    fluid = constrainedDoFs(problem, [pressureConstraint(problem.material[1].phName, p=0)])
+    bc = constrainedDoFs(problem, BC)
+    free = setdiff(fluid, bc)
+    #one = zeros(problem.non * problem.pdim) .+ 1
+    #type = :none
+    #if problem.dim == 1
+    #    type = :scalar
+    #elseif problem.dim == 2
+    #    type = :scalar
+    #else
+    #    error("problem.dim = $(problem.dim)")
+    #end
+    #α = ScalarField([], reshape(one, :,1), [0.0], [], 1, type, problem)
[…]
