+++ src/fieldtools.jl
@@ -1,0 +1,192 @@
+module FieldTools
+
+using ..LowLevelFEM
+
+export probe_field
+
+"""
+    probe_field(rr::ScalarField, x, y, z)
+
+Gyors, tisztán Julia-alapú lekérdezés (Gmsh nélkül), amely visszaadja
+a `ScalarField` értékét az `(x, y, z)` pontban.
+
+Támogatott elemtípusok: `tet4`, `tet10`, `hex8`, `hex20`, `tri3`, `tri6`.
+"""
+function probe_field(rr::ScalarField, x::Float64, y::Float64, z::Float64)
+    gmsh.model.setCurrent(rr.model.name)
+
+    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+    X = coords[1:3:end]; Y = coords[2:3:end]; Z = coords[3:3:end]
+
+    elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(rr.model.dim, -1)
+
+    @inbounds for (i, et) in enumerate(elemTypes)
+        elementName, dim, order, numNodes::Int, localCoord, numPrimary =
+            gmsh.model.mesh.getElementProperties(et)
+
+        for (j, elem) in enumerate(elemTags[i])
+            if !(elem in rr.numElem)
+                continue
+            end
+
+            node_ids = elemNodeTags[i][(j-1)*numNodes+1 : j*numNodes]
+            xnodes = X[node_ids]; ynodes = Y[node_ids]; znodes = Z[node_ids]
+
+            # gyors bbox szűrés
+            if x < minimum(xnodes)-1e-8 || x > maximum(xnodes)+1e-8 ||
+               y < minimum(ynodes)-1e-8 || y > maximum(ynodes)+1e-8 ||
+               z < minimum(znodes)-1e-8 || z > maximum(znodes)+1e-8
+                continue
+            end
+
+            ξ, η, ζ = natural_coordinates_generic(x, y, z, xnodes, ynodes, znodes, et)
+            if isnothing(ξ)
+                continue
+            end
+
+            N = lagrange_shape_generic(ξ, η, ζ, et)
+            if length(N) != numNodes
+                continue
+            end
+
+            ind = findfirst(isequal(elem), rr.numElem)
+            vals = rr.A[ind][:, 1]
+            return dot(N, vals)
+        end
+    end
+    return NaN
+end
+
+
+# ---------------------------------------------------------------------
+# Shape függvények
+# ---------------------------------------------------------------------
+
+function lagrange_shape_generic(ξ, η, ζ, et)
+    if et == 4       # tet4
+        return [1-ξ-η-ζ, ξ, η, ζ]
+
+    elseif et == 5   # hex8
+        return 0.125 .* [
+            (1-ξ)*(1-η)*(1-ζ),
+            (1+ξ)*(1-η)*(1-ζ),
+            (1+ξ)*(1+η)*(1-ζ),
+            (1-ξ)*(1+η)*(1-ζ),
+            (1-ξ)*(1-η)*(1+ζ),
+            (1+ξ)*(1-η)*(1+ζ),
+            (1+ξ)*(1+η)*(1+ζ),
+            (1-ξ)*(1+η)*(1+ζ)
+        ]
+
+    elseif et == 11  # tet10
+        L1 = 1-ξ-η-ζ; L2=ξ; L3=η; L4=ζ
+        return [
[…]

[…]
