+++ src/fieldtools.jl
@@ -1,8 +1,104 @@
 module FieldTools
 
+using LinearAlgebra: dot
 using ..LowLevelFEM
 
-export probe_field
+export probe_field_bulk
+
+"""
+    build_bbox(rr::ScalarField)
+
+Előkészíti az elemek *bounding boxait* gyors térbeli szűréshez.
+Minden elemhez `(xmin, xmax, ymin, ymax, zmin, zmax)` értékeket tárol,
+amelyeket a `probe_field_bulk` használ majd a keresés gyorsítására.
+"""
+function build_bbox(rr::ScalarField)
+    gmsh.model.setCurrent(rr.model.name)
+
+    elem_bboxes = Vector{NTuple{6,Float64}}(undef, length(rr.numElem))
+    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+    X, Y, Z = coords[1:3:end], coords[2:3:end], coords[3:3:end]
+    _, _, dim, _ = gmsh.model.mesh.getElement(rr.numElem[1])
+    elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(dim, -1)
+
+    counter = 1
+    for (i, et) in enumerate(elemTypes)
+        _, dim, order, numNodes, _, _ = gmsh.model.mesh.getElementProperties(et)
+        for j in 1:length(elemTags[i])
+            nodes = elemNodeTags[i][(j-1)*numNodes+1:j*numNodes]
+            xs, ys, zs = X[nodes], Y[nodes], Z[nodes]
+            elem_bboxes[counter] = (
+                minimum(xs), maximum(xs),
+                minimum(ys), maximum(ys),
+                minimum(zs), maximum(zs)
+            )
+            counter += 1
+        end
+    end
+
+    return elem_bboxes
+end
+
+
+
+"""
+    probe_field_bulk(rr::ScalarField, points::Matrix{Float64})
+
+Gyors, tisztán Julia-alapú tömeges lekérdezés (`(x,y,z)` pontok → mezőértékek).
+A `points` mátrix mérete `N×3` legyen (egy sor egy pontot jelent).
+Előfeltétel: a `rr.bbox` mező a `build_bbox(rr)` eredménye legyen.
+"""
+function probe_field_bulk(rr::ScalarField, points::Matrix{Float64})
+    gmsh.model.setCurrent(rr.model.name)
+    npts = size(points, 1)
+    vals = fill(NaN, npts)
+
+    rr_bbox = build_bbox(rr)
+
+    # 1. Lekérjük a hálóhoz tartozó adatokat egyszer
+    _, _, dim, _ = gmsh.model.mesh.getElement(rr.numElem[1])
+    elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(dim, -1)
+    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+    xnode, ynode, znode = coords[1:3:end], coords[2:3:end], coords[3:3:end]
+
+    # 2. Végigmegyünk az összes kiértékelendő ponton
+    @inbounds for ipt in 1:npts
+        x, y, z = points[ipt, 1], points[ipt, 2], points[ipt, 3]
+
+        # 2a. Gyors bbox szűrés
+        for (i, bbox) in enumerate(rr_bbox)
+            (xmin, xmax, ymin, ymax, zmin, zmax) = bbox
+            if !(x ≥ xmin && x ≤ xmax && y ≥ ymin && y ≤ ymax && z ≥ zmin && z ≤ zmax)
+                continue
+            end
+
+            # 2b. Megvizsgáljuk az elemen belüli interpolációt
+            _, _, dim, _ = gmsh.model.mesh.getElement(rr.numElem[i])
+            et = elemTypes[1]  # feltételezve azonos típus a mezőben
[…]
