+++ src/fieldtools.jl
@@ -5,6 +5,457 @@ using ..LowLevelFEM
 
 export probe_field_bulk
 
+#using LinearAlgebra
+
+const SUPPORTED_ELEMENT_TYPES = Set([1, 8, 2, 9, 3, 10])
+
+"""
+    build_bbox(rr::ScalarField)
+
+Előkészíti az elemek xy-síkbeli bounding boxait, és visszaad egy cache-t,
+amelyet a `probe_field_bulk` többször is fel tud használni.
+A visszatérési érték egy named tuple a következő kulcsokkal:
+    * `bbox` – kibővített (xmin, xmax, ymin, ymax) tartományok
+    * `elem_nodes` – a mezőhöz tartozó elemek csomópont-listái
+    * `elem_types` – a hozzájuk tartozó Gmsh elemtípusok
+    * `elem_coords` – előre kigyűjtött (x, y, z) csomópont-koordináták
+    * `elem_dims` – az elem topológiai dimenziója (1 vagy 2)
+"""
+function build_bbox(rr::ScalarField)
+    gmsh.model.setCurrent(rr.model.name)
+
+    elem_tags = Int.(collect(rr.numElem))
+    if isempty(elem_tags)
+        return (
+            bbox=NTuple{4,Float64}[],
+            elem_nodes=Vector{Vector{Int}}(),
+            elem_types=Int[],
+            elem_coords=Vector{NTuple{3,Vector{Float64}}}(),
+            elem_dims=Int[],
+        )
+    end
+
+    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+    xcoords = coords[1:3:end]
+    ycoords = coords[2:3:end]
+    zcoords = coords[3:3:end]
+    node_index = Dict{Int,Int}()
+    for (pos, tag_raw) in enumerate(nodeTags)
+        node_index[Int(tag_raw)] = pos
+    end
+
+    elem_nodes = Vector{Vector{Int}}(undef, length(elem_tags))
+    elem_types = Vector{Int}(undef, length(elem_tags))
+    elem_coords = Vector{NTuple{3,Vector{Float64}}}(undef, length(elem_tags))
+    elem_dims = Vector{Int}(undef, length(elem_tags))
+    elem_bboxes = Vector{NTuple{4,Float64}}(undef, length(elem_tags))
+
+    for (i, tag) in enumerate(elem_tags)
+        et, node_ids, dim, _ = gmsh.model.mesh.getElement(tag)
+        et in SUPPORTED_ELEMENT_TYPES || error("build_bbox: element type $et not supported (only line/tri/quad)")
+        dim in (1, 2) || error("build_bbox: expected 1D/2D mesh, got dim=$dim for element $tag")
+
+        node_vec = Int.(node_ids)
+        elem_nodes[i] = node_vec
+        elem_types[i] = et
+        elem_dims[i] = dim
+
+        xs = Vector{Float64}(undef, length(node_vec))
+        ys = similar(xs)
+        zs = similar(xs)
+        for (k, nid) in enumerate(node_vec)
+            idx = get(node_index, nid, nothing)
+            idx === nothing && error("build_bbox: node $nid missing from gmsh node list")
+            xs[k] = xcoords[idx]
+            ys[k] = ycoords[idx]
+            zs[k] = zcoords[idx]
+        end
+        elem_coords[i] = (xs, ys, zs)
+
+        xmin = minimum(xs)
+        xmax = maximum(xs)
+        ymin = minimum(ys)
+        ymax = maximum(ys)
[…]
