+++ src/nonlinear.jl
@@ -40,62 +40,7 @@ function nodePositionVector(problem)
     return VectorField([], reshape(r, :,1), [0], [], 1, :v3D, problem)
 end
 
-"""
-    ∇(r::Union{VectorField, ScalarField, TensorField}; nabla=:grad)
-
-Computes derivatives of `r`.
-- If `r` is a `ScalarField` and `nabla == :grad`, returns the gradient (a `VectorField`).
-- If `r` is a `VectorField` and `nabla == :grad`, returns the gradient (a `TensorField`).
-- If `r` is a `VectorField` and `nabla == :curl`, returns the curl (a `VectorField`).
-- If `r` is a `VectorField` and `nabla == :div`, returns the divergence (a `ScalarField`).
-- If `r` is a `TensorField` and `nabla == :div`, returns the divergence (a `VectorField`).
-
-Returns: `ScalarField`, `VectorField`, or `TensorField`
-
-Types:
-- `r`: `ScalarField`, `VectorField`, or `TensorField`
-- `nabla`: Symbol
-
-# 3D Examples (assumes `problem` is set as in the ∇ doc setup)
-
-```julia
-# One-time 3D setup (assumes examples/Fields/cube.geo exists with physical group "body")
-using LowLevelFEM
-gmsh.initialize()
-gmsh.open("examples/Fields/cube.geo")
-mat = material("body", E=210e3, ν=0.3, ρ=7.85e-9)
-problem = Problem([mat], type=:Solid)
-
-# 1) Gradient of a 3D scalar field: ∇f → VectorField
-f(X,Y,Z) = X^2 + Y*Z
-S = scalarField(problem, [field("body", f=f)])
-G = ∇(S)  # VectorField with 3 components
-
-# 2) Curl of a 3D vector field: ∇ × v → VectorField
-vx(X,Y,Z) = 0
-vy(X,Y,Z) = X
-vz(X,Y,Z) = 0
-V = vectorField(problem, [field("body", fx=vx, fy=vy, fz=vz)])
-C = ∇(V, nabla=:curl)  # approx (0, 0, 1) everywhere
-
-# 3) Divergence of a 3D vector field: ∇ ⋅ v → ScalarField
-v1(X,Y,Z) = X
-v2(X,Y,Z) = Y
-v3(X,Y,Z) = Z
-V2 = vectorField(problem, [field("body", fx=v1, fy=v2, fz=v3)])
-D = ∇(V2, nabla=:div)  # ≈ 3
-
-# 4) Divergence of a 3D tensor field: ∇ · T → VectorField (if T is TensorField)
-# For example, a diagonal tensor T with only Tzz = g(Z): div(T) = (0, 0, ∂g/∂Z)
-g(Z) = 10 - Z
-T = tensorField(problem, [field("body", fz=g)])
-DV = ∇(T, nabla=:div)  # VectorField
-
-# Symmetric displacement gradient via operators
-# A = (u ∘ ∇ + ∇ ∘ u) / 2
-gmsh.finalize()
-```
-"""
+#=
 function ∇_old(rr::Union{VectorField, ScalarField, TensorField}; nabla=:grad)
     problem = rr.model
     gmsh.model.setCurrent(problem.name)
@@ -362,6 +307,7 @@ function ∇_old(rr::Union{VectorField, ScalarField, TensorField}; nabla=:grad)
         end
     end
 end
+=#
 
 # --- gyors cache-ek elemtípusra (et) ---
 const _prop_cache = Dict{Int, Tuple{Int,Int,Vector{Float64}}}()   # et => (dim, numNodes, nodeCoord)
[…]
