+++ src/fieldtools.jl
@@ -3,11 +3,187 @@ module FieldTools
 using LinearAlgebra: dot
 using ..LowLevelFEM
 
-export probe_field_bulk
+export probe_field_bulk, build_surface_grid
+export probe_field_bulk_fallback
 
 #using LinearAlgebra
 
 const SUPPORTED_ELEMENT_TYPES = Set([1, 8, 2, 9, 3, 10])
+
+using Polyester
+using Base: @propagate_inbounds
+
+"""
+    build_surface_grid(rr; nx::Int=0, ny::Int=0)
+
+2D-s (alsó/felső) felülethez rács-alapú térindex. A rács celláiba az elemek
+xy-bbox alapján kerülnek. Visszaad egy `grid` named tuple-t.
+"""
+function build_surface_grid(rr; nx::Int=0, ny::Int=0)
+    gmsh.model.setCurrent(rr.model.name)
+
+    # -- egyszeri beolvasás Gmsh-ból
+    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+    Xf = Float32.(coords[1:3:end]); Yf = Float32.(coords[2:3:end]); Zf = Float32.(coords[3:3:end])
+    node_index = Dict{Int32,Int32}(Int32(tag)=>Int32(i) for (i,tag) in enumerate(nodeTags))
+
+    # csak a 2D-s mező elemei (tri/quad)
+    elem_tags = Int.(collect(rr.numElem))
+    isempty(elem_tags) && error("surface grid: üres mező")
+    _, _, dim, _ = gmsh.model.mesh.getElement(elem_tags[1])
+    dim == 2 || error("surface grid: a rr mező 2D felületen legyen (dim=2)")
+
+    elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(2, -1)
+
+    # --- elemlisták (csak azok, amelyek a rr.numElem-ben vannak)
+    elem_nodes = Vector{Vector{Int32}}()
+    elem_types = Int[]
+    elem_coords = Vector{NTuple{3,Vector{Float32}}}()
+    elem_bboxes = NTuple{4,Float32}[]
+
+    # bbox globális tartomány XY-ben
+    gxmin = typemax(Float32); gxmax = typemin(Float32)
+    gymin = typemax(Float32); gymax = typemin(Float32)
+
+    rrset = Set(rr.numElem)
+    for (t, et) in enumerate(elemTypes)
+        # csak tri3/tri6/quad4/quad9
+        et in (2,9,3,10) || continue
+        # elem tulajdonság: csomópont/elem
+        _, _, _, nper, _, _ = gmsh.model.mesh.getElementProperties(et)
+        tags = elemTags[t]; nlist = elemNodeTags[t]
+        @inbounds for j in 1:length(tags)
+            tag = tags[j]
+            tag ∈ rrset || continue
+            offs = (j-1)*nper + 1
+            nodes = Int32.(nlist[offs:offs+nper-1])
+            xs = Vector{Float32}(undef, nper)
+            ys = similar(xs); zs = similar(xs)
+            for (k,nid) in enumerate(nodes)
+                idx = node_index[nid]
+                xs[k] = Xf[idx]; ys[k] = Yf[idx]; zs[k] = Zf[idx]
+            end
+            xmin = minimum(xs); xmax = maximum(xs)
+            ymin = minimum(ys); ymax = maximum(ys)
+            push!(elem_nodes, nodes)
+            push!(elem_types, et)
+            push!(elem_coords, (xs,ys,zs))
+            push!(elem_bboxes, (xmin, xmax, ymin, ymax))
+            gxmin = min(gxmin, xmin); gxmax = max(gxmax, xmax)
+            gymin = min(gymin, ymin); gymax = max(gymax, ymax)
+        end
+    end
+
[…]
