+++ src/extra.jl
@@ -98,8 +98,11 @@ function pressureInVolume(p::ScalarField)
     tag = LowLevelFEM.getTagForPhysicalName(p.model.geometry.nameVolume)
     nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(p.model.dim + 1, tag)
     coord = reshape(coord, 3, :)
+    coord_mat = Matrix(coord')
     p1 = nodesToElements(p)
-    val = probe_field_bulk(p1, coord')
+    grid = build_surface_grid(p1)
+    #val = probe_field_bulk(p1, coord_mat, grid=grid)
+    val = probe_field_bulk_fallback(p1, coord_mat)
     pp = scalarField(p.model, p.model.geometry.nameVolume, 0)
     pp.a[nodeTags] .= val
 
@@ -108,7 +111,9 @@ function pressureInVolume(p::ScalarField)
         #nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(p.model.dim + 1, tag)
         #coord = reshape(coord, 3, :)
         h_top = ScalarField(p.model, p.model.geometry.nameGap, (x, y, z) -> z)
-        val = probe_field_bulk(h_top, coord')
+        grid = build_surface_grid(h_top)
+        #val = probe_field_bulk(h_top, coord_mat, grid=grid)
+        val = probe_field_bulk_fallback(h_top, coord_mat)
         hh = scalarField(p.model, p.model.geometry.nameVolume, 0)
         hh.a[nodeTags] .= val
         hh = nodesToElements(hh, onPhysicalGroup=p.model.geometry.nameVolume)
