+++ docs/src/news/2025-11-12_announcement.md
@@ -1,0 +1,17 @@
+# November 2025 – Package Announcement
+
+We are happy to announce that **LowLevelFEM.jl** is now officially available in the Julia General Registry! 🎉
+
+The package provides a lightweight and transparent FEM environment written entirely in Julia.  
+It supports:
+
+- 2D and 3D solid mechanics (plane stress/strain, axisymmetric, and full 3D)
+- Heat conduction and thermo-mechanical coupling
+- Gmsh integration for pre- and post-processing
+
+Unlike symbolic DSL-based FEM frameworks, LowLevelFEM gives you full control over assembly, boundary conditions, and solution steps — ideal for teaching, research, and method development.
+
+📚 [Documentation](https://perebalazs.github.io/LowLevelFEM.jl/stable/)  
+🗣 [Original announcement on Julia Discourse](https://discourse.julialang.org/t/ann-lowlevelfem-jl-a-lightweight-fem-toolbox-in-julia/)
+
+Stay tuned for upcoming examples and performance notes in future posts!
