+++ docs/src/news/index.md
@@ -1,0 +1,32 @@
+# News & Updates
+
+Welcome to the **LowLevelFEM.jl News** section!
+
+This page collects package announcements, new examples, and feature highlights related to the LowLevelFEM project.  
+Posts are written in plain Markdown and versioned together with the documentation — making it easy to maintain and extend.
+
+---
+
+## Latest posts
+
+- [November 2025 – Package Announcement](2025-11-12_announcement.md)
+
+---
+
+### How to contribute
+
+If you use **LowLevelFEM.jl** in teaching, research, or a student project and would like to share your results or an interesting example, feel free to open a pull request or suggest a short write-up.
+
+---
+
+📘 **Related links**
+
+- [Documentation home](../index.md)
+- [Julia Discourse announcement](https://discourse.julialang.org/t/ann-lowlevelfem-jl-a-lightweight-fem-toolbox-in-julia/)
+- [Project repository](https://github.com/perebalazs/LowLevelFEM.jl)
+
+---
+
+
+
+📡 **RSS Feed:** [https://perebalazs.github.io/LowLevelFEM.jl/news/feed.xml](https://perebalazs.github.io/LowLevelFEM.jl/news/feed.xml)
