### 2025-11-12 0dcf330
update news feed
