+++ v1.10.13/siteinfo.js
@@ -1,0 +1,2 @@
+var DOCUMENTER_CURRENT_VERSION = "v1.10.13";
+var DOCUMENTER_IS_DEV_VERSION = false;
