+++ src/extra.jl
@@ -171,7 +171,7 @@ end
 =#
 
 #function systemMatrix(problem, αInNodes::ScalarField, velocity::Number, height::ScalarField)
-function systemMatrix(problem, velocity::Number)
+function systemMatrix_old(problem, velocity::Number)
     gmsh.model.setCurrent(problem.name)
     if problem.type ≠ :Reynolds && problem.material[1].type ≠ :JFO
         error("systemMatrix: only Reynolds equation with JFO cavitation model is implemented.")
@@ -339,6 +339,254 @@ function systemMatrix(problem, velocity::Number)
     dropzeros!(KK)
     return SystemMatrix(K, problem), SystemMatrix(KK, problem)
     GC.gc()
+end
+
+function systemMatrix(problem, velocity::Number)
+    gmsh.model.setCurrent(problem.name)
+
+    if problem.type ≠ :Reynolds && problem.material[1].type ≠ :JFO
+        error("systemMatrix: only Reynolds equation with JFO cavitation model is implemented.")
+    end
+    if length(problem.material) ≠ 1
+        error("systemMatrix: only one material is permitted.")
+    end
+    if problem.geometry.h === nothing
+        initialize(problem)
+    end
+
+    # --- becslés az I,J,V,V2 méretére ---
+    elemTypes_all, elemTags_all, elemNodeTags_all = gmsh.model.mesh.getElements(problem.dim, -1)
+    lengthOfIJV = sum((div(length(elemNodeTags_all[i]), length(elemTags_all[i])) * problem.dim)^2 * length(elemTags_all[i]) for i in 1:length(elemTags_all))
+
+    # --- index és érték vektorok, típusosan ---
+    I  = Int[]
+    J  = Int[]
+    V  = Float64[]
+    V2 = Float64[]
+    sizehint!(I,  lengthOfIJV)
+    sizehint!(J,  lengthOfIJV)
+    sizehint!(V,  lengthOfIJV)
+    sizehint!(V2, lengthOfIJV)
+
+    # csak akkor kell, ha tényleg használod valahol:
+    nn = Vector{Matrix{Int}}()
+
+    ncoord2 = zeros(3 * problem.non)
+    phName  = problem.material[1].phName
+    η       = problem.material[1].η
+    κ       = problem.material[1].κ
+
+    dim  = problem.dim
+    pdim = problem.pdim
+    rowsOfB = dim  # 1D → 1, 2D → 2; az eredeti logika is ez volt
+
+    # (most csak egy anyagot engedünk továbbra is)
+    for ipg in 1:1
+        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+        tag     = getTagForPhysicalName(phName)
+
+        # node koordináták egyszer / physical group-onként
+        nodeTags, ncoord = gmsh.model.mesh.getNodesForPhysicalGroup(dim, tag)
+        @inbounds begin
+            ncoord2[nodeTags .* 3 .- 2] .= ncoord[1:3:length(ncoord)]
+            ncoord2[nodeTags .* 3 .- 1] .= ncoord[2:3:length(ncoord)]
+            ncoord2[nodeTags .* 3 .- 0] .= ncoord[3:3:length(ncoord)]
+        end
+
+        for idm in 1:length(dimTags)
+            edim, etag = dimTags[idm]
+            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+
+            for i in 1:length(elemTypes)
+                et = elemTypes[i]
+
+                elementName, dim_el, order, numNodes::Int, localNodeCoord, numPrimaryNodes =
+                    gmsh.model.mesh.getElementProperties(et)
+
[…]
