### 2025-11-18 857df3e
bug fix about non-symmetric tensors
