+++ src/general.jl
@@ -339,7 +339,7 @@ struct ScalarField <: AbstractField
     numElem = Int[]
 
     @inbounds for data in dataField
-        phName, f, fx, fy, fz, fxy, fyz, fzx = data
+        phName, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = data
         dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
 
         for (edim, etag) in dimTags
@@ -487,7 +487,7 @@ struct VectorField <: AbstractField
         ff = 0
         pdim = 1
         for i in 1:length(dataField)
-            phName, f, fx, fy, fz, fxy, fyz, fzx = dataField[i]
+            phName, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = dataField[i]
             dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
             for idm in 1:length(dimTags)
                 dimTag = dimTags[idm]
@@ -609,7 +609,7 @@ struct TensorField <: AbstractField
         ff = 0
         pdim = 1
         for i in 1:length(dataField)
-            phName, f, fx, fy, fz, fxy, fyz, fzx = dataField[i]
+            phName, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = dataField[i]
             dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
             for idm in 1:length(dimTags)
                 dimTag = dimTags[idm]
@@ -659,7 +659,6 @@ struct TensorField <: AbstractField
                                 else
                                     ff = fxy
                                 end
-                                ten1[9k-7, 1] = ff
                                 ten1[9k-5, 1] = ff
                             end
                             if fyz ≠ :no
@@ -668,7 +667,6 @@ struct TensorField <: AbstractField
                                 else
                                     ff = fyz
                                 end
-                                ten1[9k-3, 1] = ff
                                 ten1[9k-1, 1] = ff
                             end
                             if fzx ≠ :no
@@ -678,6 +676,29 @@ struct TensorField <: AbstractField
                                     ff = fzx
                                 end
                                 ten1[9k-6, 1] = ff
+                            end
+                            if fyx ≠ :no
+                                if fyx isa Function
+                                    ff = fyx(x, y, z)
+                                else
+                                    ff = fyx
+                                end
+                                ten1[9k-7, 1] = ff
+                            end
+                            if fzy ≠ :no
+                                if fzy isa Function
+                                    ff = fzy(x, y, z)
+                                else
+                                    ff = fzy
+                                end
+                                ten1[9k-3, 1] = ff
+                            end
+                            if fxz ≠ :no
+                                if fxz isa Function
+                                    ff = fxz(x, y, z)
+                                else
+                                    ff = fxz
+                                end
                                 ten1[9k-2, 1] = ff
                             end
                         end
@@ -693,7 +714,7 @@ struct TensorField <: AbstractField
[…]
