+++ docs/make.jl
@@ -29,6 +29,7 @@ makedocs(
         # thresholds are in BYTES (integers)
         size_threshold_warn = 300_000,  # ~293 KiB
         size_threshold = 600_000,       # ~586 KiB
+        assets = ["assets/analytics.html"],
     ),
     pages = DOC_PAGES,
     doctest = false,
