### 2025-11-23 128b994
assets
