+++ dev/assets/analytics.js
@@ -1,0 +1,5 @@
+window.dataLayer = window.dataLayer || [];
+function gtag(){dataLayer.push(arguments);}
+gtag('js', new Date());
+gtag('config', 'G-WVNCELC9FM*');  // <-- measurement-id
+
