### 2025-11-23 ee0b0ca
assets
