### 2025-11-23 25b883c
assets
