### 2025-11-25 0bbc78f
tangentVector
