+++ src/general.jl
@@ -4,7 +4,7 @@ export temperatureConstraint, heatFlux, heatSource, heatConvection
 export field, scalarField, vectorField, tensorField, ScalarField, VectorField, TensorField
 export constrainedDoFs, freeDoFs
 export elementsToNodes, nodesToElements, projectTo2D, expandTo3D, isNodal, isElementwise
-export fieldError, resultant, integrate, normalVector
+export fieldError, resultant, integrate, normalVector, tangentVector
 export rotateNodes, CoordinateSystem
 export showDoFResults, showModalResults, showBucklingResults
 export showStrainResults, showStressResults, showElementResults, showHeatFluxResults
@@ -1679,7 +1679,7 @@ function normalVector(problem::Problem, phName::String)
     for dimTag in dimTags
         edim, etag = dimTag
         if edim != 2
-            error("normalVector: physical group is not a surface.")
+            error("normalVector: physical group is not a surface (2D).")
         end
 
         elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
@@ -1752,6 +1752,225 @@ function normalVector(problem::Problem, phName::String)
     end
 
     return VectorField(normalvectors, [;;], [0.0], numElem, 1, :v3D, problem)
+end
+
+"""
+    tangentVector(problem::Problem, phName::String) -> VectorField
+
+Compute unit tangent vectors for all nodes of a curve-type physical group
+in a 3D Gmsh model.
+
+# Arguments
+- `problem::Problem`: A `Problem` structure containing the current Gmsh model
+  (name, dimension, number of nodes, etc.).
+- `phName::String`: The name of a physical curve group in Gmsh for which the
+  tangent vectors are computed.
+
+# Description
+The function sets the current model, queries the elements and nodes that belong
+to the given 1D physical group, and evaluates the gradients of the Lagrange
+basis functions to determine the local mapping derivative ∂x/∂ξ. Since this
+derivative represents the geometric tangent direction of the curve at each
+evaluation point, it is normalized to produce a unit tangent vector.
+
+Each node belonging to the physical curve group is assigned the corresponding
+3D unit tangent vector aligned with the parametric direction of the curve.
+
+# Returns
+- `VectorField`: A `VectorField` structure that stores the nodal tangent vectors
+  on the given physical curve.
+
+# Errors
+- Throws an error if the provided physical group is not of curve type (`dim != 1`).
+
+# Example
+```julia
+using LowLevelFEM
+
+# Load a 3D geometry containing a curved edge
+gmsh.initialize()
+gmsh.open("curve_geometry.msh")
+
+# Create a problem structure
+mat = material("body")
+prob = Problem([mat])
+
+# Compute tangent vectors along the curve named "leadingEdge"
+tv = tangentVector(prob, "leadingEdge")
+
+# Visualize the tangent field
+showDoFResults(tv)
+openPostProcessor()
+```
+"""
+function tangentVectorOld(problem::Problem, phName::String)
+    gmsh.model.setCurrent(problem.name)
+
+    non = problem.non
+    ncoord2 = zeros(3 * non)
+    numElem = Int[]
+    tangentvectors = Vector{Matrix{Float64}}()
+
+    dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+
+    for dimTag in dimTags
+        edim, etag = dimTag
+        if edim != 1
+            error("tangentVector: physical group is not a curve (1D).")
+        end
+
[…]
