### 2025-11-26 bf89a0a
tangent and normal vectors
