+++ src/general.jl
@@ -1620,8 +1620,12 @@ end
 """
     normalVector(problem::Problem, phName::String) -> VectorField
 
-Compute outward unit normal vectors for all nodes of a surface-type physical group
-in a 3D Gmsh model.
+Compute outward unit normal vectors for all nodes belonging to a surface-type
+physical group in a 3D Gmsh model.
+
+For curve-type physical groups, the function returns normal curvature vectors:
+the direction corresponds to the unit normal within the curve’s osculating
+plane, and the vector magnitude equals the local curvature of the curve.
 
 # Arguments
 - `problem::Problem`: A `Problem` structure containing the current Gmsh model 
@@ -1678,8 +1682,10 @@ function normalVector(problem::Problem, phName::String)
 
     for dimTag in dimTags
         edim, etag = dimTag
-        if edim != 2
-            error("normalVector: physical group is not a surface (2D).")
+        if edim == 1
+            normalVector1D(problem, phName)
+        elseif edim ≠ 2
+            error("normalVector: physical group is niether a curve (1D) nor a surface (2D).")
         end
 
         elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
@@ -1803,85 +1809,6 @@ showDoFResults(tv)
 openPostProcessor()
 ```
 """
-function tangentVectorOld(problem::Problem, phName::String)
-    gmsh.model.setCurrent(problem.name)
-
-    non = problem.non
-    ncoord2 = zeros(3 * non)
-    numElem = Int[]
-    tangentvectors = Vector{Matrix{Float64}}()
-
-    dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
-
-    for dimTag in dimTags
-        edim, etag = dimTag
-        if edim != 1
-            error("tangentVector: physical group is not a curve (1D).")
-        end
-
-        elemTypes, elemTags, elemNodeTags =
-            gmsh.model.mesh.getElements(edim, etag)
-
-        nodeTags, ncoord, _ =
-            gmsh.model.mesh.getNodes(edim, etag, true, false)
-
-        ncoord2[nodeTags*3 .- 2] = ncoord[1:3:end]
-        ncoord2[nodeTags*3 .- 1] = ncoord[2:3:end]
-        ncoord2[nodeTags*3 .- 0] = ncoord[3:3:end]
-
-        # --- Elemciklus ---
-        for i in 1:length(elemTypes)
-            et = elemTypes[i]
-
-            elementName, eldim, order, numNodes, localNodeCoord, _ =
-                gmsh.model.mesh.getElementProperties(et)
-
-            # 1D: nincs 3D kiterjesztés!
-            localNodeCoord2 = localNodeCoord
-
-            # ∂N/∂ξ értékek
-            _, dfun, _ = gmsh.model.mesh.getBasisFunctions(et, localNodeCoord2, "GradLagrange")
-            dN_dξ = reshape(dfun, :, numNodes)
-
-            nnet = zeros(Int, length(elemTags[i]), numNodes)
-
-            for j in 1:length(elemTags[i])
-                elem = elemTags[i][j]
-
-                @inbounds for k in 1:numNodes
-                    nnet[j,k] = elemNodeTags[i][(j-1)*numNodes + k]
-                end
-
-                xx = ncoord2[nnet[j,:]*3 .- 2]
-                yy = ncoord2[nnet[j,:]*3 .- 1]
-                zz = ncoord2[nnet[j,:]*3 .- 0]
-
-                tang_elem = zeros(3 * numNodes)
-
-                for k in 1:numNodes
[…]
