### 2025-11-29 c33712d
Update nonlinear.jl
