+++ src/fieldtools.jl
@@ -107,6 +107,21 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
         end
     end
 
+    # --- elem-szomszédsági mátrix építése (tri/quad)
+    elem_neighbors = build_surface_neighbors(elem_nodes, elem_types)
+
+    return (
+        elem_nodes=elem_nodes,
+        elem_types=elem_types,
+        elem_coords=elem_coords,
+        elem_bboxes=elem_bboxes,
+        # rács
+        bins=bins, nx=nx, ny=ny,
+        gxmin=gxmin, gymin=gymin, dx=dx, dy=dy,
+        # szomszédok
+        elem_neighbors=elem_neighbors,
+    )
+#=
     return (
         elem_nodes=elem_nodes,
         elem_types=elem_types,
@@ -116,8 +131,68 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
         bins=bins, nx=nx, ny=ny,
         gxmin=gxmin, gymin=gymin, dx=dx, dy=dy
     )
+=#
 end
 
+# ---------------------------------------------------------------------
+# Elem-szomszédsági mátrix építése 2D tri/quad felületre
+# ---------------------------------------------------------------------
+
+"""
+    build_surface_neighbors(elem_nodes, elem_types) -> Matrix{Int32}
+
+Elem-szomszédsági mátrixot épít tri3/tri6 (et=2,9) és quad4/quad9 (et=3,10)
+elemekre. A visszatérő `neighbors` mátrix mérete (nelem × 4):
+
+- háromszögeknél az első 3 oszlopban a 3 él szomszédja (4. oszlop = 0),
+- négyszögeknél mind a 4 oszlopban szomszéd, ha van (különben 0).
+
+Az éleket a csúcs-csomópontok alapján azonosítjuk, a gmsh töltési
+sorrendjét feltételezve (tri6: első 3 a csúcsok, quad9: első 4 a csúcsok).
+"""
+function build_surface_neighbors(elem_nodes::Vector{Vector{Int32}},
+                                 elem_types::Vector{Int})
+    nelem = length(elem_nodes)
+    nelem == length(elem_types) ||
+        error("build_surface_neighbors: elem_nodes and elem_types length mismatch")
+
+    neighbors = fill(Int32(0), nelem, 4)
+
+    # él -> (elem_index, local_edge_index)
+    edge_map = Dict{Tuple{Int32,Int32},Tuple{Int32,Int32}}()
+
+    for e in 1:nelem
+        nodes = elem_nodes[e]
+        et    = elem_types[e]
+
+        # csúcs-indexek és lokális él-lista
+        if et == 2 || et == 9  # tri3 / tri6
+            # csúcsok: 1,2,3
+            edges = ((1,2), (2,3), (3,1))
+        elseif et == 3 || et == 10 # quad4 / quad9
+            # csúcsok: 1,2,3,4
+            edges = ((1,2), (2,3), (3,4), (4,1))
+        else
+            error("build_surface_neighbors: unsupported element type $et (only tri/quad surface)")
+        end
+
+        for (le, (i1,i2)) in enumerate(edges)
+            n1 = nodes[i1]
+            n2 = nodes[i2]
+            key = n1 < n2 ? (n1, n2) : (n2, n1)
+
+            if haskey(edge_map, key)
+                e2, le2 = edge_map[key]
+                neighbors[e, le]   = e2
+                neighbors[e2, le2] = Int32(e)
+            else
+                edge_map[key] = (Int32(e), Int32(le))
+            end
+        end
+    end
+
+    return neighbors
+end
 
 """
[…]
