### 2025-12-04 f3ad3b2
improvements - visited works
