+++ src/fieldtools.jl
@@ -4,7 +4,7 @@ using LinearAlgebra: dot
 using ..LowLevelFEM
 
 export probe_field_bulk, build_surface_grid
-export probe_field_bulk_fallback
+export probe_field_bulk_fallback, probe_field_bulk_walking
 
 #using LinearAlgebra
 
@@ -24,7 +24,7 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
 
     # -- egyszeri beolvasás Gmsh-ból
     nodeTags, coords, _ = gmsh.model.mesh.getNodes()
-    Xf = Float32.(coords[1:3:end]); Yf = Float32.(coords[2:3:end]); Zf = Float32.(coords[3:3:end])
+    Xf = Float64.(coords[1:3:end]); Yf = Float64.(coords[2:3:end]); Zf = Float64.(coords[3:3:end])
     node_index = Dict{Int32,Int32}(Int32(tag)=>Int32(i) for (i,tag) in enumerate(nodeTags))
 
     # csak a 2D-s mező elemei (tri/quad)
@@ -38,12 +38,12 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
     # --- elemlisták (csak azok, amelyek a rr.numElem-ben vannak)
     elem_nodes = Vector{Vector{Int32}}()
     elem_types = Int[]
-    elem_coords = Vector{NTuple{3,Vector{Float32}}}()
-    elem_bboxes = NTuple{4,Float32}[]
+    elem_coords = Vector{NTuple{3,Vector{Float64}}}()
+    elem_bboxes = NTuple{4,Float64}[]
 
     # bbox globális tartomány XY-ben
-    gxmin = typemax(Float32); gxmax = typemin(Float32)
-    gymin = typemax(Float32); gymax = typemin(Float32)
+    gxmin = typemax(Float64); gxmax = typemin(Float64)
+    gymin = typemax(Float64); gymax = typemin(Float64)
 
     rrset = Set(rr.numElem)
     for (t, et) in enumerate(elemTypes)
@@ -57,7 +57,7 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
             tag ∈ rrset || continue
             offs = (j-1)*nper + 1
             nodes = Int32.(nlist[offs:offs+nper-1])
-            xs = Vector{Float32}(undef, nper)
+            xs = Vector{Float64}(undef, nper)
             ys = similar(xs); zs = similar(xs)
             for (k,nid) in enumerate(nodes)
                 idx = node_index[nid]
@@ -83,12 +83,12 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
         sidey = gymax - gymin
         # cél elemsűrűség ~ 16 elem/bin
         bins = max(1, round(Int, sqrt(ne/16)))
-        ar = sidex / max(sidey, eps(Float32))
+        ar = sidex / max(sidey, eps(Float64))
         nx = max(8, round(Int, bins*sqrt(Float64(ar))))
         ny = max(8, round(Int, bins/sqrt(Float64(ar))))
     end
-    dx = (gxmax - gxmin + eps(Float32)) / nx
-    dy = (gymax - gymin + eps(Float32)) / ny
+    dx = (gxmax - gxmin + eps(Float64)) / nx
+    dy = (gymax - gymin + eps(Float64)) / ny
 
     # --- bin lista: minden cellához elemindexek
     bins = [Int32[] for _ in 1:(nx*ny)]
@@ -729,12 +729,92 @@ end
 # ---------------------------------------------------------------------
 
 """
-    tri_reference_coordinates_raw(x, y, xnodes, ynodes)
[…]
