+++ src/fieldtools.jl
@@ -13,6 +13,27 @@ const SUPPORTED_ELEMENT_TYPES = Set([1, 8, 2, 9, 3, 10])
 using Polyester
 using Base: @propagate_inbounds
 
+# Heurisztika: adaptív rácsfelbontás elemszám és aspektusarány alapján
+@inline function _adaptive_grid_resolution(ne::Int, gxmin::Float64, gxmax::Float64,
+                                           gymin::Float64, gymax::Float64;
+                                           target_per_bin::Int = 8,
+                                           min_bins::Int = 64,
+                                           max_bins::Int = 200_000)
+
+    sidex = gxmax - gxmin
+    sidey = gymax - gymin
+    ar = sidex / max(sidey, eps(Float64))  # aspektusarány (x/y)
+
+    # Cél: ~target_per_bin elem/bin
+    total_bins = clamp(div(ne, target_per_bin), min_bins, max_bins)
+
+    # x irányban több bin, ha a tartomány "hosszúkás"
+    nx = max(8, Int(round(sqrt(total_bins * max(ar, 0.05)))))
+    ny = max(4, Int(clamp(round(total_bins / nx), 4, total_bins)))
+
+    return nx, ny
+end
+
 """
     build_surface_grid(rr; nx::Int=0, ny::Int=0)
 
@@ -24,7 +45,9 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
 
     # -- egyszeri beolvasás Gmsh-ból
     nodeTags, coords, _ = gmsh.model.mesh.getNodes()
-    Xf = Float64.(coords[1:3:end]); Yf = Float64.(coords[2:3:end]); Zf = Float64.(coords[3:3:end])
+    Xf = Float64.(coords[1:3:end])
+    Yf = Float64.(coords[2:3:end])
+    Zf = Float64.(coords[3:3:end])
     node_index = Dict{Int32,Int32}(Int32(tag)=>Int32(i) for (i,tag) in enumerate(nodeTags))
 
     # csak a 2D-s mező elemei (tri/quad)
@@ -77,34 +100,62 @@ function build_surface_grid(rr; nx::Int=0, ny::Int=0)
     ne = length(elem_nodes)
     ne == length(rr.A) || error("surface grid: eltér az elemek és rr.A száma")
 
-    # --- rács felbontás (heuriszta: ~10-20 elem/bin átlag)
+    # --- ha nincs explicit nx,ny → adaptív választás
     if nx == 0 || ny == 0
-        sidex = gxmax - gxmin
-        sidey = gymax - gymin
-        # cél elemsűrűség ~ 16 elem/bin
-        bins = max(1, round(Int, sqrt(ne/16)))
-        ar = sidex / max(sidey, eps(Float64))
-        nx = max(8, round(Int, bins*sqrt(Float64(ar))))
-        ny = max(8, round(Int, bins/sqrt(Float64(ar))))
+        nx, ny = _adaptive_grid_resolution(ne, gxmin, gxmax, gymin, gymax;
+                                           target_per_bin=4, # tetszőlegesen hangolható
+                                           min_bins=16,
+                                           max_bins=200_000)
     end
-    dx = (gxmax - gxmin + eps(Float64)) / nx
-    dy = (gymax - gymin + eps(Float64)) / ny
 
-    # --- bin lista: minden cellához elemindexek
-    bins = [Int32[] for _ in 1:(nx*ny)]
+    @info "nx=$nx, ny=$ny"
 
-    @inline cellid(ix,iy) = (iy-1)*nx + ix
[…]
