+++ src/extra.jl
@@ -85,7 +85,7 @@ function initialize(problem::Problem)
     nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(problem.dim, tag)
     coord = reshape(coord, 3, :)
     #val = probe_field_bulk_fallback(h0, coord')
-    grid = build_surface_grid(h0, nx=500, ny=500)
+    grid = build_surface_grid(h0, nx=0, ny=0)
     val = probe_field_bulk_walking(h0, coord', grid)
     h1 = scalarField(problem, problem.material[1].phName, (x,y,z)->z)
     h1.a[nodeTags] .-= val
@@ -112,7 +112,7 @@ function pressureInVolume(p::ScalarField)
     coord = reshape(coord, 3, :)
     coord_mat = Matrix(coord')
     p1 = nodesToElements(p)
-    grid = build_surface_grid(p1, nx=500, ny=500)
+    grid = build_surface_grid(p1, nx=0, ny=0)
     #val = probe_field_bulk(p1, coord_mat, grid=grid)
     #val = probe_field_bulk_fallback(p1, coord_mat)
     val = probe_field_bulk_walking(p1, coord_mat, grid)
@@ -124,7 +124,7 @@ function pressureInVolume(p::ScalarField)
         #nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(p.model.dim + 1, tag)
         #coord = reshape(coord, 3, :)
         h_top = ScalarField(p.model, p.model.geometry.nameGap, (x, y, z) -> z)
-        grid = build_surface_grid(h_top, nx=500, ny=500)
+        grid = build_surface_grid(h_top, nx=0, ny=0)
         #val = probe_field_bulk(h_top, coord_mat, grid=grid)
         #val = probe_field_bulk_fallback(h_top, coord_mat)
         val = probe_field_bulk_walking(h_top, coord_mat, grid)
