### 2025-12-08 044f6fb
VectorField can be given with a one vector value function
