+++ src/general.jl
@@ -12,6 +12,7 @@ export plotOnPath, showOnSurface
 export openPreProcessor, openPostProcessor, setParameter, setParameters
 export probe
 export saveField, loadField, isSaved
+export ∂x, ∂y, ∂z
 
 """
     Material(phName, type, E, ν, ρ, k, c, α, λ, μ, κ)
@@ -474,6 +475,44 @@ struct VectorField <: AbstractField
     function VectorField(A0, a0, t0, numElem0, nsteps0, type0, model)
         return new(A0, a0, t0, numElem0, nsteps0, type0, model)
     end
+    function VectorField(comps::Vector{ScalarField})
+        s1 = nodesToElements(comps[1])
+        s2 = nodesToElements(comps[2])
+        s3 = nodesToElements(comps[3])
+        comps1 = [s1, s2, s3]
+        n = length(comps1)
+        @assert n == 3 "VectorField must have exactly 3 scalar components."
+
+        # check same problem
+        prob = comps1[1].model
+        for s in comps1
+            @assert s.model === prob "All scalar fields must belong to the same problem."
+        end
+
+        # check element ordering
+        elems = comps1[1].numElem
+        nsteps = comps1[1].nsteps
+        for s in comps1
+            @assert s.numElem == elems "ScalarField element numbering mismatch."
+            @assert s.nsteps == nsteps "ScalarField time steps mismatch."
+        end
+
+        @disp comps1[1]
+        @disp s1
+
+        # assemble elementwise vector field
+        ε0 = Vector{Matrix{Float64}}(undef, length(elems))
+        #ε = zeros(Matrix{Float64}, length(elems))
+        for i in 1:length(elems)
+            # comps1[j].a[i] is a (numNodes × nsteps) matrix
+            blocks = [comps1[j].A[i] for j in 1:3]
+            ε0[i] = vcat(blocks...)   # 3*numNodes × nsteps
+        end
+        @disp ε0
+
+        a = [;;]
+        return new(copy(ε0), a, comps1[1].t, elems, comps1[1].nsteps, :v3D, prob)
+    end
     function VectorField(problem::Problem, dataField)
         if !isa(dataField, Vector)
             error("VectorField: dataField are not arranged in a vector. Put them in [...]")
@@ -818,6 +857,130 @@ function Base.show(io::IO, M::Union{ScalarField,VectorField,TensorField})
         display(M.A)
     end
 end
+
+function ∂(rr::ScalarField, dir::Int)
+    problem = rr.model
+    gmsh.model.setCurrent(problem.name)
+
+    @assert dir ∈ 1:3 "d(rr,dir): dir must be 1 (x), 2 (y), or 3 (z)."
+
+    # ha nodális tér, konvertáljuk elemisre
+    r = rr.a == [;;] ? elementsToNodes(rr) : rr
+    nsteps = r.nsteps
+
+    # globális node koordináták gyors elérése
+    nodeTags, ncoord, _ = gmsh.model.mesh.getNodes()
+    non = problem.non
+
+    ncoord2 = zeros(3 * non)
+    @inbounds begin
+        ncoord2[nodeTags .* 3 .- 2] = ncoord[1:3:end]
+        ncoord2[nodeTags .* 3 .- 1] = ncoord[2:3:end]
+        ncoord2[nodeTags .* 3 .- 0] = ncoord[3:3:end]
+    end
+
+    ε = Vector{Matrix{Float64}}()    # elemi eredmények
+    numElem = Int[]                  # elem azonosítók
+
+    # anyagcsoportok végigjárása
+    for ipg in 0:length(problem.material)
+        phName = ""
+        if ipg == 0
+            if rr.model.geometry.nameVolume ≠ ""
+                phName = rr.model.geometry.nameVolume
+            else
+                continue
+            end
+        else
[…]
