### 2025-12-09 3bcf8c6
x,y,z derivatives
