+++ src/general.jl
@@ -497,8 +497,8 @@ struct VectorField <: AbstractField
             @assert s.nsteps == nsteps "ScalarField time steps mismatch."
         end
 
-        @disp comps1[1]
-        @disp s1
+        #@disp comps1[1]
+        #@disp s1
 
         # assemble elementwise vector field
         ε0 = Vector{Matrix{Float64}}(undef, length(elems))
@@ -508,7 +508,7 @@ struct VectorField <: AbstractField
             blocks = [comps1[j].A[i] for j in 1:3]
             ε0[i] = vcat(blocks...)   # 3*numNodes × nsteps
         end
-        @disp ε0
+        #@disp ε0
 
         a = [;;]
         return new(copy(ε0), a, comps1[1].t, elems, comps1[1].nsteps, :v3D, prob)
@@ -858,14 +858,15 @@ function Base.show(io::IO, M::Union{ScalarField,VectorField,TensorField})
     end
 end
 
-function ∂(rr::ScalarField, dir::Int)
-    problem = rr.model
+function ∂(r::ScalarField, dir::Int)
+    @info "∂: nodal"
+    problem = r.model
     gmsh.model.setCurrent(problem.name)
 
-    @assert dir ∈ 1:3 "d(rr,dir): dir must be 1 (x), 2 (y), or 3 (z)."
+    @assert dir ∈ 1:3 "d(r,dir): dir must be 1 (x), 2 (y), or 3 (z)."
 
     # ha nodális tér, konvertáljuk elemisre
-    r = rr.a == [;;] ? elementsToNodes(rr) : rr
+    #r = rr.a == [;;] ? elementsToNodes(rr) : rr
     nsteps = r.nsteps
 
     # globális node koordináták gyors elérése
@@ -886,8 +887,8 @@ function ∂(rr::ScalarField, dir::Int)
     for ipg in 0:length(problem.material)
         phName = ""
         if ipg == 0
-            if rr.model.geometry.nameVolume ≠ ""
-                phName = rr.model.geometry.nameVolume
+            if r.model.geometry.nameVolume ≠ ""
+                phName = r.model.geometry.nameVolume
             else
                 continue
             end
@@ -976,9 +977,109 @@ function ∂(rr::ScalarField, dir::Int)
     return ScalarField(ε, [;;], r.t, numElem, nsteps, :scalar, problem)
 end
 
-∂x(r::ScalarField) = ∂(r, 1)
-∂y(r::ScalarField) = ∂(r, 2)
-∂z(r::ScalarField) = ∂(r, 3)
+function ∂e(rr::ScalarField, dir::Int)
+    @info "∂: elementwise"
+    problem = rr.model
+    gmsh.model.setCurrent(problem.name)
+
+    @assert dir ∈ 1:3 "d(rr,dir): dir must be 1 (x), 2 (y), or 3 (z)."
+
+    # --- IMPORTANT: rr must already be elementwise ---
+    @assert rr.A != [] "Elementwise derivative: rr must be elementwise ScalarField (A ≠ [])."
+
+    nsteps = rr.nsteps
+    ε = Vector{Matrix{Float64}}()     # element results
+    numElem = Int[]
+
+    # --- iterate physical groups ---
+    for ipg in 0:length(problem.material)
+        phName =
+            ipg == 0 ?
+                (rr.model.geometry.nameVolume ≠ "" ? rr.model.geometry.nameVolume : "") :
+                problem.material[ipg].phName
+
+        if phName == ""
+            continue
+        end
+
+        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+
+        @inbounds for (edim, etag) in dimTags
+            elemTypes, elemTags, elemNodeTags =
[…]
