### 2025-12-09 8ccebab
new type of VectorField and TensorField definition, new differential operators: d/dx, d/dy, d/dz
Testing is needed
