+++ src/general.jl
@@ -371,7 +371,6 @@ struct ScalarField <: AbstractField
 
     return ScalarField(A, [;;], [0.0], numElem, nsteps, type, problem)
 end
-#=
 function ScalarField(problem, dataField)
         if !isa(dataField, Vector)
             error("ScalarField: dataField are not arranged in a vector. Put them in [...]")
@@ -423,7 +422,6 @@ function ScalarField(problem, dataField)
         t = [0.0]
         return new(A, a, t, numElem, nsteps, type, problem)
     end
-    =#
     function ScalarField(problem::Problem, phName::String, data::Union{Number,Function})
         f = field(phName, f=data)
         return ScalarField(problem, [f])
@@ -434,6 +432,7 @@ end
     VectorField(A, a, t, numElem, nsteps, type, model)
     VectorField(problem, dataField)
     VectorField(problem::Problem, phName::String, data::Vector)
+    VectorField(Vector{ScalarField})
 
 Structure containing the data of a vector field (e.g., displacement or heat flux).
 - A: vector of element-wise vector data
@@ -631,6 +630,7 @@ end
     TensorField(A, a, t, numElem, nsteps, type, model)
     TensorField(problem, dataField)
     TensorField(problem::Problem, phName::String, data::Matrix)
+    TensorField(Matrix{ScalarField})
 
 Structure containing the data of a tensor field (e.g., stress or strain).
 - A: tensor of element-wise tensor data
@@ -797,39 +797,44 @@ struct TensorField <: AbstractField
             error("TensorField: size of data is $(size(data)).")
         end
     end
+    function TensorField(comps::Matrix{ScalarField})
+        @assert size(comps) == (3,3) "TensorField requires a 3×3 matrix of ScalarFields."
+   
+        # --- 1) Konvertálás elementwise alakra (nodesToElements) ---
+        comps2 = Matrix{ScalarField}(undef, 3, 3)
+        for i in 1:3, j in 1:3
+            comps2[i,j] = nodesToElements(comps[i,j])
+        end
+   
+        # --- 2) Ellenőrzés: minden komponens ugyanahhoz a problémához tartozik ---
+        prob   = comps2[1,1].model
+        numElem = comps2[1,1].numElem
+        nsteps = comps2[1,1].nsteps
+        t      = comps2[1,1].t
+   
+        for i in 1:3, j in 1:3
+            s = comps2[i,j]
+            @assert s.model === prob        "TensorField: ScalarFields must share same Problem."
+            @assert s.numElem == numElem    "TensorField: element list mismatch."
+            @assert s.nsteps  == nsteps     "TensorField: time step mismatch."
+        end
+   
+        # --- 3) Elemwise tensor összeállítása ---
+        A = Vector{Matrix{Float64}}(undef, length(numElem))
+   
+        for e in 1:length(numElem)
+            blocks = Matrix{Float64}[]
+            # 9 komponens 3×3 mátrixból sorfolytonosan
+            for i in 1:3, j in 1:3
+                push!(blocks, comps2[i,j].A[e])
+            end
+            A[e] = vcat(blocks...)   # méret: (9*numNodes × nsteps)
+        end
+   
+        # --- 4) TensorField visszaadása ---
+        return TensorField(A, [;;], t, numElem, nsteps, :tensor, prob)
+    end
 end
-
-#=
-"""
-TensorField(A, a, t, numElem, nsteps, type, model)
-
[…]
