### 2025-12-11 570bcae
setindex! for vector and tensor fields
