+++ src/general.jl
@@ -504,8 +504,13 @@ struct VectorField <: AbstractField
         #ε = zeros(Matrix{Float64}, length(elems))
         for i in 1:length(elems)
             # comps1[j].a[i] is a (numNodes × nsteps) matrix
-            blocks = [comps1[j].A[i] for j in 1:3]
-            ε0[i] = vcat(blocks...)   # 3*numNodes × nsteps
+            #blocks = [comps1[j].A[i] for j in 1:3]
+            block = zeros(size(comps1[1].A[i],1) * 3, size(comps1[1].A[i],2))
+            for l = 1:3
+                block[l:3:end, :] = comps1[l].A[i]
+            end
+            #ε0[i] = vcat(blocks...)   # 3*numNodes × nsteps
+            ε0[i] = block
         end
         #@disp ε0
 
@@ -1226,8 +1231,8 @@ grad_s = VectorField([gx, gy, gz])
   1 → x, 2 → y, 3 → z.
 """
 ∂x(r::ScalarField) = isNodal(r) ? ∂(r, 1) : ∂e(r, 1)
-∂y(r::ScalarField) = isNodal(r) ? ∂(r, 1) : ∂e(r, 2)
-∂z(r::ScalarField) = isNodal(r) ? ∂(r, 1) : ∂e(r, 3)
+∂y(r::ScalarField) = isNodal(r) ? ∂(r, 2) : ∂e(r, 2)
+∂z(r::ScalarField) = isNodal(r) ? ∂(r, 3) : ∂e(r, 3)
 
 """
     v[k] -> ScalarField
@@ -1261,7 +1266,7 @@ function Base.getindex(v::VectorField, k::Int)
 
         @inbounds for i in 1:numElem
             Ai = v.A[i]
-            rows = (k-1)*numNodes + 1 : k*numNodes
+            rows = k:3:3*numNodes
             Aout[i] = Ai[rows, :]
         end
 
@@ -1271,10 +1276,10 @@ function Base.getindex(v::VectorField, k::Int)
     # ha nodal mező
     if v.a != [;;]
         numNodes = size(v.a, 1) ÷ 3
-        rows = (k-1)*numNodes + 1 : k*numNodes
+        rows = k:3:3*numNodes
         aout = v.a[rows, :]
 
-        return ScalarField([;;], aout, v.t, v.numElem, v.nsteps, :scalar, v.model)
+        return ScalarField([], aout, v.t, v.numElem, v.nsteps, :scalar, v.model)
     end
 
     error("VectorField: no data found in A or a.")
@@ -1301,7 +1306,7 @@ Raises an error if i or j is outside 1:3.
 function Base.getindex(T::TensorField, i::Int, j::Int)
     @assert 1 ≤ i ≤ 3 && 1 ≤ j ≤ 3 "TensorField indexing must be T[i,j], with i,j ∈ 1:3"
 
-    block = (i-1)*3 + j   # which tensor component
+    block = (j-1)*3 + i   # which tensor component
     # --- elementwise ---
     if T.A != []
         numElem  = length(T.A)
@@ -1310,7 +1315,7 @@ function Base.getindex(T::TensorField, i::Int, j::Int)
 
         @inbounds for e in 1:numElem
             Ae = T.A[e]
-            rows = (block-1)*numNodes + 1 : block*numNodes
+            rows = block:9:9*numNodes
             Aout[e] = Ae[rows, :]
         end
 
@@ -1320,10 +1325,10 @@ function Base.getindex(T::TensorField, i::Int, j::Int)
     # --- nodal ---
     if T.a != [;;]
         numNodes = size(T.a, 1) ÷ 9
-        rows     = (block-1)*numNodes + 1 : block*numNodes
[…]
