### 2025-12-14 0e44ee1
definition of time dependent scalar field
