+++ src/general.jl
@@ -331,47 +331,57 @@ struct ScalarField <: AbstractField
     function ScalarField(A0, a0, t0, numElem0, nsteps0, type0, model)
         return new(A0, a0, t0, numElem0, nsteps0, type0, model)
     end
-    function ScalarField(problem::Problem, dataField::Vector)
-    gmsh.model.setCurrent(problem.name)
-    nodeTags, coords, _ = gmsh.model.mesh.getNodes()
-    type = :scalar
-    nsteps = 1
-    A = Vector{Matrix{Float64}}()
-    numElem = Int[]
+    function ScalarField(problem::Problem, dataField::Vector; steps=1, tmin=0.0, tmax=tmin+(steps-1))
+        gmsh.model.setCurrent(problem.name)
+        nodeTags, coords, _ = gmsh.model.mesh.getNodes()
+        type = :scalar
+        nsteps = steps
+        t = range(start=tmin, stop=tmax, length=steps)
+        A = Vector{Matrix{Float64}}()
+        numElem = Int[]
 
-    @inbounds for data in dataField
-        phName, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = data
-        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+        @inbounds for data in dataField
+            phName, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = data
+            dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
 
-        for (edim, etag) in dimTags
-            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
-            @inbounds for (i, et) in enumerate(elemTypes)
-                _, _, _, numNodes, _, _ = gmsh.model.mesh.getElementProperties(et)
-                for (j, elem) in enumerate(elemTags[i])
-                    sc1 = zeros(numNodes, nsteps)
-                    push!(numElem, elem)
-                    if f != :no
-                        for k in 1:numNodes
-                            nodeTag = elemNodeTags[i][(j-1)*numNodes + k]
-                            if f isa Function
-                                x = coords[3*nodeTag - 2]
-                                y = coords[3*nodeTag - 1]
-                                z = coords[3*nodeTag]
-                                sc1[k, 1] = f(x, y, z)
-                            else
-                                sc1[k, 1] = f
+            for (edim, etag) in dimTags
+                elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+                @inbounds for (i, et) in enumerate(elemTypes)
+                    _, _, _, numNodes, _, _ = gmsh.model.mesh.getElementProperties(et)
+                    for (j, elem) in enumerate(elemTags[i])
+                        sc1 = zeros(numNodes, nsteps)
+                        push!(numElem, elem)
+                        if f != :no
+                            for it in 1:steps
+                                for k in 1:numNodes
+                                    nodeTag = elemNodeTags[i][(j-1)*numNodes + k]
+                                    if f isa Function
+                                        x = coords[3*nodeTag - 2]
+                                        y = coords[3*nodeTag - 1]
+                                        z = coords[3*nodeTag]
+                                        if applicable(f,1,2,3,4)
+                                            sc1[k, it] = f(x, y, z, t[it])
+                                        elseif applicable(f,1,2,3)
+                                            sc1[k, it] = f(x, y, z)
+                                        else
+                                            error("ScalarField: function must have 3 or 4 arguments.")
+                                        end
+                                    else
[…]
