### 2025-12-15 1d509bc
expand a scalar field to more time steps with the same values
