### 2025-12-16 9a881ae
time derivative of fields
