+++ src/LowLevelFEM.jl
@@ -26,7 +26,7 @@ include("extra.jl")
 
 export @showfields, @showstruct, @showdef, @showtype, @showmem, @showmethods, @disp, @showsize
 export probe_field
-#=
+
 @setup_workload begin
     @compile_workload begin
         mat = material("dummy")
@@ -111,5 +111,5 @@ export probe_field
         _ = log(tfA)
     end
 end
-=#
+
 end #module
