+++ src/general.jl
@@ -12,7 +12,7 @@ export plotOnPath, showOnSurface
 export openPreProcessor, openPostProcessor, setParameter, setParameters
 export probe
 export saveField, loadField, isSaved
-export ∂x, ∂y, ∂z
+export ∂x, ∂y, ∂z, ∂t
 
 """
     Material(phName, type, E, ν, ρ, k, c, α, λ, μ, κ)
@@ -289,38 +289,115 @@ abstract type AbstractField end
 
 """
     ScalarField(A, a, t, numElem, nsteps, type, model)
-    ScalarField(problem, dataField)
-    ScalarField(problem::Problem, phName::String, data::Union{Number,Function})
     ScalarField(problem::Problem, dataField::Vector; steps=1, tmin=0.0, tmax=tmin+(steps-1))
+    ScalarField(problem::Problem, phName::String, data::Union{Number,Function};
+                steps=1, tmin=0.0, tmax=tmin+(steps-1))
     ScalarField(s::ScalarField; steps=1, tmin=0.0, tmax=tmin+(steps-1), step=1)
 
-Structure containing all data of a scalar field (e.g., temperature).
-- A: vector of element-wise scalar data
-- a: matrix of nodal values of the scalar field
-- numElem: vector of element tags
-- nsteps: number of time steps stored in `A` (for animations)
-- type: type of data (e.g., `:T` for temperature)
-- model: associated `Problem`
+Container for a time-dependent scalar field defined on a finite element mesh
+(e.g. temperature, pressure, potential).
 
-Types:
-- `A`: Vector{Vector{Float64}}
-- `a`: Matrix{Float64}
-- `t`: Vector{Float64}
-- `numElem`: Vector{Integer}
-- `nsteps`: Integer
-- `type`: Symbol
-- `model`: Problem
+A `ScalarField` can store the field either
+- **element-wise** (values at the element nodes, stored in `A`), or
+- **nodally** (values at global mesh nodes, stored in `a`).
 
-# Example
+Time dependence is handled by storing multiple time steps for each spatial
+degree of freedom.
+
+---
+
+## Stored data
+
+- `A` : Vector of matrices holding **element-wise** values  
+  (`A[e][k,i]` = value at local node `k` of element `e` at time step `i`)
+- `a` : Matrix of **nodal** values  
+  (`a[n,i]` = value at node `n` at time step `i`)
+- `t` : Vector of time instants corresponding to the stored time steps
+- `numElem` : Vector of element tags associated with `A`
+- `nsteps` : Number of stored time steps
+- `type` : Symbol identifying the physical meaning of the field
+- `model` : Associated `Problem`
+
+At a given time, either `A` or `a` is typically populated, depending on whether
+the field is element-wise or nodal.
+
+---
+
+## Constructors
+
+### Low-level constructor
+```julia
+ScalarField(A, a, t, numElem, nsteps, type, model)
+````
+
+Directly constructs a `ScalarField` from preallocated data arrays.
+
+---
+
+### From spatial field definitions
+
+```julia
+ScalarField(problem, dataField; steps, tmin, tmax)
+```
+
+Constructs an **element-wise** scalar field from a vector of field definitions
+(e.g. as returned by `field(...)`).
+The scalar value may be:
+
+* a constant,
+* a spatial function `f(x,y,z)`,
+* or a space–time function `f(x,y,z,t)`.
+
+Values are evaluated at the element nodes for each requested time step.
+
+---
+
+### From a physical group
+
+```julia
+ScalarField(problem, phName, data; steps, tmin, tmax)
+```
+
+Convenience constructor for defining a scalar field on a single physical group.
+Equivalent to calling `field(phName, f=data)` internally.
+
+---
+
+### From an existing ScalarField
+
+```julia
[…]
