+++ src/general.jl
@@ -4,7 +4,7 @@ export temperatureConstraint, heatFlux, heatSource, heatConvection
 export field, scalarField, vectorField, tensorField, ScalarField, VectorField, TensorField
 export constrainedDoFs, freeDoFs
 export elementsToNodes, nodesToElements, projectTo2D, expandTo3D, isNodal, isElementwise
-export fieldError, resultant, integrate, normalVector, tangentVector
+export fieldError, resultant, integrate, ∫, normalVector, tangentVector
 export rotateNodes, CoordinateSystem
 export showDoFResults, showModalResults, showBucklingResults
 export showStrainResults, showStressResults, showElementResults, showHeatFluxResults
@@ -328,7 +328,7 @@ the field is element-wise or nodal.
 ### Low-level constructor
 ```julia
 ScalarField(A, a, t, numElem, nsteps, type, model)
-````
+```
 
 Directly constructs a `ScalarField` from preallocated data arrays.
 
@@ -523,7 +523,7 @@ Vector components are stored in **interleaved form**
 ### Low-level constructor
 ```julia
 VectorField(A, a, t, numElem, nsteps, type, model)
-````
+```
 
 Directly constructs a `VectorField` from preallocated data arrays.
 
@@ -831,7 +831,7 @@ repeated for all nodes and time steps.
 ### Low-level constructor
 ```julia
 TensorField(A, a, t, numElem, nsteps, type, model)
-````
+```
 
 Directly constructs a `TensorField` from preallocated data arrays.
 
@@ -3812,9 +3812,11 @@ function resultant2(problem, field, phName, grad, component, offsetX, offsetY, o
 end
 
 """
-    integrate(problem::Problem, phName::String, f::Union{Function,ScalarField})
+    integrate(problem::Problem, phName::String, f::Union{Function,ScalarField}; step::Int64=1)
+    ∫(problem::Problem, phName::String, f::Union{Function,ScalarField}; step::Int64=1)
 
 Integrates the function or scalar field `f` over the physical group `phName` defined in the geometry of `problem`.
+If `f` is a `ScalarField`, the time step `step` will be integrated.
 
 Returns: integral
 
@@ -3831,7 +3833,7 @@ f(x, y, z) = x^2 + y^2
 Iz = integrate(prob, "body", f)
 ```
 """
-function integrate(problem::Problem, phName::String, f::Union{Function,ScalarField})
+function integrate(problem::Problem, phName::String, f::Union{Function,ScalarField}; step::Int64=1)
     gmsh.model.setCurrent(problem.name)
     f2 = 0
     if f isa ScalarField
@@ -3879,12 +3881,6 @@ function integrate(problem::Problem, phName::String, f::Union{Function,ScalarFie
                 s1 = 0
                 first = (l - 1) * numNodes + 1
                 @inbounds for j in 1:numIntPoints
-                    #x = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 2]
-                    #y = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 1]
-                    #z = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 0]
-                    ##x = h[:, j]' * ncoord2[nodeids * 3 .- 2]
-                    ##y = h[:, j]' * ncoord2[nodeids * 3 .- 1]
-                    ##z = h[:, j]' * ncoord2[nodeids * 3 .- 0]
                     x = 0.0
                     y = 0.0
                     z = 0.0
@@ -3901,7 +3897,7 @@ function integrate(problem::Problem, phName::String, f::Union{Function,ScalarFie
                         ff = f(x, y, z)
                     elseif f isa ScalarField
[…]
