+++ README.md
@@ -10,7 +10,9 @@
 
 # LowLevelFEM
 
-LowLevelFEM is a Julia package for finite element analysis with an engineering-first workflow. It exposes each phase of the workflow as simple functions (mesh → matrices → loads/BCs → solve → postprocess → visualize), so you can customize, combine, or inspect any step. Typical tasks such as strain energy or resultants are one-liners (for example, `U = q' * K * q / 2`).
+LowLevelFEM is a Julia package for finite element analysis with an engineering-first workflow.  
+It exposes each phase of the workflow as simple functions (mesh → matrices → loads/BCs → solve → postprocess → visualize), so you can customize, combine, or inspect any step **at matrix and field level**.  
+Typical tasks such as strain energy or resultants are one-liners (for example, `U = q' * K * q / 2`).
 
 ## Requirements
 
@@ -24,6 +26,7 @@ LowLevelFEM is a Julia package for finite element analysis with an engineering-f
 - Elements and order: standard line/triangle/quad/tetra/hex/pyramid/wedge with Lagrange order up to 10.
 - Materials: linear elastic (Hooke) and large deformation laws (St. Venant–Kirchhoff, compressible Neo-Hooke).
 - Matrices: stiffness `K`, mass `M` (lumped or consistent), proportional damping `C` (Rayleigh/Caughey), heat conduction/capacity, latent heat, and convection matrices/vectors.
+- **Matrix- and field-oriented FEM workflow: global stiffness, mass, damping and coupling matrices, together with scalar, vector and tensor fields, are explicit and user-accessible objects, enabling direct implementation, inspection and modification of PDE operators and constitutive laws at matrix and field level.**
 - Loads and constraints: nodal and distributed loads on physical groups; function-based loads and temperature BCs; elastic supports; initial displacement/velocity/temperature.
 - Thermal–structural coupling: thermal expansion, thermal stresses, and heat generated by elastic deformation.
 - Solvers: static (direct and iterative solvers with arbitrary preconditioners) and transient dynamics (central difference and HHT-α from the Newmark family).
@@ -31,7 +34,7 @@ LowLevelFEM is a Julia package for finite element analysis with an engineering-f
 - Field operators and results: gradient/divergence/curl; stress/strain and heat flux as element or nodal fields; smoothing at nodes with field jumps; user-defined scalar/vector/tensor FE fields.
 - Visualization and plots: Gmsh-based views for displacements, stresses, strains, heat flux, with animation for dynamics; plot results along user-defined paths; show results on surfaces.
 - Coordinate systems: rotate nodal DOFs with constant or function-defined local coordinate systems (incl. curvilinear).
-- Truss stuctures (static, transient, modal analysis)
+- Truss structures (static, transient, modal analysis)
 
 ## Installation
 
@@ -52,14 +55,16 @@ gmsh.open("your_model.geo")
 mat = material("body", E=2e5, ν=0.3)
 prob = Problem([mat], type=:PlaneStress)  # :Solid, :PlaneStrain, :AxiSymmetric, :HeatConduction, ...
 
-bc   = displacementConstraint("supp", ux=0, uy=0)
+bc    = displacementConstraint("supp", ux=0, uy=0)
 force = load("load", fy=-1)
 
-q = solveDisplacement(prob, [force], [bc])
-S = solveStress(q)
+u = solveDisplacement(prob, [force], [bc])
[…]

[…]
