+++ versions.js
@@ -13,5 +13,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.10.18";
+var DOCUMENTER_NEWEST = "v1.10.19";
 var DOCUMENTER_STABLE = "stable";
