+++ src/extra.jl
@@ -128,6 +128,224 @@ function initialize_old(problem::Problem)
     #gmsh.view.remove(problem.geometry.tagTop)
 end
 
+function projectScalarField(p::ScalarField; from="", to="")
+    problem = p.model
+    if isempty(from) || isempty(to)
+        error("projectScalarField: physical groups must be given.")
+    end
+    # --- 1) 2D elemek összegyűjtése ---
+    dimTags2D = gmsh.model.getEntitiesForPhysicalName(from)
+
+    elems2D = Int[]
+    conn2D = Dict{Int,Vector{Int}}()
+    etype2D = Dict{Int,Int}()
+    prin2D = Dict{Int,Int}()
+
+    for (edim, etag) in dimTags2D
+        elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+        for i in eachindex(elemTypes)
+            et = Int(elemTypes[i])
+            _, dim, _, numNodes, _, numPrimaryNodes = gmsh.model.mesh.getElementProperties(et)
+            dim == 2 || error("projectScalarField: 3D mesh as a source is not yet implemented.")
+
+            tags = elemTags[i]
+            nodes = elemNodeTags[i]
+
+            @inbounds for j in eachindex(tags)
+                elem = Int(tags[j])
+                push!(elems2D, elem)
+
+                i1 = (j - 1) * numNodes + 1
+                conn = Vector{Int}(undef, numNodes)
+                @inbounds for k in 1:numNodes
+                    conn[k] = Int(nodes[i1+k-1])
+                end
+
+                conn2D[elem] = conn
+                etype2D[elem] = et
+                prin2D[elem] = numPrimaryNodes
+            end
+        end
+    end
+
+    # ---- Elem indexelés: elemTag -> idx, és tömbök a gyors belső ciklushoz
+    ne = length(elems2D)
+    elem2idx = Dict{Int,Int}()
+    sizehint!(elem2idx, ne)
+    et_of = Vector{Int}(undef, ne)
+    pn_of = Vector{Int}(undef, ne)
+    conn_of = Vector{Vector{Int}}(undef, ne)
+
+    @inbounds for idx in 1:ne
+        e = elems2D[idx]
+        elem2idx[e] = idx
+        et_of[idx] = etype2D[e]
+        pn_of[idx] = prin2D[e]
+        conn_of[idx] = conn2D[e]
+    end
+
+    # --- 2) Bounding box + minimális elemsize ---
+    bbxmin = Inf
+    bbxmax = -Inf
+    bbymin = Inf
+    bbymax = -Inf
+    esizemin = Inf
+
+    nodeCoordCache = Dict{Int,NTuple{3,Float64}}()
+
+    dimTags = gmsh.model.getEntitiesForPhysicalName(from)
+    for (edim, etag) in dimTags
+        xmin, ymin, _, xmax, ymax, _ = gmsh.model.getBoundingBox(edim, etag)
+        bbxmin = min(bbxmin, xmin)
+        bbxmax = max(bbxmax, xmax)
+        bbymin = min(bbymin, ymin)
+        bbymax = max(bbymax, ymax)
+
+        elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+        for i in eachindex(elemTypes)
+            _, _, _, numNodes, _, _ = gmsh.model.mesh.getElementProperties(elemTypes[i])
+
+            @inbounds for j in eachindex(elemTags[i])
+                xemin = Inf
+                xemax = -Inf
+                yemin = Inf
+                yemax = -Inf
+
+                nodes = elemNodeTags[i][(j-1)*numNodes+1:j*numNodes]
+                @inbounds for n in nodes
+                    coord = get!(nodeCoordCache, Int(n)) do
[…]
