+++ src/linear.jl
@@ -3289,13 +3289,13 @@ end
 """
     getTagForPhysicalName(name)
                             
-Returns `tags` of elements of physical group `name`.
+Returns `tag` of elements of physical group `name`.
                             
-Returns: `tags`
+Returns: `tag`
                             
 Types:
 - `name`: String
-- `tags`: Vector{Integer}
+- `tag`: Integer
 """
 function getTagForPhysicalName(name)
     dimTags = gmsh.model.getPhysicalGroups(-1)
@@ -3307,6 +3307,29 @@ function getTagForPhysicalName(name)
         end
     end
     return dimTags[i][2]
+end
+
+"""
+    getDimForPhysicalName(name)
+                            
+Returns `dim` of elements of physical group `name`.
+                            
+Returns: `dim`
+                            
+Types:
+- `name`: String
+- `dim`: Integer
+"""
+function getDimForPhysicalName(name)
+    dimTags = gmsh.model.getPhysicalGroups(-1)
+    i = 1
+    while gmsh.model.getPhysicalName(dimTags[i][1], dimTags[i][2]) != name
+        i += 1
+        if i > length(dimTags)
+            error("Physical name '$name' does not exist.")
+        end
+    end
+    return dimTags[i][1]
 end
 
 """
