+++ src/extra.jl
@@ -94,6 +94,7 @@ function flowRate(name::String; q=0)
 end
 
 function initialize(problem::Problem)
+    #=
     fh(x, y, z) = gmsh.view.probe(problem.geometry.tagTop, x, y, 0, -1, 1, false, -1, [], [], [], 2)[1][1] - z # az alsó felületnek síknak kell lennie!
     fdhdx(x, y, z) = gmsh.view.probe(problem.geometry.tagTop, x, y, 0, -1, 1, true, -1, [], [], [], 2)[1][1]
     h = scalarField(problem, problem.material[1].phName, fh)
@@ -103,6 +104,16 @@ function initialize(problem::Problem)
     problem.geometry.h = h
     problem.geometry.dhdx = dhdx
     gmsh.view.remove(problem.geometry.tagTop)
+    =#
+
+    h0 = ScalarField(problem, problem.geometry.nameGap, (x, y, z) -> z)
+    h0 = elementsToNodes(h0)
+    h = projectScalarField(h0, from=problem.geometry.nameGap, to=problem.material[1].phName, gap=true)
+    h0 = nodesToElements(h, onPhysicalGroup=problem.material[1].phName)
+    dhdx = elementsToNodes(∂x(h0))
+    dhdx = elementsToNodes(dhdx)
+    problem.geometry.h = h
+    problem.geometry.dhdx = dhdx
 end
 
 function initialize_old(problem::Problem)
@@ -128,11 +139,119 @@ function initialize_old(problem::Problem)
     #gmsh.view.remove(problem.geometry.tagTop)
 end
 
-function projectScalarField(p::ScalarField; from="", to="")
+"""
+    projectScalarField(p::ScalarField; from="", to="", gap=false, binSize=0.7)
+
+Projects a scalar field defined on a 2D surface onto the nodes of another mesh
+(typically a 3D volume or a parallel surface), using a robust AABB-based
+spatial binning algorithm.
+
+This function is primarily intended for use in the Reynolds equation solver,
+where a pressure field defined on a lubricating surface must be transferred
+to another surface or to the surrounding 3D mesh.
+
+---
+
+## Functionality
+
+The projection is performed by:
+1. Collecting all 2D finite elements belonging to the physical group `from`.
+2. Assigning each source element to all spatial bins intersected by its
+   axis-aligned bounding box (AABB) in the *(x,y)* plane.
+3. Assigning target nodes (from the physical group `to`) to spatial bins
+   based on their *(x,y)* coordinates.
+4. For each target node, testing candidate source elements from the same bin:
+   - computing local coordinates using Gmsh,
+   - performing an inside-element test (triangular or quadrilateral),
+   - interpolating the scalar field using Gmsh-provided Lagrange basis functions.
+
+The algorithm is deterministic, robust with respect to bin size, and works
+for arbitrary element order supported by Gmsh.
+
+---
+
+## Arguments
+
+- `p::ScalarField`  
+  Scalar field defined on the source surface (`from`).
+  In the Reynolds context, this typically represents the pressure field.
+
+---
+
+## Keyword Arguments
+
+- `from::String`  
+  Name of the physical group defining the **source 2D surface**
+  on which the scalar field `p` is defined.
+
+- `to::String`  
+  Name of the physical group defining the **target mesh**
+  (surface or volume) whose nodes receive the projected values.
+
+- `gap::Bool = false`  
+  If `true`, the source surface nodes are temporarily projected onto the
+  *z = 0* plane during the projection.
+  
[…]
