+++ src/extra.jl
@@ -94,18 +94,6 @@ function flowRate(name::String; q=0)
 end
 
 function initialize(problem::Problem)
-    #=
-    fh(x, y, z) = gmsh.view.probe(problem.geometry.tagTop, x, y, 0, -1, 1, false, -1, [], [], [], 2)[1][1] - z # az alsó felületnek síknak kell lennie!
-    fdhdx(x, y, z) = gmsh.view.probe(problem.geometry.tagTop, x, y, 0, -1, 1, true, -1, [], [], [], 2)[1][1]
-    h = scalarField(problem, problem.material[1].phName, fh)
-    dhdx = scalarField(problem, problem.material[1].phName, fdhdx)
-
-    #dhdx = elementsToNodes(∂x(h))
-    problem.geometry.h = h
-    problem.geometry.dhdx = dhdx
-    gmsh.view.remove(problem.geometry.tagTop)
-    =#
-
     h0 = ScalarField(problem, problem.geometry.nameGap, (x, y, z) -> z)
     h0 = elementsToNodes(h0)
     h = projectScalarField(h0, from=problem.geometry.nameGap, to=problem.material[1].phName, gap=true)
@@ -114,29 +102,6 @@ function initialize(problem::Problem)
     dhdx = elementsToNodes(dhdx)
     problem.geometry.h = h
     problem.geometry.dhdx = dhdx
-end
-
-function initialize_old(problem::Problem)
-    h0 = ScalarField(problem, problem.geometry.nameGap, (x, y, z) -> z)
-    tag = getTagForPhysicalName(problem.material[1].phName)
-    nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(problem.dim, tag)
-    coord = reshape(coord, 3, :)
-    #val = probe_field_bulk_fallback(h0, coord')
-    grid = build_surface_grid(h0, nx=0, ny=0)
-    val = probe_field_bulk_walking(h0, coord', grid)
-    h1 = scalarField(problem, problem.material[1].phName, (x,y,z)->z)
-    h1.a[nodeTags] .-= val
-    h1.a .*= -1.0
-    h = h1 # nodesToElements(h1, onPhysicalGroup="bottom")
-
-    #fh(x, y, z) = gmsh.view.probe(problem.geometry.tagTop, x, y, 0)[1][1] - z
-    #h = scalarField(problem, problem.material[1].phName, fh)
-    grad_h = ∇(h)
-    ex = VectorField(problem, problem.material[1].phName, [1,0,0])
-    dhdx = grad_h * ex
-    problem.geometry.h = h
-    problem.geometry.dhdx = elementsToNodes(dhdx)
-    #gmsh.view.remove(problem.geometry.tagTop)
 end
 
 """
@@ -239,10 +204,15 @@ for arbitrary element order supported by Gmsh.
 
 ---
 """
-function projectScalarField(p::ScalarField; from="", to="", gap=false, binSize=0.7)
-    problem = p.model
+function projectScalarField(pp::ScalarField; from="", to="", gap=false, binSize=0.7)
+    problem = pp.model
     if isempty(from) || isempty(to)
         error("projectScalarField: physical groups must be given.")
+    end
+    if isElementwise(pp)
+        p = elementsToNodes(p)
+    else
+        p = pp
     end
     dimF = 0
     if gap == true
@@ -423,7 +393,6 @@ function projectScalarField(p::ScalarField; from="", to="", gap=false, binSize=0
         for (ni, nn) in enumerate(nodeTagsUpper)
             gmsh.model.mesh.setNode(nn, [coordUpper[3ni-2], coordUpper[3ni-1], 0.0], [0, 0, 0])
         end
-        #pvals = vec(h0.a)
     end
[…]
