### 2025-12-26 acdeaea
Update general.jl
