+++ src/extra.jl
@@ -1,5 +1,5 @@
 export pressureConstraint, flowRate
-export solvePressure, pressureInVolume, solveShearStress
+export solvePressure, solveShearStress, fieldsToVolume
 
 
 function showGapThickness(phName; name=phName, visible=false)
@@ -204,16 +204,13 @@ for arbitrary element order supported by Gmsh.
 
 ---
 """
-function projectScalarField(pp::ScalarField; from="", to="", gap=false, binSize=0.7)
-    problem = pp.model
+function projectScalarField(pp::Union{ScalarField,Vector{ScalarField}}; from="", to="", gap=false, binSize=0.7)
+    problem = pp isa Vector ? pp[1].model : pp.model
     if isempty(from) || isempty(to)
         error("projectScalarField: physical groups must be given.")
     end
-    if isElementwise(pp)
-        p = elementsToNodes(p)
-    else
-        p = pp
-    end
+    rows = pp isa Vector ? size(pp, 1) : 1
+    p = pp isa Vector ? [elementsToNodes(pp[i]) for i in eachindex(pp)] : [elementsToNodes(pp)]
     dimF = 0
     if gap == true
         tag = getTagForPhysicalName(from)
@@ -385,8 +382,13 @@ function projectScalarField(pp::ScalarField; from="", to="", gap=false, binSize=
     end
 
     # --- 5) Interpoláció ---
-    a = zeros(problem.non)
-    pvals = vec(p.a)
+    a = zeros(problem.non, rows)
+    pvals = Matrix{Float64}(undef, problem.non, rows)
+    for j in 1:rows
+        pvals[:, j] .= vec(p[j].a)
+    end
+
+    #pvals = [vec(p[i].a) for i in eachindex(p)]
     uvw = Vector{Float64}(undef, 3)
 
     if gap == true
@@ -395,6 +397,7 @@ function projectScalarField(pp::ScalarField; from="", to="", gap=false, binSize=
         end
     end
 
+    val = Vector{Float64}(undef, rows)
     for b in eachindex(ebins)
         elems = ebins[b]   # idx-ek
         nodes = nbins[b]
@@ -424,12 +427,23 @@ function projectScalarField(pp::ScalarField; from="", to="", gap=false, binSize=
                     uvw[3] = w
                     _, bf, _ = gmsh.model.mesh.getBasisFunctions(et, uvw, "Lagrange")
 
-                    val = 0.0
-                    @inbounds for i in eachindex(conn)
-                        val += bf[i] * pvals[conn[i]]
+                    fill!(val, 0.0)
+                    @inbounds for ii in eachindex(conn)
+                        nid = conn[ii]
+                        bfi = bf[ii]
+                        for j in 1:rows
+                            val[j] += bfi * pvals[nid, j]
+                        end
                     end
 
-                    a[node] = val
+                    #fill!(val, 0.0)
+                    #@inbounds for i in eachindex(conn), j in 1:rows
+                    #    val[j] += bf[i] * pvals[conn[i], j]
+                    #end
+
+                    @inbounds for i in 1:rows
+                        a[node, i] = val[i]
+                    end
                     break
                 end
             end
@@ -442,28 +456,38 @@ function projectScalarField(pp::ScalarField; from="", to="", gap=false, binSize=
         end
     end
 
-    ret = ScalarField([], reshape(a, :, 1), [0.0], [], 1, :scalar, problem)
-    if isNodal(pp)
-        return ret
+    if pp isa Vector
[…]
