+++ src/linear.jl
@@ -3287,52 +3287,6 @@ function applyBoundaryConditions!(heatCondMat::SystemMatrix, heatCapMat::SystemM
 end
 
 """
-    getTagForPhysicalName(name)
-                            
-Returns `tag` of elements of physical group `name`.
-                            
-Returns: `tag`
-                            
-Types:
-- `name`: String
-- `tag`: Integer
-"""
-function getTagForPhysicalName(name)
-    dimTags = gmsh.model.getPhysicalGroups(-1)
-    i = 1
-    while gmsh.model.getPhysicalName(dimTags[i][1], dimTags[i][2]) != name
-        i += 1
-        if i > length(dimTags)
-            error("Physical name '$name' does not exist.")
-        end
-    end
-    return dimTags[i][2]
-end
-
-"""
-    getDimForPhysicalName(name)
-                            
-Returns `dim` of elements of physical group `name`.
-                            
-Returns: `dim`
-                            
-Types:
-- `name`: String
-- `dim`: Integer
-"""
-function getDimForPhysicalName(name)
-    dimTags = gmsh.model.getPhysicalGroups(-1)
-    i = 1
-    while gmsh.model.getPhysicalName(dimTags[i][1], dimTags[i][2]) != name
-        i += 1
-        if i > length(dimTags)
-            error("Physical name '$name' does not exist.")
-        end
-    end
-    return dimTags[i][1]
-end
-
-"""
     applyBoundaryConditions!(stiffMat, massMat, dampMat, loadVec, supports)
                             
 Applies displacement boundary conditions `supports` on a stiffness matrix
