+++ docs/src/Poisson.md
@@ -1,0 +1,9 @@
+# Poisson equation
+
+```@autodocs
+Modules = [LowLevelFEM]
+Public = true
+Private = true
+Order = [:type, :function]
+Pages = ["src/poisson.jl"]
+```
