### 2025-12-30 33418a9
Poisson equation
getindex!, setindex for systemMatrix
