+++ src/operators.jl
@@ -1852,124 +1852,132 @@ end
 
 ###############################################################################
 #                                                                             #
-#                      System matrix operations                               #
+#                      SystemMatrix operations                                #
 #                                                                             #
 ###############################################################################
 
+"""
+    *(A::Union{SystemMatrix,Matrix}, B::VectorField)
+
+Matrix–vector multiplication between a system matrix and a nodal vector field.
+
+If the vector field is defined elementwise, it is automatically converted to
+nodal representation before multiplication.
+
+# Returns
+- `VectorField` containing the nodal result with one time step.
+"""
 function *(A::Union{SystemMatrix,Matrix}, BB::VectorField)
-    if BB.A != []
-        B = elementsToNodes(BB)
-    else
-        B = BB
-    end
-    type = :none
+    B = BB.A != [] ? elementsToNodes(BB) : BB
+
     if B.a != [;;] && B.nsteps == 1
-        if B.type == :v2D
-            type = :v2D
-        elseif B.type == :v3D
-            type = :v3D
-        elseif B.type == :other
-            type = :other
-        else
-            error("*(A::Union{SystemMatrix,Matrix}, B::VectorField): neither 2D nor 3D (type=$(B.type))?")
-        end
+        type = B.type
         C = A isa Matrix ? A * B.a : A.A * B.a
         return VectorField([], reshape(C, :, 1), [0.0], [], 1, type, B.model)
     else
-        error("*(A::Union{SystemMatrix,Matrix}, B::VectorField): Type of data is not nodal or more than one time steps ($(B.nsteps)).")
+        error("*(A, B::VectorField): vector field must be nodal with a single time step.")
     end
 end
 
+
+"""
+    *(A::Union{SystemMatrix,Matrix}, B::ScalarField)
+
+Matrix–vector multiplication between a system matrix and a nodal scalar field.
+
+If the scalar field is defined elementwise, it is automatically converted to
+nodal representation before multiplication.
+
+# Returns
+- `ScalarField` containing the nodal result with one time step.
+"""
 function *(A::Union{SystemMatrix,Matrix}, BB::ScalarField)
-    if BB.A != []
-        B = elementsToNodes(BB)
-    else
-        B = BB
-    end
-    type = :none
+    B = BB.A != [] ? elementsToNodes(BB) : BB
+
     if B.a != [;;] && B.nsteps == 1
-        type = :scalar
         C = A isa Matrix ? A * B.a : A.A * B.a
-        return ScalarField([], reshape(C, :, 1), [0.0], [], 1, type, B.model)
+        return ScalarField([], reshape(C, :, 1), [0.0], [], 1, :scalar, B.model)
     else
-        error("*(A::SystemMatrix, B::ScalarField): Type of data is not nodal or more than one time steps ($(B.nsteps)).")
+        error("*(A, B::ScalarField): scalar field must be nodal with a single time step.")
     end
 end
 
-import Base.\
+
+###############################################################################
+# Linear solves
+###############################################################################
+
+import Base:\
+"""
+    \\(A::Union{SystemMatrix,Matrix}, b::ScalarField)
+
[…]
