+++ src/poisson.jl
@@ -1,36 +1,61 @@
-function systemMatrixPoisson(problem::Problem, multiplier::Union{Number,ScalarField})
+export stiffnessMatrixPoisson
+
+"""
+    stiffnessMatrixPoisson(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
+
+Assembles the global stiffness matrix of a Poisson-type problem using the finite
+element method.
+
+The matrix corresponds to the weak form of the Poisson operator
+```
+
+K_ab = ∫_Ω (∇N_a · ∇N_b) · α(x) dΩ,
+
+```
+where `N_a` are the Lagrange shape functions and `α(x)` is a scalar coefficient.
+
+# Arguments
+- `problem::Problem`:
+  Finite element problem definition, including geometry, dimension, physical
+  groups, and discretization.
+- `coefficient::Union{Number,ScalarField}`:
+  Scalar coefficient `α(x)` in the Poisson equation.
+  - If a `Number`, a constant coefficient is used over the entire domain.
+  - If a `ScalarField`, the coefficient is given elementwise and interpolated
+    to Gauss points.
+
+# Returns
+- `SystemMatrix`:
+  The assembled global stiffness matrix associated with the Poisson problem.
+
+# Notes
+- The function assembles only the left-hand side operator of the Poisson equation.
+- The spatial dimension (2D or 3D) is taken from `problem`.
+- Boundary conditions and the right-hand side vector are handled separately.
+"""
+function stiffnessMatrixPoisson(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
     gmsh.model.setCurrent(problem.name)
     elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(problem.dim, -1)
     lengthOfIJV = sum([(div(length(elemNodeTags[i]), length(elemTags[i])) * problem.dim)^2 * length(elemTags[i]) for i in 1:length(elemTags)])
-    nn = []
-    I = []
-    J = []
-    V = []
-    V = convert(Vector{Float64}, V)
+    I = Vector{Int}(undef, lengthOfIJV)
+    J = Vector{Int}(undef, lengthOfIJV)
+    V = Vector{Float64}(undef, lengthOfIJV)
+    pos = 1
     sizehint!(I, lengthOfIJV)
     sizehint!(J, lengthOfIJV)
     sizehint!(V, lengthOfIJV)
 
-    if multiplier isa Number
-        p = multiplier
+    if coefficient isa Number
+        p = coefficient
     else
-        p = nodesToElements(multiplier)
+        p = nodesToElements(coefficient)
         pa = Dict(zip(p.numElem, p.A))
     end
 
     for ipg in 1:length(problem.material)
         phName = problem.material[ipg].phName
-        #kk = problem.material[ipg].k
         dim = problem.dim
         pdim = problem.pdim
-        b = problem.thickness
-        if problem.type == :Poisson
-            rowsOfB = 3
-        elseif problem.type == :Poisson2D
-            rowsOfB = 2
-        else
-            error("heatCondMatrixSolid: dimension is $(problem.dim), problem type is $(problem.type).")
-        end
 
         dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
         for idm in 1:length(dimTags)
@@ -49,16 +74,8 @@ function systemMatrixPoisson(problem::Problem, multiplier::Union{Number,ScalarFi
                 ∇h = reshape(dfun, :, numIntPoints)
                 nnet = zeros(Int, length(elemTags[i]), numNodes)
                 invJac = zeros(3, 3numIntPoints)
-                Iidx = zeros(Int, numNodes * pdim, numNodes * pdim)
[…]
