### 2025-12-31 c046595
Update operators.jl
mapScalarField
