+++ src/operators.jl
@@ -11,6 +11,8 @@ import Base.-
 import Base.log
 import Base.sqrt
 import Base.cbrt
+import Base.abs
+export mapScalarField
 using StaticArrays
 
 """
@@ -330,6 +332,213 @@ function *(b::Number, A::ScalarField)
     return A * b
 end
 
+"""
+    /(A::ScalarField, b::Number)
+
+Elementwise division of a scalar field by a constant.
+
+Each element-wise matrix of the scalar field is divided by the scalar `b`.
+If the field is nodal, it is first converted to elementwise form.
+
+# Returns
+- A new `ScalarField` containing the elementwise divided values.
+"""
+function /(AA::ScalarField, b::Number)
+    A = isNodal(AA) ? nodesToElements(AA) : AA
+    n = length(A.A)
+    C = Vector{Matrix{Float64}}(undef, n)
+    @inbounds for i in 1:n
+        C[i] = A.A[i] ./ b
+    end
+    return ScalarField(C, [;;], A.t, A.numElem, A.nsteps, A.type, A.model)
+end
+
+"""
+    /(b::Number, A::ScalarField)
+
+Elementwise division of a constant by a scalar field.
+
+Each element-wise matrix of the scalar field is used as the divisor of the
+constant `b`, i.e. `b ./ A`.
+
+If the field is nodal, it is first converted to elementwise form.
+
+# Returns
+- A new `ScalarField` containing the elementwise divided values.
+"""
+function /(b::Number, AA::ScalarField)
+    A = isNodal(AA) ? nodesToElements(AA) : AA
+    n = length(A.A)
+    C = Vector{Matrix{Float64}}(undef, n)
+    @inbounds for i in 1:n
+        C[i] = b ./ A.A[i]
+    end
+    return ScalarField(C, [;;], A.t, A.numElem, A.nsteps, A.type, A.model)
+end
+
+"""
+    log(A::ScalarField)
+
+Elementwise natural logarithm of a scalar field.
+
+Applies the natural logarithm to each entry of every element-wise matrix
+of the scalar field.
+
+If the field is nodal, it is first converted to elementwise form.
+
+# Returns
+- A new `ScalarField` containing the elementwise logarithmic values.
+"""
+function log(AA::ScalarField)
+    A = isNodal(AA) ? nodesToElements(AA) : AA
+    n = length(A.A)
+    C = Vector{Matrix{Float64}}(undef, n)
+    @inbounds for i in 1:n
+        C[i] = log.(A.A[i])
+    end
+    return ScalarField(C, [;;], A.t, A.numElem, A.nsteps, A.type, A.model)
+end
+
+"""
+    sqrt(A::ScalarField)
+
+Elementwise square root of a scalar field.
+
+Applies the square root to each entry of every element-wise matrix
+of the scalar field.
+
+If the field is nodal, it is first converted to elementwise form.
+
+# Returns
+- A new `ScalarField` containing the elementwise square-rooted values.
+"""
+function sqrt(AA::ScalarField)
+    A = isNodal(AA) ? nodesToElements(AA) : AA
+    n = length(A.A)
+    C = Vector{Matrix{Float64}}(undef, n)
+    @inbounds for i in 1:n
+        C[i] = sqrt.(A.A[i])
+    end
+    return ScalarField(C, [;;], A.t, A.numElem, A.nsteps, A.type, A.model)
+end
+
+"""
+    cbrt(A::ScalarField)
+
+Elementwise cubic root of a scalar field.
+
+Applies the cubic root to each entry of every element-wise matrix
+of the scalar field.
+
[…]
