### 2026-01-02 15fafb1
extended poisson
stiffness, mass, convection
