+++ src/poisson.jl
@@ -1,7 +1,10 @@
-export stiffnessMatrixPoisson
+export stiffnessMatrixPoisson_old
+export stiffnessMatrixPoisson, convectionMatrixPoisson, massMatrixPoisson, convectionMatrixReynoldsSkew
+export couetteMatrixReynoldsSkew
+
 
 """
-    stiffnessMatrixPoisson(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
+    stiffnessMatrixPoisson_old(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
 
 Assembles the global stiffness matrix of a Poisson-type problem using the finite
 element method.
@@ -33,10 +36,11 @@ where `N_a` are the Lagrange shape functions and `α(x)` is a scalar coefficient
 - The spatial dimension (2D or 3D) is taken from `problem`.
 - Boundary conditions and the right-hand side vector are handled separately.
 """
-function stiffnessMatrixPoisson(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
+function stiffnessMatrixPoisson_old(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
     gmsh.model.setCurrent(problem.name)
     elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(problem.dim, -1)
-    lengthOfIJV = sum([(div(length(elemNodeTags[i]), length(elemTags[i])) * problem.dim)^2 * length(elemTags[i]) for i in 1:length(elemTags)])
+    #lengthOfIJV = sum([(div(length(elemNodeTags[i]), length(elemTags[i])) * problem.dim)^2 * length(elemTags[i]) for i in 1:length(elemTags)])
+    lengthOfIJV = estimateLengthOfIJV(problem)
     I = Vector{Int}(undef, lengthOfIJV)
     J = Vector{Int}(undef, lengthOfIJV)
     V = Vector{Float64}(undef, lengthOfIJV)
@@ -76,15 +80,15 @@ function stiffnessMatrixPoisson(problem::Problem; coefficient::Union{Number,Scal
                 invJac = zeros(3, 3numIntPoints)
                 ∂h = zeros(dim, numNodes * numIntPoints)
                 K1 = zeros(pdim * numNodes, pdim * numNodes)
-                H = zeros(pdim * numIntPoints, pdim * numNodes)
-                @inbounds for k in 1:numIntPoints, l in 1:numNodes
-                    val = h[(k-1)*numNodes + l]
-                    for kk in 1:pdim
-                        row = (k-1)*pdim + kk
-                        col = (l-1)*pdim + kk
-                        H[row, col] = val
-                    end
-                end
+                #H = zeros(pdim * numIntPoints, pdim * numNodes)
+                #@inbounds for k in 1:numIntPoints, l in 1:numNodes
+                #    val = h[(k-1)*numNodes + l]
+                #    for kk in 1:pdim
+                #        row = (k-1)*pdim + kk
+                #        col = (l-1)*pdim + kk
+                #        H[row, col] = val
+                #    end
+                #end
                 @inbounds for j in 1:length(elemTags[i])
                     elem = elemTags[i][j]
                     @inbounds for k in 1:numNodes
@@ -151,5 +155,806 @@ function stiffnessMatrixPoisson(problem::Problem; coefficient::Union{Number,Scal
     K = sparse(I, J, V, dof, dof)
     dropzeros!(K)
     return SystemMatrix(K, problem)
+end
+
+
+
+
+
+
+
+
+
+
+
+
+
[…]
