+++ src/LowLevelFEM.jl
@@ -27,7 +27,7 @@ include("extra.jl")
 
 export @showfields, @showstruct, @showdef, @showtype, @showmem, @showmethods, @disp, @showsize
 export probe_field
-
+#=
 @setup_workload begin
     @compile_workload begin
         mat = material("dummy")
@@ -112,5 +112,5 @@ export probe_field
         _ = log(tfA)
     end
 end
-
+=#
 end #module
