### 2026-01-03 1ecdcbf
Poisson-type operators
