+++ src/LowLevelFEM.jl
@@ -27,7 +27,7 @@ include("extra.jl")
 
 export @showfields, @showstruct, @showdef, @showtype, @showmem, @showmethods, @disp, @showsize
 export probe_field
-#=
+
 @setup_workload begin
     @compile_workload begin
         mat = material("dummy")
@@ -110,7 +110,13 @@ export probe_field
         _ = sqrt(tfA)
         _ = cbrt(tfA)
         _ = log(tfA)
+        stiffnessMatrixPoisson(prob)
+        convectionMatrixPoisson(prob)
+        massMatrixPoisson(prob)
+        gradDivMatrix(prob)
+        symmetricGradientMatrix(prob)
+        curlCurlMatrix(prob)
     end
 end
-=#
+
 end #module
