+++ src/poisson.jl
@@ -1,10 +1,8 @@
-export stiffnessMatrixPoisson_old
-export stiffnessMatrixPoisson, convectionMatrixPoisson, massMatrixPoisson, convectionMatrixReynoldsSkew
-export couetteMatrixReynoldsSkew
-
+export stiffnessMatrixPoisson, convectionMatrixPoisson, massMatrixPoisson
+export gradDivMatrix, symmetricGradientMatrix, curlCurlMatrix
 
 """
-    stiffnessMatrixPoisson_old(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
+    stiffnessMatrixPoissonAllInOne(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
 
 Assembles the global stiffness matrix of a Poisson-type problem using the finite
 element method.
@@ -35,8 +33,11 @@ where `N_a` are the Lagrange shape functions and `α(x)` is a scalar coefficient
 - The function assembles only the left-hand side operator of the Poisson equation.
 - The spatial dimension (2D or 3D) is taken from `problem`.
 - Boundary conditions and the right-hand side vector are handled separately.
+
+Everything in one big function.
 """
-function stiffnessMatrixPoisson_old(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
+
+function stiffnessMatrixPoissonAllInOne(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
     gmsh.model.setCurrent(problem.name)
     elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(problem.dim, -1)
     #lengthOfIJV = sum([(div(length(elemNodeTags[i]), length(elemTags[i])) * problem.dim)^2 * length(elemTags[i]) for i in 1:length(elemTags)])
@@ -320,116 +321,6 @@ function _assemble_poissonlike!(
     return pos
 end
 
-function _assemble_reynolds_couette!(
-    I::Vector{Int}, J::Vector{Int}, V::Vector{Float64},
-    pos0::Int,
-    problem::Problem,
-    phName::String;
-    velocity::Number,
-    eta::Number = problem.material[1].η,
-    dir::Int = 1,
-    hfield::ScalarField = problem.geometry.h,        # nodal ScalarField
-    dhdxfield::ScalarField = problem.geometry.dhdx   # nodal ScalarField (∂h/∂x_dir)
-)
-    gmsh.model.setCurrent(problem.name)
-
-    pdim = problem.pdim
-    dim  = problem.dim
-    @assert 1 ≤ dir ≤ dim "Reynolds Couette: dir out of bounds"
-
-    etaV = Float64(6.0 * eta * velocity / 2.0)   # matches your old code: 6*η*V/2
-
-    pos = pos0
-    dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
-
-    for idm in 1:length(dimTags)
-        edim, etag = dimTags[idm]
-        elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
-
-        for itype in 1:length(elemTypes)
-            et = elemTypes[itype]
-            _, _, order, numNodes::Int64, _, _ = gmsh.model.mesh.getElementProperties(et)
-
-            intPoints, intWeights = gmsh.model.mesh.getIntegrationPoints(et, "Gauss" * string(2order + 1))
-            numIntPoints = length(intWeights)
-
-            comp, fun, _  = gmsh.model.mesh.getBasisFunctions(et, intPoints, "Lagrange")
[…]
