+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.11",
   "v1.10",
   "v1.9",
   "v1.8",
@@ -13,5 +14,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.10.19";
+var DOCUMENTER_NEWEST = "v1.11.0";
 var DOCUMENTER_STABLE = "stable";
