+++ src/poisson.jl
@@ -1,6 +1,7 @@
 export stiffnessMatrixPoisson, convectionMatrixPoisson, massMatrixPoisson
 export gradDivMatrix, symmetricGradientMatrix, curlCurlMatrix
-export sourceVector
+export tensorLaplaceMatrix, traceLaplaceMatrix, beltramiMichellMatrix
+export sourceVector, loadTensor
 
 """
     stiffnessMatrixPoissonAllInOne(problem::Problem; coefficient::Union{Number,ScalarField}=1.0)
@@ -569,6 +570,84 @@ end
     return nothing
 end
 
+@inline function _kernel_tensorlaplace_sym!(
+    Ke::Matrix{Float64},
+    w::Float64, k::Int,
+    h, ∂h,
+    numNodes::Int, pdim::Int, dim::Int, elem::Integer;
+    dir::Int = 1
+)
+    @assert pdim == 9 "tensorLaplaceMatrix requires pdim = 9 (TensorField)"
+
+    # Standard tensor Laplace (componentwise)
+    @inbounds for a in 1:numNodes, b in 1:numNodes
+        s = 0.0
+        @inbounds for d in 1:dim
+            s += ∂h[d, (k-1)*numNodes + a] *
+                 ∂h[d, (k-1)*numNodes + b]
+        end
+        val = s * w
+
+        ia = (a-1)*pdim
+        ib = (b-1)*pdim
+        @inbounds for α in 1:9
+            Ke[ia+α, ib+α] += val
+        end
+    end
+
+    # Project operator to symmetric tensor subspace
+    @inbounds for a in 1:numNodes, b in 1:numNodes
+        ia = (a-1)*pdim
+        ib = (b-1)*pdim
+
+        # xy / yx
+        s = 0.5 * (Ke[ia+4, ib+2] + Ke[ia+2, ib+4])
+        Ke[ia+4, ib+2] = s
+        Ke[ia+2, ib+4] = s
+
+        # xz / zx
+        s = 0.5 * (Ke[ia+7, ib+3] + Ke[ia+3, ib+7])
+        Ke[ia+7, ib+3] = s
+        Ke[ia+3, ib+7] = s
+
+        # yz / zy
+        s = 0.5 * (Ke[ia+8, ib+6] + Ke[ia+6, ib+8])
+        Ke[ia+8, ib+6] = s
+        Ke[ia+6, ib+8] = s
+    end
+
+    return nothing
+end
+
+@inline function _kernel_tracelaplace!(
+    Ke::Matrix{Float64},
+    w::Float64, k::Int,
+    h, ∂h,
+    numNodes::Int, pdim::Int, dim::Int, elem::Integer;
+    dir::Int = 1
+)
+    @assert pdim == 9 "traceLaplaceMatrix requires pdim = 9 (TensorField)"
+
+    @inbounds for a in 1:numNodes, b in 1:numNodes
+        s = 0.0
+        @inbounds for d in 1:dim
+            s += ∂h[d, (k-1)*numNodes + a] *
+                 ∂h[d, (k-1)*numNodes + b]
+        end
+        val = s * w
+
+        ia = (a-1)*pdim
+        ib = (b-1)*pdim
+
+        # trace indices: σxx, σyy, σzz
+        @inbounds for α in (1,5,9), β in (1,5,9)
+            Ke[ia+α, ib+β] += val
+        end
+    end
+
+    return nothing
+end
+
 # --- public API ---------------------------------------------------------------
 
 """
@@ -864,6 +943,110 @@ function curlCurlMatrix(problem::Problem; coefficient::Union{Number,ScalarField}
     return SystemMatrix(C, problem)
 end
 
+"""
[…]
