+++ src/LowLevelFEM.jl
@@ -27,7 +27,7 @@ include("extra.jl")
 
 export @showfields, @showstruct, @showdef, @showtype, @showmem, @showmethods, @disp, @showsize
 export probe_field
-
+#=
 @setup_workload begin
     @compile_workload begin
         mat = material("dummy")
@@ -116,7 +116,9 @@ export probe_field
         gradDivMatrix(prob)
         symmetricGradientMatrix(prob)
         curlCurlMatrix(prob)
+        tensorLaplaceMatrix(prob)
+        traceLaplaceMatrix(prob)
     end
 end
-
+=#
 end #module
