+++ src/operators.jl
@@ -2153,7 +2153,7 @@ end
 ###############################################################################
 
 """
-    *(A::Union{SystemMatrix,Matrix}, B::VectorField)
+    *(A::Union{SystemMatrix,Matrix}, B::Union{ScalarField,VectorField,TensorField})
 
 Matrix–vector multiplication between a system matrix and a nodal vector field.
 
@@ -2163,19 +2163,20 @@ nodal representation before multiplication.
 # Returns
 - `VectorField` containing the nodal result with one time step.
 """
-function *(A::Union{SystemMatrix,Matrix}, BB::VectorField)
+function *(A::Union{SystemMatrix,Matrix}, BB::Union{ScalarField,VectorField,TensorField})
     B = BB.A != [] ? elementsToNodes(BB) : BB
+    T = typeof(BB)
 
     if B.a != [;;] && B.nsteps == 1
         type = B.type
         C = A isa Matrix ? A * B.a : A.A * B.a
-        return VectorField([], reshape(C, :, 1), [0.0], [], 1, type, B.model)
+        return T([], reshape(C, :, 1), [0.0], [], 1, type, B.model)
     else
-        error("*(A, B::VectorField): vector field must be nodal with a single time step.")
+        error("*(A, B::Union{ScalarField,VectorField,TensorField}): vector field must be nodal with a single time step.")
     end
 end
 
-
+#=
 """
     *(A::Union{SystemMatrix,Matrix}, B::ScalarField)
 
@@ -2197,6 +2198,7 @@ function *(A::Union{SystemMatrix,Matrix}, BB::ScalarField)
         error("*(A, B::ScalarField): scalar field must be nodal with a single time step.")
     end
 end
+=#
 
 
 ###############################################################################
@@ -2205,28 +2207,29 @@ end
 
 import Base:\
 """
-    \\(A::Union{SystemMatrix,Matrix}, b::ScalarField)
+    \\(A::Union{SystemMatrix,Matrix}, b::Union{ScalarField,VectorField,TensorField})
 
-Solves the linear system `A * x = b` for a nodal scalar field right-hand side.
+Solves the linear system `A * x = b` for a nodal scalar, vectoror tensor field right-hand side.
 
-If the scalar field is defined elementwise, it is converted to nodal form
+If the field is defined elementwise, it will be converted to nodal form
 before solving.
 
 # Returns
-- `ScalarField` containing the solution.
+- `ScalarField` or `VectorField` or `TensorField` containing the solution.
 """
-function \(A::Union{SystemMatrix,Matrix}, BB::ScalarField)
+function \(A::Union{SystemMatrix,Matrix}, BB::Union{ScalarField,VectorField,TensorField})
     B = BB.A != [] ? elementsToNodes(BB) : BB
+    T = typeof(BB)
 
     if B.a != [;;] && B.nsteps == 1
         C = A isa Matrix ? A \ B.a : A.A \ B.a
-        return ScalarField([], reshape(C, :, 1), [0.0], [], 1, :scalar, B.model)
+        return T([], reshape(C, :, 1), [0.0], [], 1, BB.type, BB.model)
     else
-        error("\\(A, b::ScalarField): scalar field must be nodal with a single time step.")
+        error("\\(A, b::Union{ScalarField,VectorField,TensorField}): scalar field must be nodal with a single time step.")
     end
 end
 
-
+#=
 """
     \\(A::Union{SystemMatrix,Matrix}, b::VectorField)
 
@@ -2238,6 +2241,7 @@ before solving.
 # Returns
 - `VectorField` containing the solution.
 """
+
 function \(A::Union{SystemMatrix,Matrix}, BB::VectorField)
     B = BB.A != [] ? elementsToNodes(BB) : BB
 
[…]
