+++ src/general.jl
@@ -48,6 +48,105 @@ struct Material
     type::Symbol
     E::Float64
     ν::Float64
+    λ::Float64
+    μ::Float64
+    κ::Float64
+    ρ::Float64
+    k::Float64
+    c::Float64
+    α::Float64
+    η::Float64
+    p₀::Float64
+    A::Float64
+
+    function Material(
+        phName::String;
+        type = :Hooke,
+        E = nothing, ν = nothing, λ = nothing, μ = nothing, κ = nothing,
+        ρ = 7.85e-9,
+        k = 45.0,
+        c = 4.2e8,
+        α = 1.2e-5,
+        η = 1e-7,
+        p₀ = 0.1,
+        A = 1.0
+    )
+        type ∈ (:Hooke, :StVenantKirchhoff, :NeoHookeCompressible, :JFO) ||
+            error("Invalid material type: $type")
+
+        ec = elastic_constants(E=E, ν=ν, λ=λ, μ=μ, κ=κ)
+
+        return new(
+            phName,
+            type,
+            ec.E, ec.ν, ec.λ, ec.μ, ec.κ,
+            ρ, k, c, α,
+            η, p₀, A
+        )
+    end
+end
+
+function elastic_constants(; E=nothing, ν=nothing, λ=nothing, μ=nothing, κ=nothing)
+    given = Dict(:E=>E, :ν=>ν, :λ=>λ, :μ=>μ, :κ=>κ)
+    specified = filter(kv -> kv[2] !== nothing, given)
+    n = length(specified)
+    DEFAULT_POISSON = 0.3
+    DEFAULT_YOUNG = 2.0e5
+
+    if n == 0
+        E = DEFAULT_YOUNG
+        ν = DEFAULT_POISSON
+        μ = E / (2*(1+ν))
+        λ = 2μ*ν/(1-2ν)
+        κ = E / (3*(1-2ν))
+
+    elseif n == 1 && haskey(specified, :E)
+        E = specified[:E]
+        ν = DEFAULT_POISSON
+        μ = E / (2*(1+ν))
+        λ = 2μ*ν/(1-2ν)
+        κ = E / (3*(1-2ν))
+
+    elseif n == 2 && haskey(specified, :E) && haskey(specified, :ν)
+        E, ν = specified[:E], specified[:ν]
+        μ = E / (2*(1+ν))
+        λ = 2μ*ν/(1-2ν)
+        κ = E / (3*(1-2ν))
+
+    elseif n == 2 && haskey(specified, :μ) && haskey(specified, :ν)
+        μ, ν = specified[:μ], specified[:ν]
+        E = 2μ*(1+ν)
+        λ = 2μ*ν/(1-2ν)
+        κ = E / (3*(1-2ν))
+
+    elseif n == 2 && haskey(specified, :λ) && haskey(specified, :μ)
+        λ, μ = specified[:λ], specified[:μ]
+        ν = λ / (2*(λ+μ))
+        E = μ*(3λ+2μ)/(λ+μ)
+        κ = λ + 2μ/3
+
+    elseif n == 2 && haskey(specified, :κ) && haskey(specified, :μ)
+        κ, μ = specified[:κ], specified[:μ]
+        λ = κ - 2μ/3
+        ν = λ / (2*(λ+μ))
+        E = 2μ*(1+ν)
+
+    elseif n < 2
+        error("Insufficient elastic parameters. Specify at least E+ν, μ+ν, λ+μ, κ+μ, or E alone.")
+
+    else
+        error("Elastic constants are overdetermined: $(keys(specified))")
+    end
+
+    return (; E, ν, λ, μ, κ)
+end
+
+#=
+struct Material
[…]
