### 2026-01-07 3ce1352
change Tuple to BoundaryCondition
