+++ src/linear.jl
@@ -2876,7 +2876,15 @@ function elasticSupportMatrix(problem, elSupports)
     dof = pdim * non
     ncoord2 = zeros(3 * problem.non)
     for n in 1:length(elSupports)
-        name, kx, ky, kz = elSupports[n]
+        #name, kx, ky, kz = elSupports[n]
+        name = elSupports[n].phName
+        kx = elSupports[n].kx
+        ky = elSupports[n].ky
+        kz = elSupports[n].kz
+        kx = kx !== nothing ? kx : 0
+        ky = ky !== nothing ? ky : 0
+        kz = kz !== nothing ? kz : 0
+        kx = elSupports[n].h !== nothing ? h : kx
         if problem.pdim == 3
             f = [0, 0, 0]
         elseif problem.pdim == 2
@@ -3040,6 +3048,19 @@ function loadVector(problem, loads)
     if !isa(loads, Vector)
         error("loadVector: loads are not arranged in a vector. Put them in [...]")
     end
+
+    @inline function loadvec_helper(f, h, x, y, z, nnet, j, l)
+        if f isa Number
+            return f
+        elseif f isa Function
+            return f(x, y, z)
+        elseif f isa ScalarField
+            return h[:, j]' * f.a[nnet[l, :]]
+        else
+            error("loadVector: internal error.")
+        end
+    end
+
     pdim = problem.pdim
     DIM = problem.dim
     b = problem.thickness
@@ -3049,7 +3070,27 @@ function loadVector(problem, loads)
     ncoord2 = zeros(3 * problem.non)
     f = [.0]
     for n in 1:length(loads)
-        name, fx, fy, fz = loads[n]
+        #name, fx, fy, fz = loads[n]
+        name = loads[n].phName
+        fx = loads[n].fx
+        fy = loads[n].fy
+        fz = loads[n].fz
+        p = loads[n].p
+        hc = loads[n].h
+        T0 = loads[n].T
+        qn = loads[n].qn
+        hs = loads[n].h
+
+        sx = loads[n].sx
+        sy = loads[n].sy
+        sz = loads[n].sz
+        sxy = loads[n].sxy
+        syz = loads[n].syz
+        szx = loads[n].szx
+        syx = loads[n].syx
+        szy = loads[n].szy
+        sxz = loads[n].sxz
+
         if pdim == 1 || pdim == 2 || pdim == 3 || pdim == 9
             f = zeros(pdim)
         else
@@ -3064,7 +3105,7 @@ function loadVector(problem, loads)
         fz_is_number = fz isa Number
         fz_is_function = fz isa Function
         fz_is_field = fz isa ScalarField
-        if fy == 1im && DIM == 3
+        if p !== nothing && DIM == 3
             nv = -normalVector(problem, name)
             ex = VectorField(problem, [field(name, fx=1, fy=0, fz=0)])
             ey = VectorField(problem, [field(name, fx=0, fy=1, fz=0)])
@@ -3080,9 +3121,12 @@ function loadVector(problem, loads)
                 fx = elementsToNodes((nv ⋅ ex) * pp)
             end
         end
-        if fy == 1im && DIM ≠ 3
+        if p !== nothing && DIM ≠ 3
             error("loadVector: pressure can be given on a surface of a 3D solid.")
         end
+        fx = fx !== nothing ? fx : 0
+        fy = fy !== nothing ? fy : 0
+        fz = fz !== nothing ? fz : 0
         dimTags = gmsh.model.getEntitiesForPhysicalName(name)
         for i ∈ 1:length(dimTags)
             dimTag = dimTags[i]
@@ -3127,46 +3171,36 @@ function loadVector(problem, loads)
                             y = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 1]
[…]
