+++ src/general.jl
@@ -2698,10 +2698,10 @@ Types:
 - `h`: Float64
 - `Tₐ`: Float64 or Function
 """
-function heatConvection(name; h=10., Tₐ=20.)
-    p = 2im
-    hcv0 = name, h, Tₐ, p
-    return hcv0
+function heatConvection(name; h=nothing, T∞=nothing)
+    h === nothing && T∞ === nothing
+        error("heatSource: h and T∞ must be specified.")
+    return BoundaryCondition(name, h=h, T=T∞)
 end
 
 """
@@ -2870,7 +2870,11 @@ function vectorField(problem, dataField)
     field = zeros(non * pdim)
     
     for i in 1:length(dataField)
-        name, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = dataField[i]
+        #name, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = dataField[i]
+        name = dataField[i].phName
+        fx = dataField[i].fx
+        fy = dataField[i].fy
+        fz = dataField[i].fz
         phg = getTagForPhysicalName(name)
         nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(-1, phg)
         if isa(fx, Function) || isa(fy, Function) || isa(fz, Function)
@@ -2878,7 +2882,7 @@ function vectorField(problem, dataField)
             yy = coord[2:3:length(coord)]
             zz = coord[3:3:length(coord)]
         end
-        if fx != :no
+        if fx !== nothing
             nodeTagsX = copy(nodeTags)
             nodeTagsX *= pdim
             nodeTagsX .-= (pdim - 1)
@@ -2889,7 +2893,7 @@ function vectorField(problem, dataField)
                 field[nodeTagsX,:] .= fx
             end
         end
-        if fy != :no
+        if fy !== nothing
             nodeTagsY = copy(nodeTags)
             nodeTagsY *= pdim
             nodeTagsY .-= (pdim - 2)
@@ -2900,7 +2904,7 @@ function vectorField(problem, dataField)
                 field[nodeTagsY,:] .= fy
             end
         end
-        if fz != :no && pdim == 3
+        if fz !== nothing && pdim == 3
             nodeTagsZ = copy(nodeTags)
             nodeTagsZ *= pdim
             if isa(fz, Function)
@@ -2971,7 +2975,17 @@ function tensorField(problem, dataField; type=:e)
     field = zeros(non * pdim)
     
     for i in 1:length(dataField)
-        name, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = dataField[i]
+        #name, f, fx, fy, fz, fxy, fyz, fzx, fyx, fzy, fxz = dataField[i]
+        name = dataField[i].phName
+        fx = dataField[i].fx
+        fy = dataField[i].fy
+        fz = dataField[i].fz
+        fxy = dataField[i].fxy
+        fyz = dataField[i].fyz
+        fzx = dataField[i].fzx
+        fyx = dataField[i].fyx
+        fzy = dataField[i].fzy
+        fxz = dataField[i].fxz
         phg = getTagForPhysicalName(name)
         nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(-1, phg)
         if isa(fx, Function) || isa(fy, Function) || isa(fz, Function) || isa(fxy, Function) || isa(fyz, Function) || isa(fzx, Function)
@@ -2979,7 +2993,7 @@ function tensorField(problem, dataField; type=:e)
             yy = coord[2:3:length(coord)]
             zz = coord[3:3:length(coord)]
         end
-        if fx != :no
+        if fx !== nothing
             nodeTagsX = copy(nodeTags)
[…]
