+++ src/linear.jl
@@ -3005,6 +3005,18 @@ function elasticSupportMatrix(problem, elSupports)
     return SystemMatrix(C, problem)
 end
 
+@inline function _loadvec_helper(f, h, x, y, z, nnet, j, l)
+    if f isa Number
+        return f
+    elseif f isa Function
+        return f(x, y, z)
+    elseif f isa ScalarField
+        return h[:, j]' * f.a[nnet[l, :]]
+    else
+        error("loadVector: internal error.")
+    end
+end
+
 """
     loadVector(problem, loads)
                             
@@ -3029,7 +3041,7 @@ The interpretation of the physical group dimension is:
   - Surface → surface force
   - Volume → body force
 
-Load components may be specified as constants, spatial functions, or `ScalarField`s.
+Load components may be specified as constants, spatial functions, or `ScalarField`s (nodal type).
 Axisymmetric and other weighted formulations can be handled by including the
 appropriate geometric factor (e.g. `2πr`) in the load definition or coefficient.
 
@@ -3040,25 +3052,13 @@ Return: `loadVec`
                             
 Types:
 - `problem`: Problem
-- `loads`: Vector{Tuple{String, Float64, Float64, Float64}}
+- `loads`: Vector{BoundaryCondition}
 - `loadVec`: VectorField
 """
 function loadVector(problem, loads)
     gmsh.model.setCurrent(problem.name)
     if !isa(loads, Vector)
         error("loadVector: loads are not arranged in a vector. Put them in [...]")
-    end
-
-    @inline function loadvec_helper(f, h, x, y, z, nnet, j, l)
-        if f isa Number
-            return f
-        elseif f isa Function
-            return f(x, y, z)
-        elseif f isa ScalarField
-            return h[:, j]' * f.a[nnet[l, :]]
-        else
-            error("loadVector: internal error.")
-        end
     end
 
     pdim = problem.pdim
@@ -3068,7 +3068,7 @@ function loadVector(problem, loads)
     dof = pdim * non
     fp = zeros(dof)
     ncoord2 = zeros(3 * problem.non)
-    f = [.0]
+    f = nothing
     for n in 1:length(loads)
         #name, fx, fy, fz = loads[n]
         name = loads[n].phName
@@ -3087,10 +3087,9 @@ function loadVector(problem, loads)
         sxy = loads[n].sxy
         syz = loads[n].syz
         szx = loads[n].szx
-        syx = loads[n].syx
-        szy = loads[n].szy
-        sxz = loads[n].sxz
 
+        (qn !== nothing || hc !== nothing || hs !== nothing || T !== nothing) && (fx !== nothing || fy !== nothing || fz !== nothing) &&
+            error("loadVector: qn/h/T∞ and fx/fy/fz cannot be defined in the same BC.")
         if pdim == 1 || pdim == 2 || pdim == 3 || pdim == 9
             f = zeros(pdim)
         else
@@ -3110,12 +3109,12 @@ function loadVector(problem, loads)
             ex = VectorField(problem, [field(name, fx=1, fy=0, fz=0)])
             ey = VectorField(problem, [field(name, fx=0, fy=1, fz=0)])
             ez = VectorField(problem, [field(name, fx=0, fy=0, fz=1)])
-            if fx_is_number || fx_is_field
-                fy = elementsToNodes((nv ⋅ ey) * fx)
-                fz = elementsToNodes((nv ⋅ ez) * fx)
-                fx = elementsToNodes((nv ⋅ ex) * fx)
-            elseif fx_is_function
-                pp = scalarField(problem, [field(name, f=fx)])
+            if p isa Number || p isa ScalarField
+                fy = elementsToNodes((nv ⋅ ey) * p)
+                fz = elementsToNodes((nv ⋅ ez) * p)
[…]
