### 2026-01-08 8b0533f
Tuple --> BoundaryCondition: done
testing in progress
