+++ src/general.jl
@@ -1,5 +1,5 @@
-export Problem, material, getEigenVectors, getEigenValues
-export displacementConstraint, load, elasticSupport
+export Problem, Material, getEigenVectors, getEigenValues
+export displacementConstraint, load, elasticSupport, BoundaryCondition
 export temperatureConstraint, heatFlux, heatSource, heatConvection
 export field, scalarField, vectorField, tensorField, ScalarField, VectorField, TensorField
 export constrainedDoFs, freeDoFs, DoFs
@@ -1213,40 +1213,40 @@ struct BoundaryCondition
     phName::String
 
     # Dirichlet-type (primary variables)
-    ux::Union{Nothing, Float64, Function, ScalarField}
-    uy::Union{Nothing, Float64, Function, ScalarField}
-    uz::Union{Nothing, Float64, Function, ScalarField}
-    p ::Union{Nothing, Float64, Function, ScalarField}   # pressure / scalar field
-    T ::Union{Nothing, Float64, Function, ScalarField}   # temperature
+    ux::Union{Nothing, Number, Function, ScalarField}
+    uy::Union{Nothing, Number, Function, ScalarField}
+    uz::Union{Nothing, Number, Function, ScalarField}
+    p ::Union{Nothing, Number, Function, ScalarField}   # pressure / scalar field
+    T ::Union{Nothing, Number, Function, ScalarField}   # temperature
 
     # Neumann-type (fluxes / forces)
-    f::Union{Nothing, Float64, Function, ScalarField}
-    fx::Union{Nothing, Float64, Function, ScalarField}
-    fy::Union{Nothing, Float64, Function, ScalarField}
-    fz::Union{Nothing, Float64, Function, ScalarField}
-    fxy::Union{Nothing, Float64, Function, ScalarField}
-    fyz::Union{Nothing, Float64, Function, ScalarField}
-    fzx::Union{Nothing, Float64, Function, ScalarField}
-    fyx::Union{Nothing, Float64, Function, ScalarField}
-    fzy::Union{Nothing, Float64, Function, ScalarField}
-    fxz::Union{Nothing, Float64, Function, ScalarField}
-    #qx ::Union{Nothing, Float64, Function}   # heat flux / generic scalar flux
-    #qy ::Union{Nothing, Float64, Function}   # heat flux / generic scalar flux
-    #qz ::Union{Nothing, Float64, Function}   # heat flux / generic scalar flux
-    kx ::Union{Nothing, Float64, Function, ScalarField}
-    ky ::Union{Nothing, Float64, Function, ScalarField}
-    kz ::Union{Nothing, Float64, Function, ScalarField}
-    q ::Union{Nothing, Float64, Function, ScalarField}   # heat flux / generic scalar flux
-    qn ::Union{Nothing, Float64, Function, ScalarField}   # heat flux / generic scalar flux
-    h ::Union{Nothing, Float64, Function, ScalarField}
+    f::Union{Nothing, Number, Function, ScalarField}
+    fx::Union{Nothing, Number, Function, ScalarField}
+    fy::Union{Nothing, Number, Function, ScalarField}
+    fz::Union{Nothing, Number, Function, ScalarField}
+    fxy::Union{Nothing, Number, Function, ScalarField}
+    fyz::Union{Nothing, Number, Function, ScalarField}
+    fzx::Union{Nothing, Number, Function, ScalarField}
+    fyx::Union{Nothing, Number, Function, ScalarField}
+    fzy::Union{Nothing, Number, Function, ScalarField}
[…]
