+++ src/nonlinear.jl
@@ -9,6 +9,7 @@ export grad_xy
     nodePositionVector(problem)
 
 Returns the position vectors of all mesh nodes as a `VectorField` (initial configuration).
+Returning vector is always a 3D vector.
 
 Returns: `R`
 
@@ -17,27 +18,11 @@ Types:
 - `R`: VectorField
 """
 function nodePositionVector(problem)
-    dim = problem.dim
-    non = problem.non
-    if dim != 3
-        error("nodePositionVector: This function only works for 3D problems.")
-    end
-    r = zeros(non * dim)
-    for ipg in 1:length(problem.material)
-        phName = problem.material[ipg].phName
-        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
-        for idm in 1:length(dimTags)
-            dimTag = dimTags[idm]
-            edim = dimTag[1]
-            etag = dimTag[2]
-            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
-            #nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes(edim, -1, true, false)
-            nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes()
-            r[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
-            r[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
-            r[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
-        end
-    end
+    nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes()
+    r = zeros(length(nodeTags) * 3)
+    r[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
+    r[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
+    r[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
     return VectorField([], reshape(r, :,1), [0], [], 1, :v3D, problem)
 end
 
@@ -1607,7 +1592,11 @@ function nonFollowerLoadVector(r::VectorField, loads)
     F0 = ∇(r)
     F0 = elementsToNodes(F0)
     for n in 1:length(loads)
-        name, fx, fy, fz = loads[n]
+        #name, fx, fy, fz = loads[n]
+        name = loads[n].phName
+        fx = loads[n].fx
+        fy = loads[n].fy
+        fz = loads[n].fz
         if pdim == 3
             f = [.0, .0, .0]
         elseif pdim == 2
@@ -1683,26 +1672,18 @@ function nonFollowerLoadVector(r::VectorField, loads)
                         F1 = reshape(F1, 3, 3)
                         iF1 =inv(F1)
                         x = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 2]
-                        y = 0
-                        z = 0
-                        if isa(fx, Function) || isa(fy, Function) || isa(fz, Function)
-                            y = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 1]
-                            z = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 0]
+                        y = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 1]
+                        z = h[:, j]' * ncoord2[nnet[l, :] * 3 .- 0]
+                        if fx !== nothing
+                            f[1] = _loadvec_helper(fx, h, x, y, z, nnet, j, l)
                         end
-                        if fz == 2im
-                            if isa(fx, Function)
-                                error("heatConvectionVector: h cannot be a function.")
-                            end
-                            f[1] = isa(fy, Function) ? fx * fy(x, y, z) : fx * fy
-                        else
-                            f[1] = isa(fx, Function) ? fx(x, y, z) : fx
[…]

[…]
