### 2026-01-08 496d430
new loadVector & applyBoundaryCondition
