+++ src/linear.jl
@@ -3082,9 +3082,15 @@ function loadVector(problem, loads)
         qn = loads[n].qn
         hs = loads[n].h
 
+        q = loads[n].q
+
         fxy = loads[n].fxy
         fyz = loads[n].fyz
         fzx = loads[n].fzx
+        
+        fyx = loads[n].fyx
+        fzy = loads[n].fzy
+        fxz = loads[n].fxz
 
         (qn !== nothing || hc !== nothing || hs !== nothing || T !== nothing) && (fx !== nothing || fy !== nothing || fz !== nothing) &&
             error("loadVector: qn/h/T∞ and fx/fy/fz cannot be defined in the same BC.")
@@ -3093,16 +3099,7 @@ function loadVector(problem, loads)
         else
             error("loadVector: dimension of the problem is $(problem.dim).")
         end
-        fx_is_number = fx isa Number
-        fx_is_function = fx isa Function
-        fx_is_field = fx isa ScalarField
-        fy_is_number = fy isa Number
-        fy_is_function = fy isa Function
-        fy_is_field = fy isa ScalarField
-        fz_is_number = fz isa Number
-        fz_is_function = fz isa Function
-        fz_is_field = fz isa ScalarField
-        if p !== nothing && DIM == 3
+        if p !== nothing && DIM == 3 && pdim == 3
             nv = -normalVector(problem, name)
             ex = VectorField(problem, [field(name, fx=1, fy=0, fz=0)])
             ey = VectorField(problem, [field(name, fx=0, fy=1, fz=0)])
@@ -3173,6 +3170,8 @@ function loadVector(problem, loads)
                             f[1] = _loadvec_helper(qn, h, x, y, z, nnet, j, l)
                         elseif hs !== nothing
                             f[1] = _loadvec_helper(hs, h, x, y, z, nnet, j, l)
+                        elseif q !== nothing
+                            f[1] = _loadvec_helper(q, h, x, y, z, nnet, j, l)
                         elseif fx !== nothing && pdim <= 3
                             f[1] = _loadvec_helper(fx, h, x, y, z, nnet, j, l)
                         end
@@ -3195,15 +3194,21 @@ function loadVector(problem, loads)
                             end
                             if fxy !== nothing
                                 f[4] = _loadvec_helper(fxy, h, x, y, z, nnet, j, l)
-                                f[2] = f[4]
                             end
                             if fyz !== nothing
                                 f[8] = _loadvec_helper(fyz, h, x, y, z, nnet, j, l)
-                                f[6] = f[8]
                             end
                             if fzx !== nothing
                                 f[3] = _loadvec_helper(fzx, h, x, y, z, nnet, j, l)
-                                f[7] = f[3]
+                            end
+                            if fyx !== nothing
+                                f[2] = _loadvec_helper(fyx, h, x, y, z, nnet, j, l)
+                            end
+                            if fzy !== nothing
+                                f[6] = _loadvec_helper(fzy, h, x, y, z, nnet, j, l)
+                            end
+                            if fxz !== nothing
+                                f[7] = _loadvec_helper(fxz, h, x, y, z, nnet, j, l)
                             end
                         end
                         r = x
@@ -3563,34 +3568,59 @@ function applyBoundaryConditions!(stiffMat::SystemMatrix, massMat::SystemMatrix,
[…]
