+++ src/general.jl
@@ -1,8 +1,8 @@
-export Problem, Material, getEigenVectors, getEigenValues
+export Problem, Material, getEigenVectors, getEigenValues, material
 export displacementConstraint, load, elasticSupport, BoundaryCondition
 export temperatureConstraint, heatFlux, heatSource, heatConvection
 export field, scalarField, vectorField, tensorField, ScalarField, VectorField, TensorField
-export constrainedDoFs, freeDoFs, DoFs
+export constrainedDoFs, freeDoFs, allDoFs, DoFs
 export elementsToNodes, nodesToElements, projectTo2D, expandTo3D, isNodal, isElementwise
 export fieldError, resultant, integrate, ∫, normalVector, tangentVector
 export rotateNodes, CoordinateSystem
@@ -107,6 +107,14 @@ function elastic_constants(; E=nothing, ν=nothing, λ=nothing, μ=nothing, κ=n
         λ = 2μ*ν/(1-2ν)
         κ = E / (3*(1-2ν))
 
+    elseif n == 1 && haskey(specified, :κ)
+        κ = specified[:κ]
+        ν = DEFAULT_POISSON
+        E = DEFAULT_YOUNG
+        μ = E / (2*(1+ν))
+        λ = 2μ*ν/(1-2ν)
+        #κ = E / (3*(1-2ν))
+
     elseif n == 2 && haskey(specified, :E) && haskey(specified, :ν)
         E, ν = specified[:E], specified[:ν]
         μ = E / (2*(1+ν))
@@ -139,6 +147,19 @@ function elastic_constants(; E=nothing, ν=nothing, λ=nothing, μ=nothing, κ=n
     end
 
     return (; E, ν, λ, μ, κ)
+end
+
+function material(  phName::String;
+        type = :Hooke,
+        E = nothing, ν = nothing, λ = nothing, μ = nothing, κ = nothing,
+        ρ = 7.85e-9,
+        k = 45.0,
+        c = 4.2e8,
+        α = 1.2e-5,
+        η = 1e-7,
+        p₀ = 0.1,
+        A = 1.0)
+    return Material(phName, type=type, E=E, ν=ν, λ=λ, μ=μ, κ=κ, ρ=ρ, k=k, c=c, α=α, η=η, p₀=p₀, A=A)
 end
 
 #=
@@ -2861,7 +2882,7 @@ Here VectorField is defined in nodes.
 """
 function vectorField(problem, dataField)
     if !isa(dataField, Vector)
-        error("applyBoundaryConditions!: dataField are not arranged in a vector. Put them in [...]")
+        error("vectorField: dataField are not arranged in a vector. Put them in [...]")
     end
     gmsh.model.setCurrent(problem.name)
     pdim = problem.dim
@@ -2966,7 +2987,7 @@ Here TensorField is defined in nodes.
 """
 function tensorField(problem, dataField; type=:e)
     if !isa(dataField, Vector)
-        error("applyBoundaryConditions!: dataField are not arranged in a vector. Put them in [...]")
+        error("tensorField!: dataField are not arranged in a vector. Put them in [...]")
     end
     gmsh.model.setCurrent(problem.name)
     pdim = 9
@@ -3513,26 +3534,26 @@ Return: `DoFs`
 
 Types:
 - `problem`: Problem
-- `supports`: Vector{Tuple{String, Float64, Float64, Float64}}
+- `supports`: Vector{BoundaryCondition}
 - `DoFs`: Vector{Int64}
 """
 function constrainedDoFs(problem, supports)
     if !isa(supports, Vector)
-        error("applyBoundaryConditions!: supports are not arranged in a vector. Put them in [...]")
+        error("constrainedDoFs: supports are not arranged in a vector. Put them in [...]")
     end
     gmsh.model.setCurrent(problem.name)
[…]
