+++ src/linear.jl
@@ -8,7 +8,7 @@ export solveModalAnalysis, solveBuckling
 export initialDisplacement, initialDisplacement!, initialVelocity, initialVelocity!
 export nodalForce!, nodalAcceleration!
 export largestPeriodTime, smallestPeriodTime, largestEigenValue, smallestEigenValue
-export CMD, HHT, CDMaccuracyAnalysis, HHTaccuracyAnalysis
+export CDM, HHT, CDMaccuracyAnalysis, HHTaccuracyAnalysis
 
 # linear.jl tetején
 mutable struct SolidWS
@@ -3576,6 +3576,7 @@ end
         uux = ux.(xx, yy, zz)
         field.a[nodeTagsX,:] .= uux
     elseif ux isa ScalarField
+        ux = elementsToNodes(ux)
         uux = ux.a[nodeTags]
         field.a[nodeTagsX,:] .= uux
     else
@@ -3685,7 +3686,7 @@ function applyBoundaryConditions(problem::Problem, supports::Vector{BoundaryCond
         field = scalarField(problem, problem.material[1].phName, 0)
     elseif problem.pdim == 2
         field = vectorField(problem, problem.material[1].phName, [0,0,0])
-        field = projectTo2D(field)
+        #field = projectTo2D(field)
     elseif problem.pdim == 3
         field = vectorField(problem, problem.material[1].phName, [0,0,0])
     elseif problem.pdim == 9
@@ -3744,7 +3745,7 @@ end
 
 """
     solveDisplacement(problem, load, supp, elSupp;
-                      condensed = false,
+                      condensed = true,
                       iterative = false,
                       reltol::Real = sqrt(eps()),
                       maxiter::Int = problem.non * problem.dim,
@@ -3776,7 +3777,7 @@ Types:
 - `q`: VectorField
 """
 function solveDisplacement(problem, load, supp, elSupp;
-                           condensed=false,
+                           condensed=true,
                            iterative=false,
                            reltol::Real = sqrt(eps()),
                            maxiter::Int = problem.non * problem.dim,
@@ -3792,7 +3793,7 @@ function solveDisplacement(problem, load, supp, elSupp;
     f = loadVector(problem, load)
     if condensed
         fixed = constrainedDoFs(problem, supp)
-        free = setdiff(1:problem.non*problem.pdim, fixed)
+        free = freeDoFs(problem, supp)
         u = copy(f)
         fill!(u.a, 0.0)
         applyBoundaryConditions!(u, supp)
@@ -5468,35 +5469,9 @@ Types:
 - `uy`: Float64 
 - `uz`: Float64 
 """
-function initialDisplacement(problem, name; ux=1im, uy=1im, uz=1im)
-    pdim = problem.pdim
-    dim = problem.dim
-    u0 = zeros(problem.non * problem.dim)
-    phg = getTagForPhysicalName(name)
-    nodeTags, coord = gmsh.model.mesh.getNodesForPhysicalGroup(-1, phg)
-    if ux != 1im
-        for i in 1:length(nodeTags)
-            u0[nodeTags[i]*dim-(dim-1)] = ux
-        end
-    end
-    if uy != 1im
[…]
