+++ src/general.jl
@@ -109,8 +109,8 @@ function elastic_constants(; E=nothing, ν=nothing, λ=nothing, μ=nothing, κ=n
 
     elseif n == 1 && haskey(specified, :κ)
         κ = specified[:κ]
-        ν = DEFAULT_POISSON
         E = DEFAULT_YOUNG
+        ν = (1 - E/3κ) / 2
         μ = E / (2*(1+ν))
         λ = 2μ*ν/(1-2ν)
         #κ = E / (3*(1-2ν))
@@ -711,6 +711,21 @@ Each scalar field provides one vector component.
 
 ---
 
+### From an existing VectorField
+
+```julia
+VectorField(s; steps, tmin, tmax, step)
+```
+
+Creates a new `VectorField` by **replicating a selected time step** of an
+existing field.
+This is useful for initializing a time-dependent field from a static solution.
+
+* `step` selects the time index of `s` to be replicated.
+* The new field contains `steps` identical time slices.
+
+---
+
 ## Notes
 
 * Time steps do not need to be uniformly spaced.
@@ -905,6 +920,21 @@ struct VectorField <: AbstractField
         a = [;;]
         t = [0.0]
         return new(A, a, t, numElem, nsteps, type, problem)
+    end
+    function VectorField(s::VectorField; steps=1, tmin=0.0, tmax=tmin+(steps-1), step=1)
+        if step > s.nsteps || step < 1
+            error("VectorField: `step` is out of range - 0 < $step < $(s.nsteps)")
+        end
+        A = Vector{Matrix{Float64}}(undef, length(s.numElem))
+        s1 = nodesToElements(s)
+        for j = 1:length(s1.numElem)
+            A[j] = zeros(size(s1.A[j], 1), steps)
+            for i in 1:steps
+                A[j][:, i] = s1.A[j][:, step]
+            end
+        end
+        t = range(start=tmin, stop=tmax, length=steps)
+        VectorField(A, [;;], t, s1.numElem, steps, s1.type, s1.model)
     end
 end
 
@@ -1230,6 +1260,84 @@ struct TensorField <: AbstractField
     end
 end
 
+"""
+    CoordinateSystem(vec1, vec2)
+
+A structure containing the data of a coordinate system.
+- `vec1`: direction of the new x axis.
+- `vec2`: together with `vec1` determine the xy plane
+If the problem is two dimensional, it is enough to give the first two
+elements of `vec1`. Elements of `vec1` and `vec2` can be functions. In
+3D case the functions have three arguments (x, y, and z coordinates),
+otherwise (in 2D case) the number of arguments is two (x and y coordinates).
+
+Types:
+- `vec1`: Vector{Float64}
+- `vec2`: Vector{Float64}
+
+# Examples
+
+```julia
+# 2D case
+nx(x, y) = x
+ny(x, y) = y
+cs = CoordinateSystem([nx, ny])
+Q = rotateNodes(problem, "body", cs)
+q2 = Q' * q1 # where `q1` is in Cartesian, `q2` is in Axisymmetric coordinate system and
+# `q1` is a nodal displacement vector.
+S2 = Q' * S1 * Q # where `S1` is a stress field in Cartesian coordinate system while
+# `S2` is in Axisymmetric coordinate system.
+
+# 3D case
+n1x(x, y, z) = x
[…]
