### 2026-01-11 6574663
solveDisplacement, solveTemperature
