+++ src/linear.jl
@@ -2884,7 +2884,7 @@ function elasticSupportMatrix(problem, elSupports)
         kx = kx !== nothing ? kx : 0
         ky = ky !== nothing ? ky : 0
         kz = kz !== nothing ? kz : 0
-        kx = elSupports[n].h !== nothing ? h : kx
+        kx = elSupports[n].h !== nothing ? elSupports[n].h : 0
         if problem.pdim == 3
             f = [0, 0, 0]
         elseif problem.pdim == 2
@@ -3668,6 +3668,16 @@ function applyBoundaryConditions!(field::Union{ScalarField,VectorField,TensorFie
     end
 end
 
+function applyBoundaryConditions!(
+    field::Union{ScalarField,VectorField,TensorField},
+    supports::AbstractVector
+)
+    isempty(supports) && return field
+    supports isa Vector{BoundaryCondition} ||
+        error("applyBoundaryConditions!: supports must be Vector{BoundaryCondition}")
+    return applyBoundaryConditions!(field, supports::Vector{BoundaryCondition})
+end
+
 """
     applyBoundaryConditions(problem::Problem, supports::Vector{BoundaryCondition})
                             
@@ -3731,7 +3741,13 @@ Types:
 - `f`: VectorField 
 - `q`: VectorField
 """
-function solveDisplacement(K, f)
+function solveDisplacement(K::SystemMatrix, f::VectorField; 
+                           support::Vector{BoundaryCondition}=BoundaryCondition[], 
+                           iterative=false,
+                           reltol::Real = sqrt(eps()),
+                           maxiter::Int = K.model.non * K.model.dim,
+                           preconditioner = Identity(),
+                           ordering=true)
     type = :null
     if f.type == :v3D
         type = :v3D
@@ -3740,7 +3756,23 @@ function solveDisplacement(K, f)
     else
         error("solveDisplacement: wrong type of 'f': ($(f.type))")
     end
-    return VectorField([], reshape(K.A \ f.a, :, 1), [0.0], [], 1, type, K.model)
+    problem = K.model
+
+    fixed = constrainedDoFs(problem, support)
+    free = freeDoFs(problem, support)
+    u = copy(f)
+    fill!(u.a, 0.0)
+    applyBoundaryConditions!(u, support)
+    f_kin = K.A[:, fixed] * u.a[fixed]
+    #u.a[free] = cholesky(Symmetric(K.A[free, free])) \ (f.a[free] - f_kin[free])
+    if iterative
+        u.a[free] = cg(K.A[free,free], f.a[free] - f_kin[free], Pl=preconditioner, reltol=reltol, maxiter=maxiter)
+    elseif ordering == false
+        u.a[free] = lu(K.A[free, free], q=nothing) \ (f.a[free] - f_kin[free])
+    else
+        u.a[free] = (K.A[free, free]) \ (f.a[free] - f_kin[free])
+    end
+    return u
 end
 
 """
@@ -3776,7 +3808,10 @@ Types:
 - `preconditioner`: preconditioner object (made by eg. `ilu` or `ichol`)
 - `q`: VectorField
 """
-function solveDisplacement(problem, load, supp, elSupp;
+function solveDisplacement(problem::Problem; 
+                           load::Vector{BoundaryCondition}=BoundaryCondition[], 
+                           support::Vector{BoundaryCondition}=BoundaryCondition[], 
+                           elasticSupport::Vector{BoundaryCondition}=BoundaryCondition[],
                            condensed=true,
                            iterative=false,
                            reltol::Real = sqrt(eps()),
@@ -3787,16 +3822,16 @@ function solveDisplacement(problem, load, supp, elSupp;
         return nothing
     end
     K = stiffnessMatrix(problem)
-    if elSupp ≠ []
[…]
