+++ src/heat.jl
@@ -657,7 +657,6 @@ Types:
 - `T₀`: ScalarField
 - `thermLoadVec`: VectorField
 """
-#function thermalLoadVector(problem, T; T₀=1im)
 function thermalLoadVector(problem, T; T₀=ScalarField([], zeros(problem.non, 1), [0], [], 1, :scalar, problem))
     if problem.type == :AxiSymmetric
         return thermalLoadVectorAXI(problem, T, T₀)
@@ -951,64 +950,65 @@ function applyHeatConvection!(heatCondMat, heatFluxVec, heatConv)
 end
 
 """
-    solveTemperature(K, q)
+    solveTemperature(K::SystemMatrix, q::ScalarField; temperatureConstraint::Vector{BoundaryCondition})
 
-Solves the equation K*T=q for the temperature field `T`. `K` is the heat conduction matrix,
-`q` is the heat flux vector.
+Solves the equation K*T=q for the temperature field `T` after applying `temperatureConstraint`. 
+`K` is the heat conduction matrix, `q` is the heat flux vector.
 
 Return: `T`
 
 Types:
 - `K`: SystemMatrix 
-- `q`: ScalarField 
+- `q`: ScalarField
+- `temperatureConstraint`: Vector{BoundaryCondition}
 - `T`: ScalarField
 """
-function solveTemperature(K::SystemMatrix, q::ScalarField)
-    return K \ q
+function solveTemperature(K::SystemMatrix, q::ScalarField, temperatureConstraint::Vector{BoundaryCondition}=BoundaryCondition[])
+    bc = constrainedDoFs(problem, temperatureConstraint)
+    free = freeDoFs(problem, temperatureConstraint)
+    T = applyBoundaryConditions(problem, temperatureConstraint)
+    q0 = K[:,bc] * DoFs(T)[bc]
+    DoFs(T)[free] = K[free,free] \ (DoFs(q)[free] - q0[free])
+    return T
 end
 
 """
-    solveTemperature(problem, flux, temp)
+    solveTemperature(problem::Problem; heatFlux::Vector{BoundaryCondition}=[], temperatureConstraint::Vector{BoundaryCondition}=[], heatConvection::Vector{BoundaryCondition}=[])
 
-Solves the temperature field `T` of `problem` with given heat flux `flux` and
-temperature `temp`.
+Solves the temperature field `T` of `problem` with given heat flux `heatFlux`,
+temperature `temperatureConstraint` and heat convection `heatConvection`.
 
 Return: `T`
 
 Types:
 - `problem`: Problem 
-- `flux`: Vector{Tuple} 
-- `temp`: Vector{Tuple}
+- `heatFlux`: Vector{BoundaryCondition} 
+- `temperatureConstraint`: Vector{BoundaryCondition}
+- `heatConvection`: Vector{BoundaryCondition}
 - `T`: ScalarField
 """
-function solveTemperature(problem, flux, temp)
+function solveTemperature(problem::Problem; 
+                          heatFlux::Vector{BoundaryCondition}=BoundaryCondition[], 
+                          temperatureConstraint::Vector{BoundaryCondition}=BoundaryCondition[], 
+                          heatConvection::Vector{BoundaryCondition}=BoundaryCondition[]
+                          )
+    bc = constrainedDoFs(problem, temperatureConstraint)
+    free = freeDoFs(problem, temperatureConstraint)
+    T = applyBoundaryConditions(problem, temperatureConstraint)
     K = heatConductionMatrix(problem)
-    q = heatFluxVector(problem, flux)
-    applyBoundaryConditions!(K, q, temp)
-    return K \ q
-end
[…]
