### 2026-01-12 37511b2
docstring for solveDisplacement and solveTemperature
