+++ src/linear.jl
@@ -3729,17 +3729,57 @@ function applyElasticSupport!(stiffMat::SystemMatrix, elastSupp)
 end
 
 """
-    solveDisplacement(K, q)
-                            
-Solves the equation K*q=f for the displacement vector `q`. `K` is the stiffness Matrix,
-`q` is the load vector.
-                            
-Return: `q`
-                            
-Types:
-- `K`: SystemMatrix 
-- `f`: VectorField 
-- `q`: VectorField
+    solveDisplacement(K, f; support = BoundaryCondition[],
+                      iterative = false,
+                      reltol = sqrt(eps()),
+                      maxiter = K.model.non * K.model.dim,
+                      preconditioner = Identity(),
+                      ordering = true)
+
+Solves the linear system
+
+    K * u = f
+
+for the displacement vector `u`, where `K` is a stiffness matrix and `f` is the
+load vector. Essential (Dirichlet) boundary conditions are imposed via the
+`support` argument.
+
+This is a **low-level solver** operating directly on a preassembled
+[`SystemMatrix`] and load [`VectorField`]. It assumes that the stiffness matrix
+already includes all material contributions (and optional elastic supports).
+
+If `iterative = true`, the reduced system is solved using the conjugate gradient
+method. Otherwise, a direct solver is used.
+
+# Arguments
+- `K::SystemMatrix`:
+  Assembled stiffness matrix associated with a [`Problem`].
+- `f::VectorField`:
+  Load vector.
+- `support::Vector{BoundaryCondition}` (keyword, optional):
+  Essential boundary conditions (Dirichlet-type constraints).
+  Default is an empty vector (no supports).
+- `iterative::Bool` (keyword):
+  If `true`, uses the conjugate gradient method.
+- `reltol::Real` (keyword):
+  Relative tolerance for the iterative solver.
+- `maxiter::Int` (keyword):
+  Maximum number of iterations for the iterative solver.
+- `preconditioner` (keyword):
+  Preconditioner object for iterative solution (e.g. from `ilu` or `ichol`).
+- `ordering::Bool` (keyword):
+  If `false`, disables column reordering in the direct solver
+  (`lu(A, q=nothing)`).
+
+# Returns
+- `u::VectorField`:
+  Displacement vector satisfying the prescribed boundary conditions.
+
+# Notes
+- Only the unconstrained degrees of freedom are solved for; constrained values
+  are imposed explicitly.
+- This function is primarily intended for internal or advanced use. Most users
+  should prefer the high-level `solveDisplacement(problem; ...)` interface.
 """
 function solveDisplacement(K::SystemMatrix, f::VectorField; 
                            support::Vector{BoundaryCondition}=BoundaryCondition[], 
@@ -3776,37 +3816,60 @@ function solveDisplacement(K::SystemMatrix, f::VectorField;
 end
 
 """
-    solveDisplacement(problem, load, supp, elSupp;
+    solveDisplacement(problem;
+                      load = BoundaryCondition[],
+                      support = BoundaryCondition[],
+                      elasticSupport = BoundaryCondition[],
                       condensed = true,
                       iterative = false,
-                      reltol::Real = sqrt(eps()),
-                      maxiter::Int = problem.non * problem.dim,
+                      reltol = sqrt(eps()),
+                      maxiter = problem.non * problem.dim,
                       preconditioner = Identity(),
-                      ordering=true)
-                            
-Computes the displacement vector `q` for the given `problem` subject to 
-loads `load`, supports `supp` and elastic supports `elSupp`. When `condensed` is `true`, the 
-reduced stiffness matrix and load vector are used in the solution.
[…]

[…]
