### 2026-01-13 8d6c9cf
bugfix - elasticSupportMatrix
