### 2026-01-14 1c25aa3
fix CDM+HHT+FDM
