### 2026-01-17 69fdaa6
ScalarField BC in time integrators
