+++ src/linear.jl
@@ -3014,13 +3014,14 @@ function elasticSupportMatrix(problem, elSupports)
     return SystemMatrix(C, problem)
 end
 
-@inline function _loadvec_helper(f, h, x, y, z, nnet, j, l)
+@inline function _loadvec_helper(f, h, x, y, z, nnet, j, l, nsteps)
     if f isa Number
-        return f
+        return fill(f, nsteps)
     elseif f isa Function
-        return f(x, y, z)
-    elseif f isa ScalarField
-        return h[:, j]' * f.a[nnet[l, :]]
+        return fill(f(x, y, z), nsteps)
+    elseif f isa ScalarField && isNodal(f)
+        v = h[:, j]' * f.a[nnet[l, :],:]
+        return vec(v)
     else
         error("loadVector: internal error.")
     end
@@ -3075,11 +3076,11 @@ function loadVector(problem, loads)
     b = problem.thickness
     non = problem.non
     dof = pdim * non
-    fp = zeros(dof)
     ncoord2 = zeros(3 * problem.non)
     f = nothing
+    fp = nothing
+    nsteps = 1
     for n in 1:length(loads)
-        #name, fx, fy, fz = loads[n]
         name = loads[n].phName
         fx = loads[n].fx
         fy = loads[n].fy
@@ -3101,10 +3102,35 @@ function loadVector(problem, loads)
         fzy = loads[n].fzy
         fxz = loads[n].fxz
 
+        fx  = (fx  isa ScalarField && isElementwise(fx))  ? elementsToNodes(fx)  : fx
+        fy  = (fy  isa ScalarField && isElementwise(fy))  ? elementsToNodes(fy)  : fy
+        fz  = (fz  isa ScalarField && isElementwise(fz))  ? elementsToNodes(fz)  : fz
+        T   = (T   isa ScalarField && isElementwise(T))   ? elementsToNodes(T)   : T
+        p   = (p   isa ScalarField && isElementwise(p))   ? elementsToNodes(p)   : p
+        hc  = (hc  isa ScalarField && isElementwise(hc))  ? elementsToNodes(hc)  : hc
+        T0  = (T0  isa ScalarField && isElementwise(T0))  ? elementsToNodes(T0)  : T0
+        qn  = (qn  isa ScalarField && isElementwise(qn))  ? elementsToNodes(qn)  : qn
+        hs  = (hs  isa ScalarField && isElementwise(hs))  ? elementsToNodes(hs)  : hs
+        q   = (q   isa ScalarField && isElementwise(q))   ? elementsToNodes(q)   : q
+        fxy = (fxy isa ScalarField && isElementwise(fxy)) ? elementsToNodes(fxy) : fxy
+        fyz = (fyz isa ScalarField && isElementwise(fyz)) ? elementsToNodes(fyz) : fyz
+        fzx = (fzx isa ScalarField && isElementwise(fzx)) ? elementsToNodes(fzx) : fzx
+        fyx = (fyx isa ScalarField && isElementwise(fyx)) ? elementsToNodes(fyx) : fyx
+        fzy = (fzy isa ScalarField && isElementwise(fzy)) ? elementsToNodes(fzy) : fzy
+        fxz = (fxz isa ScalarField && isElementwise(fxz)) ? elementsToNodes(fxz) : fxz
+    
+        nsteps = 1
+        for i in [fx, fy , fz, T,  p, hc, T0, qn, hs, q, fxy, fyz, fzx, fyx, fzy, fxz]
+            if i isa ScalarField
+                nsteps = max(nsteps, i.nsteps)
+            end
+        end
+        fp = zeros(dof, nsteps)
+
[…]
