+++ src/heat.jl
@@ -1329,38 +1329,204 @@ end
 =#
 
 """
-    FDM(K, C, q, T0, tₘₐₓ, Δt; ϑ=...)
+    FDM(K, C, q, bc, T0, n, Δt; ϑ=0.5)
 
-Solves a transient heat conduction problem using Finite Difference Method (FDM).
-Introducing a `ϑ` parameter, special cases can be used as the Forward Euler (explicit, ϑ=0),
-Backward Euler (implicit, ϑ=1), Crank-Nicolson (ϑ=0.5) and intermediate cases (0<ϑ<1).
-(This method is known as ϑ-method. See [^5].)
-`K` is the heat conduction matrix, `C` is the heat capacity matrix,
-`q` is the heat flux vector, `T0` is the initial temperature, `tₘₐₓ` is the upper 
-bound of the time intervall (lower bound is zero)
-and `Δt` is the time step size. Returns the nodal temperature vectors in each time 
-step arranged in the columns of the matrix `T`
-and a vector `t` of the time instants used.
+Solves a transient diffusion-type problem (e.g. heat conduction) using the
+finite difference method in time (ϑ-method).
 
-The critical (largest allowed) time step is `Δtₘₐₓ = 2 / ((1-2ϑ)*λₘₐₓ)`
-where `λₘₐₓ` is the largest eigenvalue of (**K**+λ**C**)**θ**=**0** 
-eigenvalue problem and `ϑ` is the parameter of the ϑ-method. Default value of `ϑ`
-is 1/2.
+The semi-discrete system
+```
 
-[^5]: Bathe, K. J.: Finite element procedures, Wiley, 1983, <https://doi.org/10.1002/nag.1610070412>
+C * Ẋ(t) + K * X(t) = q(t)
 
-Return: `T`
+```
+is integrated in time using the ϑ-scheme:
+- ϑ = 0   : Forward Euler (explicit)
+- ϑ = 1/2 : Crank–Nicolson
+- ϑ = 1   : Backward Euler (implicit)
+- 0 < ϑ < 1 : intermediate schemes
 
-Types:
-- `K`: SystemMatrix
-- `C`: SystemMatrix
-- `q`: ScalarField
-- `T0`: ScalarField
-- `tₘₐₓ`: Float64
-- `Δt`: Float64 
-- `T`: ScalarField
+The method supports:
+- time-independent or time-dependent load vectors `q`
+- time-independent or time-dependent Dirichlet boundary conditions
+- `ScalarField` and `VectorField` unknowns
+- consistent treatment of constraint-induced volume terms
+  (i.e. contributions of prescribed values appear automatically in the RHS)
+
+Boundary conditions are applied *solver-side*:
+on constrained DOFs the prescribed values override the initial condition,
+while on free DOFs the initial condition `T0` is used.
+
+If `q.nsteps == 1`, the load is treated as time-independent.
+If `q.nsteps == n`, a true ϑ-weighted load
+```
+
+q^{n+ϑ} = (1-ϑ) q^n + ϑ q^{n+1}
+
+```
+is used.
+
+For `ϑ = 0` and diagonal `C`, a fully explicit update is used.
+
+---
+
+### Arguments
+- `K::SystemMatrix`  
+  Diffusion (conductivity / stiffness) matrix.
+- `C::SystemMatrix`  
+  Capacity (mass / heat capacity) matrix.
+- `q::Union{ScalarField,VectorField}`  
+  Load / source vector. May be time-independent (`nsteps = 1`)
+  or time-dependent (`nsteps = n`).
+- `bc::Vector{BoundaryCondition}`  
[…]

[…]
