### 2026-01-18 6126eb8
final corrections
hopefully everything is working
