+++ src/linear.jl
@@ -9,6 +9,7 @@ export initialDisplacement, initialDisplacement!, initialVelocity, initialVeloci
 export nodalForce!, nodalAcceleration!
 export largestPeriodTime, smallestPeriodTime, largestEigenValue, smallestEigenValue
 export CDM, HHT, CDMaccuracyAnalysis, HHTaccuracyAnalysis
+export printElasticConstantsTable
 
 # linear.jl tetején
 mutable struct SolidWS
@@ -3078,7 +3079,7 @@ function loadVector(problem, loads)
     dof = pdim * non
     ncoord2 = zeros(3 * problem.non)
     f = nothing
-    fp = nothing
+    fp = zeros(dof,1)
     nsteps = 1
     for n in 1:length(loads)
         name = loads[n].phName
@@ -3313,299 +3314,6 @@ function loadVector(problem, loads)
     end
 end
 
-"""
-    applyBoundaryConditions!(stiffMat, loadVec, supports)
-                            
-Applies displacement boundary conditions `supports` on a stiffness matrix
-`stiffMat` and load vector `loadVec`. Mesh details are in `problem`. `supports`
-is a tuple of `name` of physical group and prescribed displacements `ux`, `uy`
-and `uz`.
-                            
-Returns: nothing
-                            
-Types:
-- `stiffMat`: SystemMatrix 
-- `loadVec`: VectorField
-- `supports`: Vector{Tuple{String, Float64, Float64, Float64}}
-"""
-function applyBoundaryConditions!(stiffMat::SystemMatrix, loadVec, supports)
-    if !isa(supports, Vector)
-        error("applyBoundaryConditions!: supports are not arranged in a vector. Put them in [...]")
-    end
-    dof, dof = size(stiffMat.A)
-    massMat = SystemMatrix(spzeros(dof, dof), stiffMat.model)
-    dampMat = SystemMatrix(spzeros(dof, dof), stiffMat.model)
-    applyBoundaryConditions!(stiffMat, massMat, dampMat, loadVec, supports)
-    #massMat.A *= 0
-    #dampMat.A *= 0
-    return
-end
-
-"""
-    applyBoundaryConditions(stiffMat, loadVec, supports)
-                            
-Applies displacement boundary conditions `supports` on a stiffness matrix
-`stiffMat` and load vector `loadVec`. Mesh details are in `problem`. `supports`
-is a tuple of `name` of physical group and prescribed displacements `ux`, `uy`
-and `uz`. Creates a new stiffness matrix and load vector.
-                            
-Returns: `stiffMat`, `loadVec`
-                            
-Types:
-- `stiffMat`: SystemMatrix 
-- `loadVec`: VectorField
-- `supports`: Vector{Tuple{String, Float64, Float64, Float64}}
-"""
-function applyBoundaryConditions(stiffMat0, loadVec0, supports)
-    problem = stiffMat0.model
-    if !isa(supports, Vector)
-        error("applyBoundaryConditions: supports are not arranged in a vector. Put them in [...]")
-    end
-    dof, dof = size(stiffMat0.A)
-    massMat = SystemMatrix(spzeros(dof, dof), stiffMat0.model)
-    dampMat = SystemMatrix(spzeros(dof, dof), stiffMat0.model)
-    stiffMat = copy(stiffMat0)
-    loadVec = copy(loadVec0)
-    applyBoundaryConditions!(stiffMat, massMat, dampMat, loadVec, supports)
-    #massMat.A = []
-    #dampMat.A = []
-    return stiffMat, loadVec
-end
-
-"""
-    applyBoundaryConditions!(heatCondMat, heatCapMat, heatFluxVec, supports)
-                            
[…]
