+++ src/linear.jl
@@ -3704,6 +3704,211 @@ Types:
 - `q`: VectorField
 - `E`: TensorField
 """
+function solveStrain(q; DoFResults=false)
+    problem = q.model
+    gmsh.model.setCurrent(problem.name)
+    
+    if !isa(q, VectorField) 
+        error("solveStrain:argument must be a VectorField. Now it is '$(typeof(q))'.")
+    end
+    if q.type != :v3D && q.type != :v2D
+        error("solveStrain: argument must be a displacement vector. Now it is '$(q.type)'.")
+    end
+    if q.A != []
+        error("solveStrain: q.A != []")
+    end
+    nsteps = q.nsteps
+    ε = []
+    numElem = Int[]
+    ncoord2 = zeros(3 * problem.non)
+    dim = problem.dim
+    pdim = problem.pdim
+    non = problem.non
+    type = :e
+    if DoFResults == true
+        E1 = zeros(non * 9, nsteps)
+        pcs = zeros(Int64, non * dim)
+    end
+    
+    for ipg in 1:length(problem.material)
+        phName = problem.material[ipg].phName
+        ν = problem.material[ipg].ν
+        dim = 0
+        if problem.dim == 3 && problem.type == :Solid
+            dim = 3
+            rowsOfB = 6
+            b = 1
+        elseif problem.dim == 2 && problem.type == :PlaneStress
+            dim = 2
+            rowsOfB = 3
+            b = problem.thickness
+        elseif problem.dim == 2 && problem.type == :PlaneStrain
+            dim = 2
+            rowsOfB = 3
+            b = 1
+        elseif problem.dim == 2 && problem.type == :AxiSymmetric
+            dim = 2
+            rowsOfB = 4
+            b = 1
+        else
+            error("solveStrain: dimension is $(problem.dim), problem type is $(problem.type).")
+        end
+        
+        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+        for idm in 1:length(dimTags)
+            dimTag = dimTags[idm]
+            edim = dimTag[1]
+            etag = dimTag[2]
+            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
+            #nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes(dim, -1, true, false)
+            nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes()
+            ncoord2[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
+            ncoord2[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
+            ncoord2[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
+            for i in 1:length(elemTypes)
+                et = elemTypes[i]
+                elementName, dim, order, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(et)
+                #e0 = zeros(rowsOfB * numNodes)
+                nodeCoord = zeros(numNodes * 3)
+                for k in 1:dim, j = 1:numNodes
+                    nodeCoord[k+(j-1)*3] = localNodeCoord[k+(j-1)*dim]
+                end
+                comp, dfun, ori = gmsh.model.mesh.getBasisFunctions(et, nodeCoord, "GradLagrange")
+                ∇h = reshape(dfun, :, numNodes)
+                comp, fun, ori = gmsh.model.mesh.getBasisFunctions(et, nodeCoord, "Lagrange")
+                h = reshape(fun, :, numNodes)
+                nnet = zeros(Int, length(elemTags[i]), numNodes)
[…]
