### 2026-01-20 c0dca05
Navier-Stokes operators
