+++ src/poisson.jl
@@ -3,6 +3,7 @@ export poissonMatrix, reactionMatrix, advectionMatrix
 export ∫∇oN_c_∇oN_dΩ, ∫∇N_c_∇N_dΩ, ∫∇xN_c_∇xN_dΩ, ∫N_c_N_dΩ, ∫N_c_∂N∂x_dΩ, ∫N_c_∂N∂x_dΩ, ∫N_c_∂N∂y_dΩ, ∫N_c_∂N∂z_dΩ
 export ∫N_c_dΩ
 export gradDivMatrix, symmetricGradientMatrix, curlCurlMatrix
+export gradMatrix, navierStokesAdvectionMatrix
 export tensorLaplaceMatrix, traceLaplaceMatrix, beltramiMichellMatrix, tensorDivDivMatrix
 export sourceVector, loadTensor
 export solveField
@@ -1413,3 +1414,560 @@ function solveField(K::SystemMatrix, f::Union{ScalarField,VectorField};
     end
     return u
 end
+
+"""
+    gradMatrix(problem_u::Problem, problem_p::Problem)
+
+Assembles the mixed gradient matrix G mapping a scalar field (pressure)
+to a vector field (velocity), corresponding to the weak form
+
+    (G p, v) = -∫_Ω p ∇·v dΩ
+
+Notes
+- Requires problem_u.pdim == problem_u.dim (vector field).
+- Requires problem_p.pdim == 1 (scalar field).
+- Requires problem_u.dim == problem_p.dim.
+- The divergence matrix is obtained as D = -G'.
+"""
+function gradMatrix(problem_u::Problem, problem_p::Problem)
+
+    @assert problem_u.dim == problem_p.dim
+    @assert problem_u.pdim == problem_u.dim
+    @assert problem_p.pdim == 1
+
+    gmsh.model.setCurrent(problem_u.name)
+
+    dim   = problem_u.dim
+    pdim  = problem_u.pdim
+    non_u = problem_u.non
+    non_p = problem_p.non
+
+    # Conservative IJV estimate
+    lengthOfIJV = estimateLengthOfIJV(problem_u)
+
+    I = Vector{Int}(undef, lengthOfIJV)
+    J = Vector{Int}(undef, lengthOfIJV)
+    V = Vector{Float64}(undef, lengthOfIJV)
+    pos = 1
+
+    for ipg in 1:length(problem_u.material)
+        phName = problem_u.material[ipg].phName
+        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
+
+        for (edim, etag) in dimTags
+            elemTypes, elemTags, elemNodeTags =
+                gmsh.model.mesh.getElements(edim, etag)
+
+            for itype in 1:length(elemTypes)
+                et = elemTypes[itype]
+
+                _, _, order, numNodes::Int, _, _ =
+                    gmsh.model.mesh.getElementProperties(et)
+
+                intPoints, intWeights =
+                    gmsh.model.mesh.getIntegrationPoints(
+                        et, "Gauss" * string(2order + 1)
+                    )
+
+                numIntPoints = length(intWeights)
+
+                # --- pressure basis (scalar) ---
+                _, fun_p, _ =
+                    gmsh.model.mesh.getBasisFunctions(et, intPoints, "Lagrange")
+                h_p = reshape(fun_p, :, numIntPoints)
+
+                # --- velocity basis (vector) ---
+                _, dfun_u, _ =
+                    gmsh.model.mesh.getBasisFunctions(et, intPoints, "GradLagrange")
+                ∇h_u = reshape(dfun_u, :, numIntPoints)
+
+                nnet = zeros(Int, length(elemTags[itype]), numNodes)
+
+                invJac = zeros(3, 3numIntPoints)
[…]
