### 2026-01-21 8f47867
Large deformations with Poisson-type operators
