+++ src/poisson.jl
@@ -1971,3 +1971,506 @@ function _assemble_ns_advection!(
     return pos
 end
 =#
+
+
+function poissonMatrixVector(
+    problem::Problem;
+    coefficient::TensorField
+)
+    @assert problem.pdim == problem.dim
+    gmsh.model.setCurrent(problem.name)
+
+    dim  = problem.dim
+    pdim = problem.pdim
+    non  = problem.non
+    dof  = pdim * non
+
+    # elementwise tensor field
+    S = nodesToElements(coefficient)
+    Se = Dict(zip(S.numElem, S.A))
+
+    lengthOfIJV = estimateLengthOfIJV(problem)
+    I = Vector{Int}(undef, lengthOfIJV)
+    J = Vector{Int}(undef, lengthOfIJV)
+    V = Vector{Float64}(undef, lengthOfIJV)
+    pos = 1
+
+    for mat in problem.material
+        for (edim, etag) in gmsh.model.getEntitiesForPhysicalName(mat.phName)
+
+            elemTypes, elemTags, elemNodeTags =
+                gmsh.model.mesh.getElements(edim, etag)
+
+            for it in eachindex(elemTypes)
+                et = elemTypes[it]
+                _, _, order, numNodes, _, _ =
+                    gmsh.model.mesh.getElementProperties(et)
+
+                intPoints, intWeights =
+                    gmsh.model.mesh.getIntegrationPoints(et, "Gauss" * string(2order + 1))
+
+                _, _, _ = gmsh.model.mesh.getBasisFunctions(et, intPoints, "Lagrange")
+                _, dfun, _ =
+                    gmsh.model.mesh.getBasisFunctions(et, intPoints, "GradLagrange")
+
+                ∇h = reshape(dfun, :, length(intWeights))
+
+                for (e, elem) in enumerate(elemTags[it])
+                    nodeTags = elemNodeTags[it][(e-1)*numNodes+1 : e*numNodes]
+
+                    jac, jacDet, _ =
+                        gmsh.model.mesh.getJacobian(elem, intPoints)
+                    Jac = reshape(jac, 3, :)
+
+                    Ke = zeros(pdim*numNodes, pdim*numNodes)
+
+                    # tensor at element nodes (9×numNodes)
+                    Snode = Se[elem]
+
+                    for k in eachindex(intWeights)
+                        invJ = inv(Jac[1:dim, 3k-2:3k])'
+                        w = jacDet[k] * intWeights[k]
+
+                        # interpolate S to GP (Voigt-free, full tensor)
+                        Sgp = zeros(dim, dim)
+                        for a in 1:numNodes
+                            Sa = reshape(Snode[9a-8:9a, 1], dim, dim)
+                            Sgp .+= Sa * (1.0 / numNodes)   # lumped interp
+                        end
+
+                        for a in 1:numNodes, b in 1:numNodes
+                            ∇Na = invJ * ∇h[3a-2:3a-(3-dim), k]
+                            ∇Nb = invJ * ∇h[3b-2:3b-(3-dim), k]
+
+                            for i in 1:dim, j in 1:dim
+                                ia = (a-1)*pdim + i
+                                ib = (b-1)*pdim + j
+                                Ke[ia, ib] += ∇Na[i] * Sgp[i,j] * ∇Nb[j] * w
+                            end
+                        end
+                    end
+
+                    for a in 1:pdim*numNodes, b in 1:pdim*numNodes
+                        I[pos] = (nodeTags[div(a-1,pdim)+1]-1)*pdim + mod(a-1,pdim)+1
+                        J[pos] = (nodeTags[div(b-1,pdim)+1]-1)*pdim + mod(b-1,pdim)+1
[…]
