+++ doc/nonlinear-poisson-notes.tex
@@ -1,0 +1,249 @@
+\documentclass[a4paper,11pt]{article}
+
+\usepackage{amsmath,amssymb}
+\usepackage{bm}
+\usepackage{geometry}
+\usepackage{hyperref}
+
+\geometry{margin=2.5cm}
+
+\title{Total Lagrange formuláció a LowLevelFEM-ben\\
+	Gyenge alak és operátor-alapú implementáció}
+\author{}
+\date{}
+
+\begin{document}
+	\maketitle
+	
+	\section{Kiinduló probléma (Total Lagrange)}
+	
+	A referencia-konfigurációban ($\Omega_0$) felírt egyensúlyi egyenlet:
+	\[
+	\mathrm{Div}\,\bm P + \bm b_0 = \bm 0
+	\quad \text{in } \Omega_0
+	\]
+	
+	ahol
+	\[
+	\bm P = \bm F \bm S,
+	\qquad
+	\bm F = \bm I + \nabla_0 \bm u.
+	\]
+	
+	Peremfeltételek:
+	\[
+	\bm u = \bar{\bm u} \quad \text{on } \Gamma_{u,0},
+	\qquad
+	\bm P \bm N = \bm t_0 \quad \text{on } \Gamma_{t,0}.
+	\]
+	
+	---
+	
+	\section{Gyenge alak}
+	
+	A virtuális elmozdulás $\delta\bm u$ alkalmazásával:
+	\[
+	\int_{\Omega_0} \delta\bm F : \bm P \, \mathrm d\Omega_0
+	=
+	\int_{\Omega_0} \delta\bm u \cdot \bm b_0 \, \mathrm d\Omega_0
+	+
+	\int_{\Gamma_{t,0}} \delta\bm u \cdot \bm t_0 \, \mathrm d\Gamma_0,
+	\]
+	ahol
+	\[
+	\delta \bm F = \nabla_0 \delta \bm u.
+	\]
+	
+	Ez az egyenlet a nemlineáris reziduál:
+	\[
+	\bm r(\bm u) = \bm f_{\mathrm{int}}(\bm u) - \bm f_{\mathrm{ext}}(\bm u) = \bm 0.
+	\]
+	
+	---
+	
+	\section{Newton-iteráció}
+	
+	Newton-módszerrel:
+	\[
+	\bm K(\bm u^k)\,\Delta \bm u
+	=
+	\bm f_{\mathrm{ext}}(\bm u^k)
+	-
+	\bm f_{\mathrm{int}}(\bm u^k),
+	\]
+	
+	ahol a teljes tangens:
+	\[
+	\bm K =
+	\bm K_{\mathrm{mat}}
+	+
+	\bm K_{\mathrm{geo}}
+	-
+	\bm K_{\mathrm{ext}}.
+	\]
+	
+	---
+	
+	\section{Operátor-alapú bontás a LowLevelFEM-ben}
+	
+	A LowLevelFEM-ben a gyenge alak térfogati és felületi tagjai
+	különálló, "buta" operátorokként jelennek meg.
+	A nemlinearitás kizárólag a mezőkön (pl.\ $\bm F$, $\bm S$, $\bm P$) keresztül lép be.
+	
+	\subsection{\texttt{poissonMatrixVector}}
+	
+	\paragraph{Gyenge alakbeli eredet}
+	\[
+	\int_{\Omega_0}
+	\nabla_0 \delta \bm u : \bm S : \nabla_0 \bm u
+	\,\mathrm d\Omega_0
+	\]
+	
+	\paragraph{Szerep}
+	\begin{itemize}
+		\item geometriai merevség,
+		\item II.~PK-feszültséggel súlyozott gradiens--gradiens tag,
+		\item térfogati integrál.
+	\end{itemize}
+	
+	\paragraph{Megjegyzés}
[…]

[…]
