+++ src/LowLevelFEM.jl
@@ -27,7 +27,7 @@ include("extra.jl")
 
 export @showfields, @showstruct, @showdef, @showtype, @showmem, @showmethods, @disp, @showsize
 export probe_field
-#=
+
 @setup_workload begin
     @compile_workload begin
         mat = material("dummy")
@@ -119,5 +119,5 @@ export probe_field
         traceLaplaceMatrix(prob)
     end
 end
-=#
+
 end #module
