### 2026-01-28 dfb90db
Large deformation - ready :-)
next: hyperelastic materials
