+++ src/linear.jl
@@ -3015,6 +3015,31 @@ function elasticSupportMatrix(problem, elSupports)
     return SystemMatrix(C, problem)
 end
 
+
+@inline function _rotation_from_F(F::AbstractMatrix{<:Real})
+    U, _, Vt = svd(F)
+    R = U * Vt
+    if det(R) < 0
+        U[:, end] .*= -1
+        R = U * Vt
+    end
+    return R
+end
+
+@inline function _loadvec_helper(f, h, x, y, z, nnet, j, l, nsteps)
+    if f isa Number
+        return fill(f, nsteps)
+    elseif f isa Function
+        return fill(f(x, y, z), nsteps)
+    elseif f isa ScalarField && isNodal(f)
+        v = h[:, j]' * f.a[nnet[l, :], :]
+        return vec(v)
+    else
+        error("loadVector: internal error.")
+    end
+end
+
+#=
 @inline function _loadvec_helper(f, h, x, y, z, nnet, j, l, nsteps)
     if f isa Number
         return fill(f, nsteps)
@@ -3027,9 +3052,10 @@ end
         error("loadVector: internal error.")
     end
 end
+=#
 
 """
-    loadVector(problem, loads)
+    loadVector(problem, loads; F=nothing)
                             
 Assembles the right-hand-side vector associated with external mechanical loads
 in weak-form finite element problems.
@@ -3058,6 +3084,9 @@ appropriate geometric factor (e.g. `2πr`) in the load definition or coefficient
 
 Returns a `VectorField` for vector-valued problems (`pdim > 1`) and a `ScalarField`
 for scalar problems (`pdim == 1`).
+
+If `F` is the deformation gradient (a `TensorField`), then the function solves the follower load
+for large displacement problems (in 3D).
                             
 Return: `loadVec`
                             
@@ -3065,8 +3094,310 @@ Types:
 - `problem`: Problem
 - `loads`: Vector{BoundaryCondition}
 - `loadVec`: VectorField
+- `F`: Union{Nothing,TensorField}
 """
-function loadVector(problem, loads)
+function loadVector(problem, loads; F=nothing)
+    gmsh.model.setCurrent(problem.name)
+    if !isa(loads, Vector)
+        error("loadVector: loads are not arranged in a vector. Put them in [...]")
+    end
+
+    Fe = nothing
+    Fmap = nothing
+    if F !== nothing
+        Fe = nodesToElements(F)
+        Fmap = Dict(zip(Fe.numElem, Fe.A))
+    end
+
+    pdim = problem.pdim
+    DIM = problem.dim
+    b = problem.thickness
+    non = problem.non
+    dof = pdim * non
+    ncoord2 = zeros(3 * problem.non)
+    f = nothing
+    fp = zeros(dof,1)
+    nsteps = 1
+    for n in 1:length(loads)
+        name = loads[n].phName
+        fx = loads[n].fx
+        fy = loads[n].fy
+        fz = loads[n].fz
+        T = loads[n].T
+        p = loads[n].p
+        hc = loads[n].h
+        T0 = loads[n].T
+        qn = loads[n].qn
+        hs = loads[n].h
+
+        q = loads[n].q
+
+        fxy = loads[n].fxy
+        fyz = loads[n].fyz
+        fzx = loads[n].fzx
+        
+        fyx = loads[n].fyx
+        fzy = loads[n].fzy
+        fxz = loads[n].fxz
+
+        fx  = (fx  isa ScalarField && isElementwise(fx))  ? elementsToNodes(fx)  : fx
+        fy  = (fy  isa ScalarField && isElementwise(fy))  ? elementsToNodes(fy)  : fy
[…]
