+++ src/nonlinear.jl
@@ -4,6 +4,7 @@ export equivalentNodalForce, nonFollowerLoadVector
 export applyDeformationBoundaryConditions!, suppressDeformationAtBoundaries!, suppressDeformationAtBoundaries
 export solveDeformation, showDeformationResults
 export grad_xy
+export materialTangentMatrix, initialStressMatrix, externalTangentFollower, internalForceVector
 
 """
     nodePositionVector(problem)
@@ -2159,3 +2160,1380 @@ end
 function showDeformationResults(r::VectorField; name="vector", visible=false)
     return showDeformationResults(r, :vector, name=name, visible=visible)
 end
+
+#############################################################
+# New approach ##############################################
+#############################################################
+
+"""
+    materialTangentMatrix(problem::Problem; F::TensorField, C::AbstractMatrix)
+
+Assembles the **material (constitutive) tangent stiffness matrix**
+for a 3D solid under **finite deformation (Total Lagrange formulation)**,
+using a user-provided deformation gradient field `F` and a material
+tangent matrix `C` given in **6×6 Mandel/Voigt form**.
+
+This function builds the matrix
+```
+
+K_mat = ∫_Ω B(F)ᵀ · C · B(F) dΩ
+
+```
+where `B(F)` is the nonlinear strain–displacement matrix associated with
+the Green–Lagrange strain, and `C` is the material tangent
+∂S/∂E expressed in Mandel notation.
+
+The routine is **material-agnostic**: it does not assume any specific
+constitutive law. The material behavior is entirely defined by the
+supplied `C`, which may be spatially varying.
+
+---
+
+## Arguments
+
+- `problem::Problem`  
+  The finite element problem definition.  
+  Must satisfy:
+  - `problem.dim == 3`
+  - `problem.pdim == 3`
+
+---
+
+## Keyword arguments
+
+- `F::TensorField`  
+  Nodal deformation gradient field.
+  Each node stores the full 3×3 deformation gradient `F`,
+  which is interpolated to Gauss points during integration.
+
+- `C::AbstractMatrix` (size `6×6`)  
+  Material tangent matrix in Mandel/Voigt notation.
+
+  Each entry `C[i,j]` may be either:
+  - a `Number` (constant material tangent), or
+  - a `ScalarField` (spatially varying material tangent component).
+
+  All entries must satisfy:
+```
+
+C[i,j] isa Number || C[i,j] isa ScalarField
+
+```
+---
+
+## Mathematical formulation
+
+- Strain measure: **Green–Lagrange strain**
+```
+
+E = 1/2 (FᵀF − I)
+
+```
+- Variation:
+```
+
+δE = sym(Fᵀ δF)
+
+```
+- Tangent contribution:
+```
+
+δS = C : δE
+
+```
+- Element stiffness contribution:
+```
+
+K_e = ∫ B(F)ᵀ · C · B(F) dΩ
+
+```
+The strain–displacement matrix `B(F)` is constructed consistently with
+Mandel notation (shear components scaled by 2).
+
+---
+
+## Notes
+
+- This function assembles **only the material part** of the tangent
+stiffness matrix.  
+The **geometric stiffness** (stress-dependent part) must be assembled
+separately.
+
+- The formulation is suitable for:
+- hyperelastic materials,
+- finite rotations,
+- finite strains,
+provided that `C` is consistent with the stress measure used elsewhere.
+
+- The material tangent `C` is evaluated at Gauss points by interpolating
+nodal `ScalarField` values when necessary.
+
+---
+
+## Returns
+
+- `SystemMatrix`  
+Sparse global material tangent stiffness matrix of size `(ndof × ndof)`.
+
+---
+
+## Typical usage
+
[…]
