### 2026-01-29 e5a4431
plan for free energy function
