+++ Project.toml
@@ -36,3 +36,9 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test"]
+
+[weakdeps]
+Tensors = "48a634ad-e948-5137-8d70-aa71f2a747f4"
+
+[extensions]
+LowLevelFEMTensorsExt = ["Tensors"]
