### 2026-01-29 d17f2a1
free energy based material tangent and initial stress matrix is working :-)
