+++ ext/LowLevelFEMTensorsExt.jl
@@ -2,33 +2,62 @@ module LowLevelFEMTensorsExt
 
 using LowLevelFEM
 using Tensors
+using StaticArrays
 
-# Ide később jönnek majd:
-# - Tensor/Matrix konverziók
-# - Gauss-ponti stress/tangent kernelfüggvények
-# - hyperelastic_from_energy implementáció (ha akarod)
+# =========================
+# Stress (PK2)
+# =========================
+function LowLevelFEM._stress_from_energy(
+    ψ,
+    Cmat::SMatrix{3,3,Float64},
+    p
+)
+    # SMatrix -> Tensor
+    C = symmetric(Tensor{2,3}(Cmat))
 
-function neo_hooke_energy(C::SymmetricTensor{2,3}, μ, K)
-    I1 = tr(C)
-    J = sqrt(det(C))
-    return μ / 2 * (I1 / J^(2 / 3) - 3) + K / 2 * (log(J))^2
+    # AD: first derivative only
+    S = 2 * gradient(C -> ψ(C, p), C)
+
+    # Tensor -> SMatrix
+    return @SMatrix [
+        S[1,1] S[1,2] S[1,3];
+        S[2,1] S[2,2] S[2,3];
+        S[3,1] S[3,2] S[3,3]
+    ]
 end
 
-function PK2_stress(C::SymmetricTensor{2,3}, μ, K)
-    return 2 * gradient(C -> neo_hooke_energy(C, μ, K), C)
+
+# =========================
+# Tangent (material)
+# =========================
+function LowLevelFEM._tangent_from_energy(
+    ψ,
+    Cmat::SMatrix{3,3,Float64},
+    p
+)
+    C = symmetric(Tensor{2,3}(Cmat))
+
+    # AD: second derivative only
+    C4 = 4 * hessian(C -> ψ(C, p), C)
+
+    # ---- convert 4th order tensor to Voigt 6x6 SMatrix ----
+    return tensor4_to_voigt(C4)
 end
 
-function material_tangent(C::SymmetricTensor{2,3}, μ, K)
-    return 4 * hessian(C -> neo_hooke_energy(C, μ, K), C)
+const voigt = (
+    (1,1),
+    (2,2),
+    (3,3),
+    (2,3),
+    (1,3),
+    (1,2)
+)
+
+@inline function tensor4_to_voigt(C4)
+    @SMatrix [
+        C4[i,j,k,l] for (i,j) in voigt, (k,l) in voigt
+    ]
 end
 
-function LowLevelFEM._tensor_extension_test()
-    F = Tensor{2,3}((1.2, 0.3, 0.0,
-        0.0, 1.1, 0.0,
-        0.0, 0.0, 1.0))
-
-    C = symmetric(F' ⋅ F)
-    return tr(C)
-end
 
 end # module
