+++ src/nonlinear.jl
@@ -2306,12 +2306,21 @@ K = Kint + Kmat
 function materialTangentMatrix(
     problem::Problem;
     F::TensorField,
-    C::AbstractMatrix)
+    C::Union{AbstractMatrix,Nothing}=nothing,
+    energy::Union{Nothing,Function}=nothing,
+    params=nothing)
     
+    @assert xor(C !== nothing, energy !== nothing) """
+    Either provide `C` (6×6, Number or ScalarField entries),
+    or provide (`energy`, `params`) to compute the tangent at Gauss points.
+    """
+
     @assert problem.dim == 3
     @assert problem.pdim == 3
-    @assert size(C) == (6, 6)
-    @assert all(x -> (x isa Number) || (x isa ScalarField), C)
+    if C !== nothing
+        @assert size(C) == (6, 6)
+        @assert all(x -> (x isa Number) || (x isa ScalarField), C)
+    end
 
     gmsh.model.setCurrent(problem.name)
 
@@ -2325,14 +2334,17 @@ function materialTangentMatrix(
 
     # --- Preprocess C entries:
     # For ScalarField entries: build element->nodal-values map once.
-    Centry = Matrix{Any}(undef, 6, 6)
-    for I in CartesianIndices(C)
-        cij = C[I]
-        if cij isa ScalarField
-            ce = nodesToElements(cij)
-            Centry[I] = Dict(zip(ce.numElem, ce.A))  # elem => (numNodes×1) nodal values
-        else
-            Centry[I] = Float64(cij)
+    Centry = nothing
+    if C !== nothing
+        Centry = Matrix{Any}(undef, 6, 6)
+        for I in CartesianIndices(C)
+            cij = C[I]
+            if cij isa ScalarField
+                ce = nodesToElements(cij)
+                Centry[I] = Dict(zip(ce.numElem, ce.A))  # elem => (numNodes×1) nodal values
+            else
+                Centry[I] = Float64(cij)
+            end
         end
     end
 
@@ -2389,18 +2401,67 @@ function materialTangentMatrix(
                             Fgp .+= Na * reshape(Fnode[9a-8:9a, 1], 3, 3)
                         end
 
-                        # --- C at Gauss: Cgp_ij = Σ_a N_a Cij_a  (ha ScalarField)
                         Cgp = zeros(6, 6)
-                        for ii in 1:6, jj in 1:6
-                            cij = Centry[ii, jj]
-                            if cij isa Float64
-                                Cgp[ii, jj] = cij
-                            else
-                                # dict elem => nodalValues(numNodes×1)
-                                nod = cij[elem][:, 1]
-                                Cgp[ii, jj] = dot(nod, H[:, k])
+
+                        if energy !== nothing
+                            # -----------------------------
+                            # NEW PATH: energy-based tangent
+                            # -----------------------------
+                            # Build right Cauchy–Green at Gauss: C = F^T F
+                            # (use SMatrix for speed + wrapper expects SMatrix{3,3})
+                            Fgpm = @SMatrix [
+                                Fgp[1,1] Fgp[1,2] Fgp[1,3];
+                                Fgp[2,1] Fgp[2,2] Fgp[2,3];
+                                Fgp[3,1] Fgp[3,2] Fgp[3,3]
+                            ]
+   
+                            Cmat = transpose(Fgpm) * Fgpm  # SMatrix{3,3}
+   
+                            # tangent_from_energy returns 6×6 (Voigt) in your extension convention
+                            Ctan = tangent_from_energy(energy, Cmat, params)  # SMatrix{6,6}
+   
+                            # Convert to Matrix{Float64} for B' * Cgp * B
+                            Cgp .= Matrix(Ctan)
+   
[…]
