### 2026-01-30 93dc87d
merge vector fields
