+++ src/nonlinear.jl
@@ -2166,12 +2166,20 @@ end
 #############################################################
 
 """
-    materialTangentMatrix(problem::Problem; F::TensorField, C::AbstractMatrix)
+    materialTangentMatrix(
+        problem::Problem;
+        F::TensorField,
+        C::AbstractMatrix,
+        energy::Function,
+        params
+    )
 
 Assembles the **material (constitutive) tangent stiffness matrix**
 for a 3D solid under **finite deformation (Total Lagrange formulation)**,
 using a user-provided deformation gradient field `F` and a material
-tangent matrix `C` given in **6×6 Mandel/Voigt form**.
+tangent matrix `C` given in **6×6 Mandel/Voigt form**. Alternatively,
+it can compute the material tangent from a **free-energy density**
+`energy` (hyperelastic formulation) via `Tensors.jl`.
 
 This function builds the matrix
 ```
@@ -2219,6 +2227,14 @@ supplied `C`, which may be spatially varying.
 C[i,j] isa Number || C[i,j] isa ScalarField
 
 ```
+
+- `energy::Function`  
+  Free-energy density `ψ(C, params)` of a hyperelastic material, evaluated
+  at the **right Cauchy–Green tensor** `C = FᵀF`. If provided, the tangent
+  is computed at Gauss points via `Tensors.jl` and `C` must be omitted.
+
+- `params`  
+  Optional parameter container passed through to `energy`.
 ---
 
 ## Mathematical formulation
@@ -2265,6 +2281,11 @@ separately.
 - finite strains,
 provided that `C` is consistent with the stress measure used elsewhere.
 
+- If `energy` is supplied, the routine constructs `C = FᵀF` at Gauss points
+  and obtains the constitutive tangent from the free-energy function using
+  `Tensors.jl`. This enables arbitrary hyperelastic laws without providing
+  a closed-form `C`.
+
 - The material tangent `C` is evaluated at Gauss points by interpolating
 nodal `ScalarField` values when necessary.
 
@@ -2309,58 +2330,49 @@ function materialTangentMatrix(
     C::Union{AbstractMatrix,Nothing}=nothing,
     energy::Union{Nothing,Function}=nothing,
     params=nothing)
-    
-    @assert xor(C !== nothing, energy !== nothing) """
-    Either provide `C` (6×6, Number or ScalarField entries),
-    or provide (`energy`, `params`) to compute the tangent at Gauss points.
-    """
 
-    @assert problem.dim == 3
-    @assert problem.pdim == 3
-    if C !== nothing
-        @assert size(C) == (6, 6)
-        @assert all(x -> (x isa Number) || (x isa ScalarField), C)
+    if problem.type == :dummy
+        return nothing
     end
+    @assert xor(C !== nothing, energy !== nothing)
+    @assert problem.dim == 3 && problem.pdim == 3
 
     gmsh.model.setCurrent(problem.name)
 
-    dim = 3
     pdim = 3
     dof = pdim * problem.non
 
-    # --- elementwise nodal F (9×numNodes per element)
+    # --- elementwise nodal F
     Fe = nodesToElements(F)
     Fmap = Dict(zip(Fe.numElem, Fe.A))
 
-    # --- Preprocess C entries:
-    # For ScalarField entries: build element->nodal-values map once.
+    # --- preprocess constant / field C (old path)
     Centry = nothing
     if C !== nothing
         Centry = Matrix{Any}(undef, 6, 6)
-        for I in CartesianIndices(C)
-            cij = C[I]
+        for Icart in CartesianIndices(C)
+            cij = C[Icart]
             if cij isa ScalarField
                 ce = nodesToElements(cij)
-                Centry[I] = Dict(zip(ce.numElem, ce.A))  # elem => (numNodes×1) nodal values
[…]
