+++ src/nonlinear.jl
@@ -26,275 +26,6 @@ function nodePositionVector(problem)
     return VectorField([], reshape(r, :,1), [0], [], 1, :v3D, problem)
 end
 
-#=
-function ∇_old(rr::Union{VectorField, ScalarField, TensorField}; nabla=:grad)
-    problem = rr.model
-    gmsh.model.setCurrent(problem.name)
-    if rr.a == [;;]
-        r = elementsToNodes(rr)
-    else
-        r = rr
-    end
-    DoFResults=false
-    
-    nsteps = r.nsteps
-    ε = []
-    numElem = Int[]
-    ncoord2 = zeros(3 * problem.non)
-    dim = problem.dim
-    pdim = problem.pdim
-    non = problem.non
-    type = :e
-    if DoFResults == true
-        E1 = zeros(non * 9, nsteps)
-        pcs = zeros(Int64, non * dim)
-    end
-
-    for ipg in 0:length(problem.material)
-        phName = ""
-        if ipg == 0
-            if rr.model.geometry.nameVolume ≠ ""
-                phName = rr.model.geometry.nameVolume
-            else
-                continue
-            end
-        else
-            phName = problem.material[ipg].phName
-        end
-        #ν = problem.material[ipg].ν
-        dim = 0
-        if problem.dim == 3 && r isa VectorField && nabla == :grad
-            dim = 3
-            pdim = 3
-            rowsOfB = 9
-            b = 1
-            sz = 9
-        elseif problem.dim == 3 && r isa VectorField && nabla == :div
-            dim = 3
-            pdim = 3
-            rowsOfB = 9
-            b = 1
-            sz = 1
-        elseif problem.dim == 3 && r isa VectorField && nabla == :curl
-            dim = 3
-            pdim = 3
-            rowsOfB = 9
-            b = 1
-            sz = 3
-        elseif (problem.dim == 1 || problem.dim == 2 || problem.dim == 3) && r isa ScalarField
-            dim = 3
-            pdim = 1
-            rowsOfB = 3
-            b = 1
-            sz = 3
-        elseif r isa TensorField && nabla == :div
-            dim = 3
-            pdim = 9
-            rowsOfB = 3
-            b = 1
-            sz = 3
-        elseif problem.dim == 2 && problem.type == :AxiSymmetric
-            dim = 2
-            rowsOfB = 4
-            b = 1
-        else
-            error("∇: dimension is $(problem.dim), problem type is $(problem.type).")
-        end
-
-        dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
-        for idm in 1:length(dimTags)
-            dimTag = dimTags[idm]
-            edim = dimTag[1]
-            etag = dimTag[2]
-            elemTypes, elemTags, elemNodeTags = gmsh.model.mesh.getElements(edim, etag)
-            #nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes(dim, -1, true, false)
-            nodeTags, ncoord, parametricCoord = gmsh.model.mesh.getNodes()
-            ncoord2[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
-            ncoord2[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
-            ncoord2[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
-            for i in 1:length(elemTypes)
-                et = elemTypes[i]
-                elementName, dim, order, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(et)
-                #e0 = zeros(rowsOfB * numNodes)
-                nodeCoord = zeros(numNodes * 3)
-                for k in 1:dim, j = 1:numNodes
[…]
