### 2026-02-07 #9150 @prashjha
Hi @TimJosephson, @turnerjennings, @lamBOOO, @alizma, and @yangbai90, I am looking for a reviewer for the above JOSS submission, and I am reaching out as I think this may be in your domain. Would you be interested in reviewing this?

You can find pdf at this link [pdf link](https://github.com/openjournals/joss-papers/blob/joss.09150/joss.09150/10.21105.joss.09150.pdf), and the corresponding software is at this [link](https://github.com/perebalazs/LowLevelFEM.jl).

This [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
