+++ doc/LowLevelFEM – API logical structure (draft, function-level).md
@@ -1,0 +1,307 @@
+# LowLevelFEM – API logical structure (draft, function-level)
+
+⚠️ Working document.
+This is NOT documentation, but a planning map for later API restructuring.
+
+---
+
+## Core
+
+Fundamental data structures and concepts.
+
+### Problem and materials
+
+- Problem
+- Material
+- material
+- elastic_constants
+- elasticSupport
+- elasticSupportMatrix
+- printElasticConstantsTable
+
+### Fields
+
+- ScalarField
+- VectorField
+- TensorField
+- scalarField
+- vectorField
+- tensorField
+- field
+- mergeFields
+- projectScalarField
+- fieldError
+- loadField
+- saveField
+- isNodal
+- isElementwise
+- isSaved
+
+### Mesh and geometry
+
+- Geometry
+- CoordinateSystem
+- generateMesh
+- rotateNodes
+- projectTo2D
+- expandTo3D
+- getDimForPhysicalName
+- getTagForPhysicalName
+
+### Mapping and integration
+
+- nodesToElements
+- elementsToNodes
+- integrate
+- normalVector
+- normalVector1D
+- tangentVector
+
+---
+
+## Preprocessing
+
+Preparation of FEM problems.
+
+### Boundary conditions
+
+- BoundaryCondition
+- BoundaryConditionFields
+- displacementConstraint
+- temperatureConstraint
+- pressureConstraint
+- elasticSupport
+- applyBoundaryConditions
+- applyBoundaryConditions!
+- applyElasticSupport!
+- suppressDeformationAtBoundaries
+- suppressDeformationAtBoundaries!
+
+### Initial conditions
+
+- initialDisplacement
+- initialDisplacement!
+- initialVelocity
+- initialVelocity!
+- initialTemperature
+- initialTemperature!
+
+### Loads and sources
+
+- loadVector
+- loadVectorOld
+- loadVectorNew
+- nonFollowerLoadVector
+- equivalentNodalForce
+- thermalLoadVector
+- thermalLoadVectorAXI
+- thermalLoadVectorSolid
+- heatSource
+- heatSourceVector
+- heatFlux
+- heatFluxVector
+- heatConvection
+- heatConvectionVector
+
+---
+
+## Operators
+
+Weak-form building blocks, physics-independent.
+
+### Scalar field operators (Poisson / Laplace)
+
+- poissonMatrix
+- poissonMatrixVector
+- poissonMatrixSymGradF
+- stiffnessMatrixPoissonAllInOne
+- traceLaplaceMatrix
+- tensorLaplaceMatrix
+- reactionMatrix
+- sourceVector
+- solveField
+
+### Vector field operators
+
+- gradMatrix
+- gradDivMatrix
+- gradDivMatrixF
+- curlCurlMatrix
+- symmetricGradientMatrix
+- tensorDivDivMatrix
+- beltramiMichellMatrix
+
+### Advection and flow-related operators
+
+- advectionMatrix
+- navierStokesAdvectionMatrix
+
+### Mass and auxiliary operators
+
+- massMatrix
+- mapScalarField
+- unitTensor
+- estimateLengthOfIJV
+
+---
+
+## Mechanics
+
+### Linear mechanics
+
+Small-deformation solid mechanics.
+
+#### Stiffness and mass
+
+- stiffnessMatrix
+- stiffnessMatrixSolid
+- stiffnessMatrixSolidParallel
+- stiffnessMatrixSolidParallelNew
+- stiffnessMatrixAXI
+- stiffnessMatrixTruss
+- massMatrixSolid
+- massMatrixSolidParallel
+- massMatrixTruss
+- dampingMatrix
+
+#### Constitutive relations
+
[…]
