### 2026-02-20 #10096 @editorialbot
[REVIEW]: LowLevelFEM.jl: A lightweight finite element toolbox in Julia
**Submitting author:** @perebalazs (<a href="http://orcid.org/0000-0002-1161-4206">Balázs Pere</a>)
**Repository:** https://github.com/perebalazs/LowLevelFEM.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.9.1
