### 2026-02-20 #10096 @prashjha
Hi @perebalazs, please fix the JOSS draft to address missing sections; see [above](https://github.com/openjournals/joss-reviews/issues/10096#issuecomment-3937545414). Read the [JOSS paper format guidelines](https://joss.readthedocs.io/en/latest/paper.html).

This needs to be fixed soon so that reviewers can review updated article and software.
