+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 LowLevelFEM.jl is a finite element method (FEM) [@zienkiewicz2005] toolbox written entirely in the Julia programming language [@bezanson2017]. Its design philosophy emphasizes **simplicity, transparency, and performance**, making it suitable for both educational and research purposes in mechanics and engineering. Unlike large frameworks that rely on domain-specific languages or compiled backends, LowLevelFEM provides users with direct access to FEM building blocks in Julia, enabling full control over discretization, assembly, and solution steps.
 
-The package currently supports two- and three-dimensional solid mechanics problems including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimalistic design lowers the entry barrier for students while still offering enough flexibility for advanced users to **inspect and control algorithms step by step, process intermediate results during the solution procedure, and perform operations with scalar, vector, and tensor fields**. LowLevelFEM uses **Gmsh** [@geuzaine2009] as its pre- and post-processor, ensuring compatibility with a widely adopted meshing and visualization tool.
+The package currently supports two- and three-dimensional solid mechanics problems including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimalistic design lowers the entry barrier for students while still offering enough flexibility for advanced users to **inspect and control algorithms step by step, process intermediate results during the solution procedure, and perform operations with scalar, vector, and tensor fields**. LowLevelFEM uses **Gmsh** [@geuzaine2009] as its pre- and post-processor, ensuring compatibility with a widely adopted meshing and visualization tool. The toolbox also supports geometrically nonlinear formulations based on a Total Lagrangian framework and energy-driven constitutive modeling, enabling research-level investigations beyond linear elasticity.
 
 Thanks to Julia’s just-in-time compilation and multiple dispatch, the code remains concise while achieving performance comparable to traditional FEM codes in C/C++ or Fortran. LowLevelFEM is released under the MIT license and distributed via the Julia General registry. Documentation, tutorials, and examples are available at [https://juliahub.com/ui/Packages/General/LowLevelFEM](https://juliahub.com/ui/Packages/General/LowLevelFEM) or [https://perebalazs.github.io/LowLevelFEM.jl/stable/](https://perebalazs.github.io/LowLevelFEM.jl/stable/).
 
@@ -42,11 +42,13 @@ prob = Problem([mat], type=:PlaneStress)  # :Solid, :PlaneStrain, :AxiSymmetric,
 bc   = displacementConstraint("supp", ux=0, uy=0)
 force = load("load", fy=-1)
 
-u = solveDisplacement(prob, [force], [bc])
+u = solveDisplacement(prob, load=[force], support=[bc])
 S = solveStress(u)
 
 showDoFResults(u)
+showDoFResults(u, :ux)
 showStressResults(S)
+showStressResults(S, :sxy, name="Shear stress")
 
 openPostProcessor()
 gmsh.finalize()
@@ -59,13 +61,14 @@ Alternatively, a lower-level sequence:
 ```julia
 K = stiffnessMatrix(prob)
 f = loadVector(prob, [force])
-applyBoundaryConditions!(K, f, [bc])
-u = K \ f
+u = solveDisplacement(K, f, support=[bc])
 
 # Simple Hooke's law stress computation
+E = mat.E
+ν = mat.ν
 A = (u ∘ ∇ + ∇ ∘ u) / 2
-I = unitTensor(A)
-S = E / (1 + ν) * (A + ν / (1 - 2 ν) * trace(A) * I)
+I = TensorField(prob, "body", [1 0 0; 0 1 0; 0 0 1])
[…]
