+++ doc/MultiFields - 🎯 FŐ VONAL.md
@@ -1,0 +1,371 @@
+Nagyon jó ötlet most megállni és „lefagyasztani” az architektúrát 💡  
+Az alábbi összefoglaló kifejezetten arra készült, hogy pár nap múlva gyorsan vissza tudj kapcsolódni.
+
+Nem implementációs részletek, hanem **struktúra + tervezési döntések + indoklás**.
+
+---
+
+# 🎯 FŐ VONAL
+
+```
+assemble_operator
+        ↓
+SystemMatrix  (lehet nxm)
+        ↓
+blokk-összeállítás  [A  B'; B  C]
+        ↓
+solveMultiFields(K, F; supports=...)
+```
+
+Cél:
+
+- egymezős eset változatlan marad
+
+- többmezős természetes kiterjesztés
+
+- nincs fix u,p
+
+- tetszőleges számú mező
+
+- nincs redundáns információ
+
+---
+
+# 1️⃣ assemble_operator
+
+### Filozófia
+
+Az operátor már most is két mező között épül:
+
+```julia
+assemble_operator(Prow, Pcol, Op_row, Op_col; coefficient=...)
+```
+
+Ez természetesen nxm mátrixot ad.
+
+---
+
+### Mit kell mostantól tudnia a visszatérő SystemMatrix-nak?
+
+A visszaadott `SystemMatrix` tartalmazza:
+
+- `A::SparseMatrixCSC`
+
+- `row_problem::Problem`
+
+- `col_problem::Problem`
+
+Egymezős esetben:
+
+```
+row_problem = col_problem = P
+```
+
+Kereszt operátornál:
+
+```
+row_problem = Pp
+col_problem = Pu
+```
+
+Ez kulcsfontosságú, mert a blokk-összeállítás innen tudja,  
+melyik mező melyikhez kapcsolódik.
+
+---
+
+# 2️⃣ SystemMatrix kibővített koncepció
+
+A jelenlegi:
+
+```julia
+SystemMatrix(A, model)
+```
+
+Jövőbeli (logikai szerkezet):
+
+```julia
+SystemMatrix
+    A
+    row_problem
+    col_problem
+    problems   # csak blokk esetben
+    offsets    # csak blokk esetben
+```
+
+### Egymezős kompatibilitás
+
+Ha row_problem == col_problem:
+
+```
+problems = [model]
+offsets = [0]
+```
+
+Semmi nem törik.
+
+---
+
+# 3️⃣ Blokk-összeállítás
+
+Cél:
+
+```julia
+K = SystemMatrix([A  B';
+                  B  C])
+```
+
+ahol:
+
+- A: (Pu, Pu)
+
+- B: (Pp, Pu)
+
+- C: (Pp, Pp)
+
+---
+
+### Mit csinál a blokk-konstruktor?
+
+1. Végigmegy minden blokkon
+
+2. Összegyűjti az összes unique Problem-et
+
+3. Meghatározza a sorrendet (first appearance rule)
+
+4. Offseteket számol:
+
+```
+offset[i] = sum(ndofs(problems[1:i-1]))
+```
+
+5. Összerakja a nagy sparse mátrixot
+
+6. Tárolja:
+
+```
+K.problems
+K.offsets
+```
+
+---
+
+### Eredmény
+
+A solver már nem találgat.  
+A blokkstruktúra az assembly-ből jön.  
+Nincs redundáns Pu/Pp megadás.
+
+---
+
+# 4️⃣ solveMultiFields
+
+### Cél
+
+A mostani `solveField` logikájának többmezős kiterjesztése.
+
+Most:
+
+```
+K_ff u_f = f_f − K_fc u_c
+```
+
+Ez marad.
+
+---
+
+### Globális működés
+
[…]
