### 2026-02-23 adece70
bug fixes
new getNumber function to import values of variables from Gmsh.
