### 2026-02-25 4fcd957
structures box and rectangle meshes for testing
