+++ src/general.jl
@@ -15,6 +15,7 @@ export openPreProcessor, openPostProcessor, setParameter, setParameters, getPara
 export probe
 export saveField, loadField, isSaved
 export ∂x, ∂y, ∂z, ∂t
+export structured_rect_mesh, structured_box_mesh
 
 """
     Material(phName, type, E, ν, ρ, k, c, α, λ, μ, κ)
@@ -4997,8 +4998,10 @@ function showDoFResults(q::Union{ScalarField,VectorField,TensorField}, comp::Sym
     elseif ff == 1 || ff == 2
         gmsh.view.option.setNumber(uvec, "ShowTime", 6)
         gmsh.view.option.setNumber(uvec, "VectorType", 5)
+    elseif factor != 0
+        gmsh.view.option.setNumber(uvec, "VectorType", 5)
+        gmsh.view.option.setNumber(uvec, "DisplacementFactor", factor)
     end
-    gmsh.view.option.setNumber(uvec, "DisplacementFactor", factor)
     return uvec
 end
 
@@ -5073,8 +5076,9 @@ Types:
 - `smooth`: Boolean
 - `tag`: Integer
 """
-function showStrainResults(E, comp; name=comp, visible=false, smooth=true)
+function showStrainResults(E0, comp; name=comp, visible=false, smooth=true)
     #gmsh.fltk.openTreeItem("0Modules/Post-processing")
+    E = nodesToElements(E0)
     problem = E.model
     gmsh.model.setCurrent(problem.name)
     gmsh.option.setNumber("Mesh.VolumeEdges", 0)
@@ -5239,8 +5243,9 @@ Types:
 - `smooth`: Boolean
 - `tag`: Integer
 """
-function showStressResults(S::TensorField, comp; name=comp, visible=false, smooth=false)
+function showStressResults(S0::TensorField, comp; name=comp, visible=false, smooth=false)
     #gmsh.fltk.openTreeItem("0Modules/Post-processing")
+    S = nodesToElements(S0)
     problem = S.model
     gmsh.model.setCurrent(problem.name)
     gmsh.option.setNumber("Mesh.VolumeEdges", 0)
@@ -5348,7 +5353,8 @@ Types:
 - `smooth`: Boolean
 - `tag`: Integer
 """
-function showHeatFluxResults(S::VectorField, comp; name=comp, visible=false, smooth=true, factor=0)
+function showHeatFluxResults(S0::VectorField, comp; name=comp, visible=false, smooth=true, factor=0)
+    S = nodesToElements(S0)
     problem = S.model
     #gmsh.fltk.openTreeItem("0Modules/Post-processing")
     gmsh.model.setCurrent(problem.name)
@@ -5447,7 +5453,8 @@ function showHeatFluxResults(S::VectorField, comp; name=comp, visible=false, smo
     return SS
 end
 
-function showScalarResults(S; name="ScalarField", visible=false, smooth=false, factor=0)
+function showScalarResults(S0; name="ScalarField", visible=false, smooth=false, factor=0)
+    S = nodesToElements(S0)
     problem = S.model
     #gmsh.fltk.openTreeItem("0Modules/Post-processing")
     gmsh.model.setCurrent(problem.name)
@@ -5957,4 +5964,230 @@ Types:
 """
 function isSaved(fileName::String)
     return isfile(fileName * "-LLF-Data.jld2")
+end
+
+"""
+    structured_rect_mesh(; x0=0.0, y0=0.0,
+                           lx=1.0, ly=1.0,
+                           n=10,
+                           dx=lx/n, dy=ly/n,
+                           order=1)
+
[…]
