+++ src/linear.jl
@@ -3082,14 +3082,15 @@ Types:
 - `F`: Union{Nothing,TensorField}
 """
 function loadVector(problem, loads; F=nothing)
+
     if problem.type == :dummy
         return nothing
     end
-    
+
     gmsh.model.setCurrent(problem.name)
-    if !isa(loads, Vector)
-        error("loadVector: loads are not arranged in a vector. Put them in [...]")
-    end
+
+    !isa(loads, Vector) &&
+        error("loadVector: loads must be in a vector.")
 
     Fe = nothing
     Fmap = nothing
@@ -3103,49 +3104,78 @@ function loadVector(problem, loads; F=nothing)
     b = problem.thickness
     non = problem.non
     dof = pdim * non
-    ncoord2 = zeros(3 * problem.non)
-    f = nothing
-    fp = zeros(dof,1)
+
+    ncoord2 = zeros(3 * non)
+    fp = zeros(dof, 1)
     nsteps = 1
-    for n in 1:length(loads)
-        name = loads[n].phName
-        fx = loads[n].fx
-        fy = loads[n].fy
-        fz = loads[n].fz
-        T = loads[n].T
-        p = loads[n].p
-        hc = loads[n].h
-        T0 = loads[n].T
-        qn = loads[n].qn
-        hs = loads[n].h
 
-        q = loads[n].q
+    for load in loads
 
-        fxy = loads[n].fxy
-        fyz = loads[n].fyz
-        fzx = loads[n].fzx
-        
-        fyx = loads[n].fyx
-        fzy = loads[n].fzy
-        fxz = loads[n].fxz
+        # multi-field filter
+        if load.problem !== nothing && load.problem !== problem
+            continue
+        end
 
-        fx  = (fx  isa ScalarField && isElementwise(fx))  ? elementsToNodes(fx)  : fx
-        fy  = (fy  isa ScalarField && isElementwise(fy))  ? elementsToNodes(fy)  : fy
-        fz  = (fz  isa ScalarField && isElementwise(fz))  ? elementsToNodes(fz)  : fz
-        T   = (T   isa ScalarField && isElementwise(T))   ? elementsToNodes(T)   : T
-        p   = (p   isa ScalarField && isElementwise(p))   ? elementsToNodes(p)   : p
-        hc  = (hc  isa ScalarField && isElementwise(hc))  ? elementsToNodes(hc)  : hc
-        T0  = (T0  isa ScalarField && isElementwise(T0))  ? elementsToNodes(T0)  : T0
-        qn  = (qn  isa ScalarField && isElementwise(qn))  ? elementsToNodes(qn)  : qn
-        hs  = (hs  isa ScalarField && isElementwise(hs))  ? elementsToNodes(hs)  : hs
-        q   = (q   isa ScalarField && isElementwise(q))   ? elementsToNodes(q)   : q
-        fxy = (fxy isa ScalarField && isElementwise(fxy)) ? elementsToNodes(fxy) : fxy
-        fyz = (fyz isa ScalarField && isElementwise(fyz)) ? elementsToNodes(fyz) : fyz
-        fzx = (fzx isa ScalarField && isElementwise(fzx)) ? elementsToNodes(fzx) : fzx
-        fyx = (fyx isa ScalarField && isElementwise(fyx)) ? elementsToNodes(fyx) : fyx
-        fzy = (fzy isa ScalarField && isElementwise(fzy)) ? elementsToNodes(fzy) : fzy
-        fxz = (fxz isa ScalarField && isElementwise(fxz)) ? elementsToNodes(fxz) : fxz
-    
+        name = load.phName
+        vals = load.values
+
+        # validate logic
+        _check_load_keys(problem, vals)
+
[…]
