+++ src/general.jl
@@ -1,5 +1,6 @@
 export Problem, Material, getEigenVectors, getEigenValues, material
 export displacementConstraint, load, elasticSupport, BoundaryCondition, BoundaryConditionFields
+export LoadCondition
 export temperatureConstraint, heatFlux, heatSource, heatConvection
 export field, scalarField, vectorField, tensorField, ScalarField, VectorField, TensorField
 export mergeFields
@@ -249,6 +250,136 @@ struct Problem
     thickness::Float64
     non::Int64
     geometry::Geometry
+    field::Symbol
+    rhs_field::Symbol
+    Problem() = new()
+    Problem(name, type, dim, pdim, material, thickness, non, geometry, field, rhs_field) =
+        new(name, type, dim, pdim, material, thickness, non, geometry, field, rhs_field)
+    Problem(name, type, dim, pdim, material, thickness, non, geometry) =
+        new(name, type, dim, pdim, material, thickness, non, geometry, :unknown, :rhs)
+    function Problem(mat; thickness=1.0, type=:Solid, bandwidth=:none,
+        nameTopSurface="", nameVolume="", dim::Int=3,
+        field::Symbol=:unknown, rhs_field::Symbol=:rhs)
+        if type == :dummy
+            return new("dummy", :dummy, 0, 0, mat, 0, 0, Geometry(), field, rhs_field)
+        end
+        pdim = 3
+        dim0 = dim
+
+        #if Sys.CPU_THREADS != Threads.nthreads()
+        #    @warn "Number of threads($(Threads.nthreads())) ≠ logical threads in CPU($(Sys.CPU_THREADS))."
+        #end
+        geometry = Geometry()
+
+        if type == :Solid
+            dim = 3
+            pdim = 3
+        elseif type == :PlaneStress
+            dim = 2
+            pdim = 2
+        elseif type == :PlaneStrain
+            dim = 2
+            pdim = 2
+        elseif type == :AxiSymmetric
+            dim = 2
+            pdim = 2
+        elseif type == :PlaneHeatConduction
+            dim = 2
+            pdim = 1
+        elseif type == :HeatConduction
+            dim = 3
+            pdim = 1
+        elseif type == :AxiSymmetricHeatConduction
+            dim = 2
+            pdim = 1
+        elseif type == :Truss
+            dim = 3
+            pdim = 3
+        elseif type == :ScalarField
+            dim0 < 1 && error("Problem: dimension of a :ScalarField problem must be one, two or three.")
+            dim0 > 3 && error("Problem: dimension of a :ScalarField problem must be equal or less than three.")
+            dim = dim0
+            pdim = 1
+        elseif type == :VectorField
+            dim0 < 1 && error("Problem: dimension of a :VectorField problem must be one, two or three.")
+            dim0 == 1 && error("Problem: dimension of a :VectorField problem must be greater than one.")
+            dim0 > 3 && error("Problem: dimension of a :VectorField problem must be two or three.")
+            dim = dim0
+            pdim = dim0
+        elseif type == :TensorField
+            dim0 < 1 && error("Problem: dimension of a :TensorField problem must be one, two or three.")
+            dim0 == 1 && error("Problem: dimension of a :TensorField problem must be greater than one.")
+            dim0 > 3 && error("Problem: dimension of a :TensorField problem must be two or three.")
+            dim = dim0
+            pdim = 9
+        elseif type == :Reynolds
+            geometry = Geometry(nameTopSurface, nameVolume)
+            dim = geometry.dim
+            pdim = 1
+        else
[…]
