### 2026-03-03 c9f8d56
Refactored BC and load system to field-based design. Solid and heat tests passing.
