+++ src/general.jl
@@ -4,7 +4,7 @@ export LoadCondition
 export temperatureConstraint, heatFlux, heatSource, heatConvection
 export field, scalarField, vectorField, tensorField, ScalarField, VectorField, TensorField
 export mergeFields
-export SystemMatrix
+export SystemMatrix, SystemVector
 export constrainedDoFs, freeDoFs, allDoFs, DoFs
 export elementsToNodes, nodesToElements, projectTo2D, expandTo3D, isNodal, isElementwise
 export fieldError, resultant, integrate, ∫, normalVector, tangentVector
@@ -512,6 +512,8 @@ struct Transformation
     dim::Int64
 end
 
+ndofs(P::Problem) = P.non * P.pdim
+
 """
     SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem)
     SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem, test_model::Problem)
@@ -526,10 +528,142 @@ Types:
 """
 struct SystemMatrix
     A::SparseMatrixCSC
-    model::Problem
-    test_model::Problem
+    model::Union{Problem,Nothing}
+    test_model::Union{Problem,Nothing}
+    problems::Union{Vector{Problem},Nothing}      # többmezős meta
+    offsets::Union{Vector{Int},Nothing}           # globális kezdő indexek
     SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem, test_model::Problem) = new(A, model, test_model)
     SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem) = new(A, model, model)
+    function SystemMatrix(blocks::Matrix{SystemMatrix})
+
+        nrows, ncols = size(blocks)
+   
+        # ----------------------------------------------------------
+        # 1) Collect trial problems (column-based order)
+        # ----------------------------------------------------------
+        trial_problems = Problem[]
+   
+        function push_unique!(vec, P)
+            P === nothing && return
+            if all(q -> q !== P, vec)
+                push!(vec, P)
+            end
+        end
+   
+        for j in 1:ncols
+            for i in 1:nrows
+                blk = blocks[i, j]
+                if blk.model !== nothing
+                    push_unique!(trial_problems, blk.model)
+                    break
+                end
+            end
+        end
+   
+        isempty(trial_problems) &&
+            error("No trial Problems found in block matrix.")
+   
+        # ----------------------------------------------------------
+        # 2) Collect test problems (row-based order)
+        # ----------------------------------------------------------
+        test_problems = Problem[]
+   
+        for i in 1:nrows
+            for j in 1:ncols
+                blk = blocks[i, j]
+                if blk.test_model !== nothing
+                    push_unique!(test_problems, blk.test_model)
+                    break
+                end
+            end
+        end
+   
+        isempty(test_problems) &&
+            error("No test Problems found in block matrix.")
+   
+        # ----------------------------------------------------------
+        # 3) Field-level square check
+        # ----------------------------------------------------------
+        if length(trial_problems) != length(test_problems) ||
+           any(trial_problems[i] !== test_problems[i]
+               for i in eachindex(trial_problems))
+            error("Block system is not square in field sense. Trial and test spaces differ.")
+        end
+   
+        problems = trial_problems
+   
+        # ----------------------------------------------------------
+        # 4) Compute global offsets (based on trial ordering)
+        # ----------------------------------------------------------
+        offsets = Vector{Int}(undef, length(problems))
[…]
