+++ src/multifield.jl
@@ -1,0 +1,1505 @@
+
+"""
+LowLevelFEM Multifield Operator Kernel
+
+This module implements the low-level operator infrastructure used to
+assemble finite element matrices from weak form expressions.
+
+The file contains three main layers:
+
+1. **Operator kernel**
+
+   Defines differential operators (`GradOp`, `DivOp`, `SymGradOp`, …)
+   and constructs element operator matrices through `build_B!`.
+
+2. **Matrix assembly**
+
+   The function
+
+       assemble_operator
+
+   performs element integration and builds the global sparse matrix.
+
+3. **Weak-form DSL**
+
+   A lightweight domain-specific language allows the construction of
+   weak forms using mathematical notation:
+
+       ∫( Grad(Pu) ⋅ Grad(Pu); Ω="domain")
+
+       ∫( SymGrad(Pu) ⋅ C ⋅ SymGrad(Pu); Ω="solid")
+
+The implementation is designed to be:
+
+- transparent
+- extensible
+- suitable for multifield problems
+- compatible with Gmsh meshes.
+
+See also:
+`SystemMatrix`, `SystemVector`, `solveField`.
+"""
+
+export Grad
+export Div
+export Curl
+export SymGrad
+export Id
+export TensorDiv
+export Adv
+
+export ∫
+export ∫Ω
+export ∫Γ
+
+export solveField
+
+export ε
+
+export ⋅
+
+
+"""
+Abstract base type for finite element operators.
+
+Operators act on fields defined by a `Problem` and are used to build
+weak-form expressions.
+
+Examples include
+
+    IdOp       identity operator
+    GradOp     gradient
+    DivOp      divergence
+    CurlOp     curl
+    SymGradOp  symmetric gradient
+
+Operators are applied to a `Problem` via helper constructors:
+
+    Grad(P)
+    Div(P)
+    SymGrad(P)
+
+and used inside weak forms such as
+
+    ∫( Grad(Pu) ⋅ Grad(Pu) )
+"""
+abstract type AbstractOp end
+
+"""
+    IdOp()
+
+Identity operator.
+
+Represents the operator
+
+    u
+
+in a weak form expression.
+
+Example
+
+    ∫( Id(Pu) ⋅ Id(Pu) )
+
+which corresponds to
+
+    ∫ u · v dΩ
+"""
+struct IdOp <: AbstractOp end     # u
+
+"""
+    GradOp()
+
+Gradient operator.
+
+For scalar fields
+
+    grad(u)
+
+For vector fields
+
+    ∇u
+
+The output dimension depends on the field type:
+
+Scalar field
+
+    grad(u) → dim
+
+Vector field
+
+    ∇u → dim × pdim
+"""
+struct GradOp <: AbstractOp end     # ∇u
+
+"""
+    DivOp()
+
+Divergence operator acting on a vector field.
+
+Represents
+
+    div(u)
+
+The operator assumes that
+
+    pdim == dim
+
+Example
+
+    ∫( Div(Pu) ⋅ Div(Pu) )
+"""
+struct DivOp <: AbstractOp end     # div u
+
+"""
+    CurlOp()
+
+Curl operator acting on a vector field.
+
+2D
+
+    curl(u) = ∂u_y/∂x − ∂u_x/∂y
+
+3D
+
+    curl(u) = ∇ × u
+"""
+struct CurlOp <: AbstractOp end    # rot u
+
+"""
+    SymGradOp()
+
+Symmetric gradient operator used in linear elasticity.
+
+Represents
+
+    ε(u) = (∇u + ∇uᵀ)/2
+
+Engineering strain components are returned.
+
+2D
+
+    [ εxx
[…]
