+++ src/general.jl
@@ -515,16 +515,86 @@ end
 ndofs(P::Problem) = P.non * P.pdim
 
 """
-    SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem)
-    SystemMatrix(A::SparseMatrixCSC{Float64}, model::Problem, test_model::Problem)
+    SystemMatrix_(blocks::Matrix{SystemMatrix})
 
-Structure containing the stiffness/mass/heat conduction/heat capacity/latent heat/... matrix and the 
-associated `Problem` (with its trial field) and another `Problem` (with a test field).
+Construct a global block system matrix from a matrix of `SystemMatrix` blocks.
 
-Types:
-- `A`: SparseMatrixCSC{Float64}
-- `model`: Problem
-- `test_model`: Problem
+This function assembles a global sparse matrix from a block structure where each
+block represents the discretization of a bilinear form between two fields
+(test and trial problems). The blocks are assumed to correspond to different
+fields in a multifield formulation.
+
+The function performs the following steps:
+
+1. Determine the ordering of **trial problems** from the columns of the block matrix.
+
+2. Determine the ordering of **test problems** from the rows of the block matrix.
+
+3. Verify that the block system is **square in the field sense**, i.e.
+   
+       trial_problems[i] === test_problems[i]
+   
+   for all fields.
+
+4. Compute global DOF offsets for each field.
+
+5. Assemble the global sparse matrix by inserting the entries of each block
+   at the appropriate offset position.
+
+Each block `blocks[i,j]` must satisfy
+
+    blocks[i,j].test_model  → test Problem (row space)
+    blocks[i,j].model       → trial Problem (column space)
+    blocks[i,j].A           → sparse block matrix
+
+The resulting global matrix has the structure
+
+    [ A₁₁  A₁₂  ⋯ ]
+    [ A₂₁  A₂₂  ⋯ ]
+    [ ⋮    ⋮    ⋱ ]
+
+where `Aᵢⱼ` couples the `j`-th trial field with the `i`-th test field.
+
+# Arguments
+
+- `blocks::Matrix{SystemMatrix}`: Matrix of block matrices representing the
+  bilinear operators between fields.
+
+# Returns
+
+- `SystemMatrix`: A global system matrix containing the assembled sparse matrix
+  and metadata about the involved `Problem`s and their DOF offsets.
+
+# Errors
+
+An error is thrown if:
+
+- no trial or test problems are found in the block matrix,
+- the block system is not square in the field sense,
+- a block contains inconsistent matrix dimensions.
+
+# Example
+
+```julia
+Auu = ∫(ε(Pu) ⋅ C ⋅ ε(Pu))
+Aup = ∫(Div(Pu) ⋅ Id(Pp))
+Apu = ∫(Div(Pu) ⋅ Id(Pp))'
+App = ∫(Grad(Pp) ⋅ Grad(Pp) * δ)
+
+K = SystemMatrix_([
+    Auu  Aup
+    Apu  App
+])
+```
+
+This assembles the global block system
+
+```
+[ Kuu  Kup ]
+[ Kpu  Kpp ]
+```
+
+corresponding to a two-field problem.  
 """
 struct SystemMatrix
     A::SparseMatrixCSC
@@ -1545,7 +1615,11 @@ end
 """
     SystemVector
 
-Unified global RHS vector for single-field and multi-field systems.
+Unified global right-hand-side (RHS) vector for both single-field and
+multi-field finite element systems.
+
+This type stores the global load vector together with metadata describing
+how the vector is associated with one or multiple `Problem`s.
 
 # Fields
 - `a::Vector{Float64}`
[…]
