+++ docs/src/manual/boundary-and-loads.md
@@ -1,0 +1,13 @@
+# Boundary Conditions and Loads
+
+This section documents how to define support constraints, mechanical/thermal loads,
+and conversion between boundary-condition and load-condition representations.
+
+## Status
+
+Placeholder page.
+
+## Related API
+
+- [Reference: Preprocessing](../reference/preprocessing.md)
+- [Reference: Heat](../reference/heat.md)
