+++ docs/src/reference/fields.md
@@ -1,0 +1,32 @@
+# Fields API
+
+Field containers and common field-level utilities.
+
+## Field Construction and Storage
+
+```@docs
+field
+scalarField
+vectorField
+tensorField
+mergeFields
+saveField
+loadField
+isSaved
+```
+
+## Field Access and Conversion
+
+```@docs
+DoFs
+elementsToNodes
+nodesToElements
+isNodal
+isElementwise
+```
+
+## Field Differential Helpers
+
+```@docs
+∂x
+```
