+++ docs/src/getting-started/index.md
@@ -1,0 +1,12 @@
+# Getting Started
+
+This section covers the minimum path from installation to a first finite element model in LowLevelFEM.
+
+## Roadmap
+
+1. [Installation](installation.md)
+2. [First Problem](first-problem.md)
+3. [Mesh and Physical Groups](mesh-and-physical-groups.md)
+4. [Fields and Results](fields-and-results.md)
+
+If you are already familiar with the package, jump to [Tutorials](../tutorials/index.md) or [API Reference](../reference/index.md).

[…]
