+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.13",
   "v1.12",
   "v1.11",
   "v1.10",
@@ -15,5 +16,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.12.3";
+var DOCUMENTER_NEWEST = "v1.13.0";
 var DOCUMENTER_STABLE = "stable";
