### 2026-03-09 98c6c2d
bug in assemble_operator - fixed
