+++ src/multifield.jl
@@ -1,4 +1,3 @@
-
 """
 LowLevelFEM Multifield Operator Kernel
 
@@ -659,7 +658,7 @@ Sparse matrix representing the discretized bilinear form.
 function assemble_operator(
     Pu::Problem, op_u::AbstractOp,
     Ps::Problem, op_s::AbstractOp;
-    coefficient::Union{Number,ScalarField}=1.0,
+    coefficient::Union{Number,ScalarField,AbstractMatrix}=1.0,
     weight=nothing,
     domain=nothing)
 
@@ -695,7 +694,7 @@ function assemble_operator(
         # build dictionaries for all ScalarFields inside matrix
         for c in coefficient
             if c isa ScalarField
-                pa = _build_elemwise_coeff_dict(c)
+                merge!(pa, _build_elemwise_coeff_dict(c))
             end
         end
     end
@@ -809,10 +808,8 @@ function assemble_operator(
 
                     # integrate
                     @inbounds for k in 1:numIntPoints
-                        #cval = (coefficient isa ScalarField) ? _coeff_at_gp(pa, elem, view(h, :, k)) : pconst
-                        #w = jacDet[k] * intWeights[k] * cval
                         Cgp = _eval_tensor_at_gp(coefficient, pa, elem, view(h, :, k))
-                        if Cgp isa Number
+                        if Cgp isa Number && weight === nothing
 
                             w = jacDet[k] * intWeights[k] * Cgp
 
@@ -821,7 +818,7 @@ function assemble_operator(
 
                             mul!(Ke, transpose(Bs), Bu, w, 1.0)
 
-                        else
+                        elseif Cgp isa AbstractMatrix && weight === nothing
 
                             w = jacDet[k] * intWeights[k]
 
@@ -831,30 +828,11 @@ function assemble_operator(
                             mul!(tmpBu, Cgp, Bu)
                             mul!(Ke, transpose(Bs), tmpBu, w, 1.0)
 
+                        else
+                            error("assemble_operator error")
                         end
-
-                        build_B!(Bu, op_u, Pu, k, h, ∂h, numNodes)
-
-                        # optional weighting of Bu (vector or matrix)
-                        if weight !== nothing
-                            if weight isa AbstractVector
-                                @inbounds for r in 1:size(Bu, 1)
-                                    wr = weight[r]
-                                    @inbounds for c in 1:size(Bu, 2)
-                                        Bu[r, c] *= wr
-                                    end
-                                end
-                            else
-                                mul!(tmpBu, weight, Bu)
-                                Bu .= tmpBu
-                            end
-                        end
-
-                        build_B!(Bs, op_s, Ps, k, h, ∂h, numNodes)
-
-                        # Ke += Bs' * Bu * w
-                        mul!(Ke, transpose(Bs), Bu, w, 1.0)
                     end
+
 
                     # scatter Ke(s,u) -> global IJV
                     @inbounds for a_loc in 1:ndofs_s_loc
@@ -862,20 +840,20 @@ function assemble_operator(
                         comp_a = mod(a_loc - 1, Ps.pdim) + 1
                         Ia_node = nnet[e, node_a]
                         Ia = (Ia_node - 1) * Ps.pdim + comp_a
-
+                
                         @inbounds for b_loc in 1:ndofs_u_loc
                             node_b = div(b_loc - 1, Pu.pdim) + 1
                             comp_b = mod(b_loc - 1, Pu.pdim) + 1
                             Jb_node = nnet[e, node_b]
[…]
