+++ src/multifield.jl
@@ -541,8 +541,66 @@ end
     return Dict(zip(p.numElem, p.A))  # elemTag => coeff nodal vector(s)
 end
 
-@inline function _coeff_at_gp(pa::Dict{<:Integer,<:AbstractMatrix}, elem::Integer, hcol::AbstractVector)
-    return dot(view(pa[elem], :, 1), hcol)
+"""
+    _prepare_coefficient(C)
+
+Prepare coefficient for assembly.
+
+Returns one of:
+
+Float64
+Dict(elem => nodal array)
+Matrix{Any} with entries Float64 or Dict(elem => nodal array)
+"""
+function _prepare_coefficient(C)
+
+    # scalar constant
+    if C isa Number
+        return Float64(C)
+    end
+
+    # scalar field
+    if C isa ScalarField
+        return _build_elemwise_coeff_dict(C)
+    end
+
+    # tensor coefficient
+    if C isa AbstractMatrix
+
+        W = Matrix{Any}(undef, size(C)...)
+
+        for I in CartesianIndices(C)
+
+            cij = C[I]
+
+            if cij isa Number
+                W[I] = Float64(cij)
+
+            elseif cij isa ScalarField
+                W[I] = _build_elemwise_coeff_dict(cij)
+
+            else
+                error("Matrix coefficient entries must be Number or ScalarField, got $(typeof(cij))")
+            end
+
+        end
+
+        return W
+    end
+
+    error("Unsupported coefficient type $(typeof(C))")
+end
+
+#@inline function _coeff_at_gp(pa::Dict{<:Integer,<:AbstractMatrix}, elem::Integer, hcol::AbstractVector)
+#    return dot(view(pa[elem], :, 1), hcol)
+#end
+
+@inline function _coeff_at_gp(
+    pa::Dict{<:Integer,<:AbstractMatrix},
+    elem::Integer,
+    hcol::AbstractVector)
+
+    return dot(view(pa[Int(elem)], :, 1), hcol)
 end
 
 """
@@ -590,6 +648,52 @@ function _eval_tensor_at_gp(C, pa, elem, hgp)
 
     error("Unsupported coefficient type $(typeof(C))")
 
+end
+
+"""
+    _eval_coefficient_at_gp(Cprep, elem, hgp)
+
+Evaluate prepared coefficient at Gauss point.
+
+Supports:
+Float64
+Dict(elem => nodal array)
+Matrix{Any} with entries Float64 or Dict(elem => nodal array)
+"""
+function _eval_coefficient_at_gp(Cprep, elem, hgp)
+
+    # scalar constant
+    if Cprep isa Number
+        return Cprep
+    end
+
+    # scalar field
+    if Cprep isa Dict
+        return _coeff_at_gp(Cprep, elem, hgp)
+    end
+
+    # tensor coefficient
+    if Cprep isa AbstractMatrix
+
+        m, n = size(Cprep)
+        W = Matrix{Float64}(undef, m, n)
+
+        for I in CartesianIndices(Cprep)
+
+            cij = Cprep[I]
+
+            if cij isa Number
+                W[I] = cij
+            else
+                W[I] = _coeff_at_gp(cij, elem, hgp)
+            end
+
+        end
+
+        return W
+    end
+
+    error("Unsupported prepared coefficient type $(typeof(Cprep))")
 end
 
 """
@@ -681,28 +785,11 @@ function assemble_operator(
         end
     end
 
-    # coefficient prep
-    pconst = 1.0
-    pa = Dict{Int,Any}()
-    if coefficient isa Number
-        pconst = Float64(coefficient)
+    # ------------------------------------------------------------
+    # prepare coefficient
[…]
