+++ src/LowLevelFEM.jl
@@ -28,7 +28,7 @@ include("extra.jl")
 
 export @showfields, @showstruct, @showdef, @showtype, @showmem, @showmethods, @disp, @showsize
 export probe_field
-
+#=
 @setup_workload begin
     @compile_workload begin
         mat = material("dummy")
@@ -136,5 +136,5 @@ export probe_field
         end
     end
 end
-
+=#
 end #module
