### 2026-03-10 c3da444
B'*A*C*D*B type operators
