### 2026-03-10 11b0dea
new point version
Matrix chain in weak form
