+++ src/general.jl
@@ -5686,7 +5686,7 @@ end
 Loads nodal results into a View in Gmsh. `q` is the field to show, `comp` is
 the component of the field (:vector, :uvec, :ux, :uy, :uz, :vvec, :vx, :vy, :vz,
 :qvec, :qx, :qy, :qz, :T, :p, :qn, :s, :sx, :sy, :sz, :sxy, :syx, :syz,
-:szy, :szx, :sxz, :e, :ex, :ey, :ez, :exy, :eyx, :eyz, :ezy, :ezx, :exz, :seqv, :scalar, :tensor),
+:szy, :szx, :sxz, :e, :ex, :ey, :ez, :exy, :eyx, :eyz, :ezy, :ezx, :exz, :x, :y, :z, :xy, :yz, :zx, :yx, :zy, :xz, :seqv, :scalar, :tensor),
 `name` is a title to display and `visible` is a Boolean to toggle the initial visibility in Gmsh on or off.
 If `q` has more columns, then a sequence of results will be shown (e.g., as an animation).
 `factor` multiplies the DoF result to increase for better visibility.
@@ -5753,7 +5753,7 @@ function showDoFResults(q::Union{ScalarField,VectorField,TensorField}, comp::Sym
     if size(q.a, 2) != length(t)
         error("showDoFResults: number of time steps missmatch ($(size(q.a,2)) <==> $(length(t))).")
     end
-    if comp == :s || comp == :e
+    if comp == :s || comp == :e || q isa TensorField
         pdim = 9
     end
     for j in 1:length(t)
@@ -5782,23 +5782,23 @@ function showDoFResults(q::Union{ScalarField,VectorField,TensorField}, comp::Sym
             end
         else
             nc = 1
-            if comp == :ux || comp == :vx || comp == :p || comp == :T || comp == :qx  || comp == :qn || comp == :sx || comp == :ex || comp == :scalar
+            if comp == :ux || comp == :vx || comp == :p || comp == :T || comp == :qx  || comp == :qn || comp == :sx || comp == :ex || comp == :scalar || (comp == :x && q isa VectorField) || ((comp == :x || comp == :xx) && q isa TensorField)
                 k = 1
-            elseif comp == :uy || comp == :vy || comp == :qy || comp == :syx || comp == :eyx
+            elseif comp == :uy || comp == :vy || comp == :qy || comp == :syx || comp == :eyx || (comp == :y && q isa VectorField) || (comp == :yx && q isa TensorField)
                 k = 2
-            elseif comp == :uz || comp == :vz || comp == :qz || comp == :szx || comp == :ezx
+            elseif comp == :uz || comp == :vz || comp == :qz || comp == :szx || comp == :ezx || (comp == :z && q isa VectorField) || (comp == :zx && q isa TensorField)
                 k = 3
-            elseif comp == :sxy || comp == :exy
+            elseif comp == :sxy || comp == :exy || (comp == :xy && q isa TensorField)
                 k = 4
-            elseif comp == :sy || comp == :ey
+            elseif comp == :sy || comp == :ey || ((comp == :y || comp == :yy) && q isa TensorField)
                 k = 5
-            elseif comp == :szy || comp == :ezy
+            elseif comp == :szy || comp == :ezy || (comp == :zy && q isa TensorField)
                 k = 6
-            elseif comp == :sxz || comp == :exz
+            elseif comp == :sxz || comp == :exz || (comp == :xz && q isa TensorField)
                 k = 7
-            elseif comp == :syz || comp == :eyz
+            elseif comp == :syz || comp == :eyz || (comp == :yz && q isa TensorField)
                 k = 8
-            elseif comp == :sz || comp == :ez
[…]
