### 2026-03-13 09c6c93
linear operators (loads)
