+++ src/multifield.jl
@@ -549,13 +549,22 @@ function build_B!(B::AbstractMatrix, op::AdvOp,
     return B
 end
 
-@inline function _build_elemwise_coeff_dict(coef::ScalarField)
-    p = nodesToElements(coef)
+struct DomainSpec
+    kind::Symbol     # :Ω vagy :Γ
+    name::String
+end
+
+Base.show(io::IO, d::DomainSpec) =
+    print(io, "$(d.kind)=\"$(d.name)\"")
+
+@inline function _build_elemwise_coeff_dict(coef::ScalarField, domain::DomainSpec)
+    p = elementsToNodes(coef)
+    p = nodesToElements(p, onPhysicalGroup=domain.name)
     return Dict(zip(p.numElem, p.A))  # elemTag => coeff nodal vector(s)
 end
 
 """
-    _prepare_coefficient(C)
+    _prepare_coefficient(C, domain)
 
 Prepare coefficient for assembly.
 
@@ -565,7 +574,7 @@ Float64
 Dict(elem => nodal array)
 Matrix{Any} with entries Float64 or Dict(elem => nodal array)
 """
-function _prepare_coefficient(C)
+function _prepare_coefficient(C, domain)
 
     # scalar constant
     if C isa Number
@@ -574,7 +583,7 @@ function _prepare_coefficient(C)
 
     # scalar field
     elseif C isa ScalarField
-        return _build_elemwise_coeff_dict(C)
+        return _build_elemwise_coeff_dict(C, domain)
     #end
 
     # tensor coefficient
@@ -590,7 +599,7 @@ function _prepare_coefficient(C)
                 W[I] = Float64(cij)
 
             elseif cij isa ScalarField
-                W[I] = _build_elemwise_coeff_dict(cij)
+                W[I] = _build_elemwise_coeff_dict(cij, domain)
 
             else
                 error("Matrix coefficient entries must be Number or ScalarField, got $(typeof(cij))")
@@ -602,8 +611,22 @@ function _prepare_coefficient(C)
     #end
 
     elseif C isa AbstractVector
-        mats = [_prepare_coefficient(M) for M in C]
+        mats = [_prepare_coefficient(M, domain) for M in C]
         return mats
+
+    #elseif C isa VectorField
+    #    if C.type == :v2D
+    #        nc = 2
+    #    elseif C.type == :v3D
+    #        nc = 3
+    #    else
+    #        error("Unknown VectorField type $(C.type)")
+    #    end
+    #    return [_prepare_coefficient(C[i]) for i in 1:nc]
+    #
+    #elseif C isa TensorField
+    #    return [_prepare_coefficient(C[i]) for i in 1:9]
+
     end
 
     error("Unsupported coefficient type $(typeof(C))")
@@ -614,6 +637,18 @@ end
 #end
 
 @inline function _coeff_at_gp(
+    pa::Dict{<:Integer,<:AbstractMatrix},
+    elem::Integer,
+    hcol::AbstractVector)
+
+    vals = get(pa, Int(elem), nothing)
+
+    vals === nothing && return 0.0
+
+    return dot(view(vals, :, 1), hcol)
+end
+
+@inline function _coeff_at_gp_old(
     pa::Dict{<:Integer,<:AbstractMatrix},
     elem::Integer,
     hcol::AbstractVector)
@@ -704,6 +739,27 @@ function _eval_coefficient_at_gp(Cprep, elem, hgp)
                 W[I] = cij
             else
                 W[I] = _coeff_at_gp(cij, elem, hgp)
+            end
+
+        end
+
+        return W
+    end
+
+    # vector coefficient
+    if Cprep isa AbstractVector
+
[…]
