+++ src/multifield.jl
@@ -355,7 +355,7 @@ Number of nodes per element.
 This function is called inside the element integration
 loop of `assemble_operator`.
 """
-function build_B!(B::AbstractMatrix, ::IdOp, P::Problem, k::Int, h, ∂h, numNodes::Int)
+function build_B!(B::AbstractMatrix, ::IdOp, P::Problem, k::Int, h, ∂h, numNodes::Integer)
     fill!(B, 0.0)
     pdim = P.pdim
     @inbounds for a in 1:numNodes
@@ -369,7 +369,7 @@ function build_B!(B::AbstractMatrix, ::IdOp, P::Problem, k::Int, h, ∂h, numNod
     return B
 end
 
-function build_B!(B::AbstractMatrix, ::GradOp, P::Problem, k::Int, h, ∂h, numNodes::Int)
+function build_B!(B::AbstractMatrix, ::GradOp, P::Problem, k::Int, h, ∂h, numNodes::Integer)
     fill!(B, 0.0)
     pdim = P.pdim
     dim = P.dim
@@ -399,7 +399,7 @@ function build_B!(B::AbstractMatrix, ::GradOp, P::Problem, k::Int, h, ∂h, numN
     return B
 end
 
-function build_B!(B::AbstractMatrix, ::DivOp, P::Problem, k::Int, h, ∂h, numNodes::Int)
+function build_B!(B::AbstractMatrix, ::DivOp, P::Problem, k::Int, h, ∂h, numNodes::Integer)
     fill!(B, 0.0)
     pdim = P.pdim
     dim = P.dim
@@ -417,7 +417,7 @@ function build_B!(B::AbstractMatrix, ::DivOp, P::Problem, k::Int, h, ∂h, numNo
 end
 
 function build_B!(B::AbstractMatrix, ::CurlOp,
-    P::Problem, k::Int, h, ∂h, numNodes::Int)
+    P::Problem, k::Int, h, ∂h, numNodes::Integer)
     fill!(B, 0.0)
     dim = P.dim
     pdim = P.pdim
@@ -463,7 +463,7 @@ function build_B!(B::AbstractMatrix, ::CurlOp,
 end
 
 function build_B!(B::AbstractMatrix, ::SymGradOp,
-    P::Problem, k::Int, h, ∂h, numNodes::Int)
+    P::Problem, k::Int, h, ∂h, numNodes::Integer)
     fill!(B, 0.0)
     dim = P.dim
     pdim = P.pdim
@@ -512,7 +512,7 @@ function build_B!(B::AbstractMatrix, ::SymGradOp,
 end
 
 function build_B!(B::AbstractMatrix, ::TensorDivOp,
-    P::Problem, k::Int, h, ∂h, numNodes::Int)
+    P::Problem, k::Int, h, ∂h, numNodes::Integer)
     fill!(B, 0.0)
     dim = P.dim
     pdim = P.pdim  # dim^2
@@ -532,7 +532,7 @@ function build_B!(B::AbstractMatrix, ::TensorDivOp,
 end
 
 function build_B!(B::AbstractMatrix, op::AdvOp,
-    P::Problem, k::Int, h, ∂h, numNodes::Int)
+    P::Problem, k::Int, h, ∂h, numNodes::Integer)
 
     fill!(B, 0.0)
     b = op.b
@@ -1119,23 +1119,28 @@ function assemble_linear(
 
     gmsh.model.setCurrent(P.name)
 
-    op isa IdOp || error("assemble_linear currently supports only Id(P) ⋅ g forms.")
+    outdim = op_outdim(op, P)
 
+    # allow matrix coefficient for tensor operators
+    if g isa AbstractMatrix
+        if length(g) != outdim
+            error("matrix coefficient size $(size(g)) incompatible with operator output dimension $outdim")
[…]
