### 2026-03-14 951d689
linear operators
