### 2026-03-14 10047a5
gmsh initialize in structured mesh
