### 2026-03-15 a0a0b4f
Linear term in weak form -> working, documented
