+++ src/multifield.jl
@@ -1051,6 +1051,19 @@ function assemble_operator(
     return SystemMatrix(K, Pu, Ps)
 end
 
+"""
+    matmul_sf(A, B)
+
+Matrix–matrix or matrix–field multiplication used internally
+when collapsing operator chains.
+
+Supports combinations of:
+- matrices
+- scalar/vector/tensor fields
+- numeric arrays
+
+Returns the resulting matrix or field.
+"""
 function matmul_sf(A, B)
 
     m,k = size(A)
@@ -1110,6 +1123,27 @@ function _field_pdim(f)
     end
 end
 
+"""
+    assemble_linear(P::Problem, op, rhs; weight=nothing, domain)
+
+Assemble a linear finite element operator of the form
+
+    ∫ v ⋅ op ⋅ rhs
+
+where `v` is the test field associated with problem `P`.
+
+Arguments
+---------
+- `P::Problem` : finite element problem definition
+- `op` : operator or matrix chain acting on the right-hand side
+- `rhs` : scalar, vector, tensor field, or numeric vector
+- `weight` : optional quadrature weight or coefficient
+- `domain` : domain specification (`Ω` or `Γ`)
+
+Returns
+-------
+Global right-hand side vector.
+"""
 function assemble_linear(
     P::Problem,
     op::AbstractOp,
@@ -1676,7 +1710,7 @@ struct BilinearTerm
 end
 
 """
-    LinearTerm(chain)
+    LinearTerm(chain::MatrixChain)
 
 Represents a linear weak-form term where the test-field operator
 appears once and the other factors are known coefficient fields
@@ -1687,6 +1721,9 @@ Examples
     f ⋅ Pu
     p ⋅ n ⋅ Pu
     σT ⋅ SymGrad(Pu)
+    Pu ⋅ A ⋅ g
+
+Used internally by the `∫` assembler.
 """
 struct LinearTerm
     chain
@@ -1708,7 +1745,8 @@ end
 # Operator combination
 # ------------------------------------------------------------
 
-function _matvec_sf(A, v)
+function _matvec_sf(A, v::AbstractVector)
+
     m, n = size(A)
     @assert n == length(v)
 
@@ -1726,82 +1764,48 @@ function _matvec_sf(A, v)
     return w
 end
 
+_to_components(v::Number) = [v]
 
+_to_components(v::AbstractVector) = v
+
+_to_components(v::ScalarField) = [v]
+
+_to_components(v::VectorField) = [v[i] for i in 1:v.model.pdim]
+
+function _to_components(T::TensorField)
+    if T.model.dim == 2
+        return [T[1], T[2], T[4], T[5]]   # 2×2 → 4 komponens
+    else
+        return [T[i] for i in 1:9]        # 3×3 → 9 komponens
+    end
+end
+
+# Collapse operator chain A⋅B⋅C⋅v → A(B(C(v)))
 function collapse_chain(mats, v)
-    w = v
+
+    w = _to_components(v)
+
     for i in length(mats):-1:1
         w = _matvec_sf(mats[i], w)
     end
+
     return w
 end
 
 """
-    ⋅(a,b)
+    MatrixChain
 
-Weak-form inner product operator used in the DSL.
-
-Constructs bilinear forms of the type
-
-    (Op_s v) ⋅ C ⋅ (Op_u u)
-
-Supported forms
-
-Operator pair
-
-    Grad(Pu) ⋅ Grad(Pu)
-
-Operator with tensor coefficient
-
-    SymGrad(Pu) ⋅ C ⋅ SymGrad(Pu)
-
-Matrix chain
-
-    Grad(Pu) ⋅ A ⋅ B ⋅ C ⋅ Grad(Pu)
[…]
