### 2026-03-16 cac71d1
adjoint for ScalarField
