+++ docs/src/reference/preprocessing.md
@@ -20,11 +20,14 @@ setParameter
 setParameters
 getParameter
 openPreProcessor
+openGeometry
 ```
 
 ## Boundary Conditions and Loads
 
 ```@docs
+BoundaryCondition
+LoadCondition
 displacementConstraint
 temperatureConstraint
 elasticSupport
