+++ src/general.jl
@@ -19,33 +19,80 @@ export ∂x, ∂y, ∂z, ∂t
 export structured_rect_mesh, structured_box_mesh, openGeometry
 
 """
-    Material(phName, type, E, ν, ρ, k, c, α, λ, μ, κ)
+    Material(phName; kwargs...)
 
-Structure containing the material type and constants.
-- type: constitutive law (`:Hooke`, `:StVenantKirchhoff`, `:NeoHookeCompressible`)
-- E: elastic modulus
-- ν: Poisson's ratio
-- ρ: mass density
-- k: thermal conductivity
-- c: specific heat
-- α: thermal expansion coefficient
-- λ: Lamé parameter
-- μ: Lamé parameter
-- κ: bulk modulus
-`phName` is the name of the physical group where the material is used.
+Material parameters assigned to a physical group.
 
-Types:
-- `phName`: String
-- `type`: Symbol
-- `E`: Float64
-- `ν`: Float64
-- `ρ`: Float64
-- `k`: Float64
-- `c`: Float64
-- `α`: Float64
-- `λ`: Float64
-- `μ`: Float64
-- `κ`: Float64
+This structure stores mechanical, thermal and fluid-related material
+properties used during assembly. The material is associated with a
+Gmsh physical group via `phName`.
+
+# Arguments
+
+- `phName::String`  
+  Name of the physical group where the material is applied.
+
+# Keyword Arguments
+
+## Elastic properties
+At least one valid combination of elastic constants must be provided.
+Missing parameters are automatically completed.
+
+- `E::Float64`  
+  Young's modulus
+- `ν::Float64`  
+  Poisson's ratio
+- `λ::Float64`  
+  First Lamé parameter
+- `μ::Float64`  
+  Second Lamé parameter (shear modulus)
+- `κ::Float64`  
+  Bulk modulus
+
+## Density
+- `ρ::Float64`  
+  Mass density
+
+## Thermal properties
+- `k::Float64`  
+  Thermal conductivity
+- `c::Float64`  
+  Specific heat capacity
+- `α::Float64`  
+  Thermal expansion coefficient
+
+## Fluid / tribology-related properties
+- `η::Float64`  
+  Dynamic viscosity
+- `p₀::Float64`  
+  Cavitation pressure
+
+## Structural property
+- `A::Float64`  
+  Cross-sectional area (used e.g. in truss elements)
+
+# Notes
+
+- Elastic constants are internally completed to a consistent set
+  using `elastic_constants`.
+- Default values are provided for all optional parameters.
+- The material definition is independent of the field type
+  (scalar, vector, tensor) and can be used in multifield problems.
+
+# Examples
+
+```julia
+mat = Material("solid", E=210e3, ν=0.3)
+
+mat = Material("fluid",
+    η=1e-6,
+    p₀=0.1
+)
+
+mat = Material("truss",
+    E=210e3,
+    A=1.0
+)
 """
 struct Material
     phName::String
@@ -213,33 +260,112 @@ mutable struct Geometry
 end
 
 """
-    Problem(materials; thickness=..., type=..., bandwidth=..., dim=..., fdim=...)
+    Problem(materials; kwargs...)
 
-Structure containing key data for a problem.
-- Parts of the model with their material constants. Multiple materials can be provided (see `material`).
-- Problem type: `:Solid`, `:PlaneStrain`, `:PlaneStress`, `:AxiSymmetric`, `:HeatConduction`, `:PlaneHeatConduction`, 
-`:AxiSymmetricHeatConduction`, `:Truss`, `:General`.
-  For `:AxiSymmetric`, the symmetry axis is `y`, and the geometry must be drawn in the positive `x` half-plane.
[…]
