+++ src/multifield.jl
@@ -2436,3 +2436,592 @@ KΓ = ∫Γ("loaded_boundary", Id(Pu) ⋅ Id(Pu))
 ∫Γ(name, expr) = ∫(expr; Γ=name)
 
 const ε = SymGrad
+
+ndofs(P) = P.non * P.pdim
+
+"""
+    multifield_bc_data(K::SystemMatrix,
+                       bc::Vector{BoundaryCondition};
+                       nsteps::Int=1)
+
+Compute global constrained and free DOFs together with the prescribed
+Dirichlet values for a multifield block system.
+
+Returns
+- `free  :: Vector{Int}`
+- `fixed :: Vector{Int}`
+- `xD    :: Matrix{Float64}` of size `(ndof, nsteps)`
+
+Notes
+- This function requires `K` to be a block `SystemMatrix`, i.e.
+  `K.problems !== nothing` and `K.offsets !== nothing`.
+- In multifield problems every `BoundaryCondition` must explicitly
+  refer to a `problem`.
+- Prescribed values are assembled from `applyBoundaryConditions(problem, [bc]; steps=nsteps)`.
+"""
+function multifield_bc_data(
+    K::SystemMatrix,
+    bc::Vector{BoundaryCondition};
+    nsteps::Int=1
+)
+    K.problems === nothing &&
+        error("multifield_bc_data: K is not a block SystemMatrix (K.problems === nothing).")
+
+    K.offsets === nothing &&
+        error("multifield_bc_data: K is not a block SystemMatrix (K.offsets === nothing).")
+
+    problems = K.problems
+    offsets = K.offsets
+
+    length(problems) == length(offsets) ||
+        error("multifield_bc_data: inconsistent metadata in K (length(problems) != length(offsets)).")
+
+    ndof = size(K.A, 1)
+    size(K.A, 1) == size(K.A, 2) ||
+        error("multifield_bc_data: K must be square.")
+
+    fixed = Int[]
+    xD = zeros(ndof, nsteps)
+
+    for bci in bc
+        bci.problem === nothing &&
+            error("multifield_bc_data: in multifield systems every BoundaryCondition must have an explicit `problem`.")
+
+        idx = findfirst(P -> P === bci.problem, problems)
+        idx === nothing &&
+            error("multifield_bc_data: BC refers to a Problem that is not present in K.problems.")
+
+        P = problems[idx]
+        off = offsets[idx]
+
+        local_fixed = constrainedDoFs(P, [bci])
+        isempty(local_fixed) && continue
+
+        global_fixed = off .+ local_fixed
+        append!(fixed, global_fixed)
+
+        field_bc = applyBoundaryConditions(P, [bci]; steps=nsteps)
+
+        # safety checks
+        size(field_bc.a, 1) == ndofs(P) ||
+            error("multifield_bc_data: applyBoundaryConditions returned incompatible field size for problem $(P.field).")
+
+        if field_bc.nsteps == 1
+            @inbounds for it in 1:nsteps
+                xD[global_fixed, it] .= field_bc.a[local_fixed, 1]
+            end
+        else
+            field_bc.nsteps == nsteps ||
+                error("multifield_bc_data: BC field nsteps = $(field_bc.nsteps), expected 1 or $nsteps.")
+            xD[global_fixed, :] .= field_bc.a[local_fixed, 1:nsteps]
+        end
+    end
+
+    fixed = unique(fixed)
+    sort!(fixed)
+
+    free = setdiff(collect(1:ndof), fixed)
+
+    return free, fixed, xD
+end
+
+
+"""
+    multifield_constrainedDoFs(K::SystemMatrix,
+                               bc::Vector{BoundaryCondition})
+
+Return global constrained DOFs for a multifield block system.
+"""
+function multifield_constrainedDoFs(
[…]
