### 2026-03-24 658e66f
FDM for multifield + truss with operators
