+++ src/multifield.jl
@@ -59,6 +59,8 @@ export SymGrad
 export Id
 export TensorDiv
 export Adv
+export AxialGrad
+export TangentialGrad
 
 export ∫
 export ∫Ω
@@ -69,6 +71,8 @@ export solveField
 export ε
 
 export ⋅
+
+export solveEigenFields
 
 
 """
@@ -245,12 +249,57 @@ struct AdvOp <: AbstractOp
     b::NTuple{3,Float64}
 end
 
-#"""
-#    ndofs(problem::Problem)
-#
-#Return total number of dofs for a single-field problem.
-#"""
-#ndofs(P::Problem) = P.non * P.pdim
+"""
+    AxialGradOp()
+
+Axial (directional) gradient operator.
+
+For vector fields:
+    AxialGrad(u) = t ⋅ ∇u
+
+Returns a scalar field representing the derivative along the element axis.
+"""
+struct AxialGradOp <: AbstractOp end
+
+#AxialGrad(P) = AxialGradOp()
+
+"""
+    TangentialGrad(P)
+
+Create a weak-form DSL tangential gradient operator applied to `P`.
+
+# Description
+The tangential gradient operator computes the projection of the gradient of a vector field
+onto the element axis:
+
+    TangentialGrad(u) = (∇u) ⋅ t
+
+where `t` is the tangent vector of the element in the physical domain.
+
+For vector fields `u ∈ ℝᵈ`, the result is a vector field of dimension `d`.
+
+# Arguments
+- `P`: Field descriptor (`Problem`) used in the weak form.
+
+# Returns
+- `OpApplied`: Operator application object representing `(∇u) ⋅ t`.
+
+# Example
+```julia
+Kg = ∫(TangentialGrad(Pu) ⋅ N0 ⋅ TangentialGrad(Pu); Ω="truss")
+```
+
+# Notes
+
+- Unlike `AxialGrad`, which produces a scalar strain measure `t ⋅ ∇u ⋅ t`,  
+  `TangentialGrad` returns a vector quantity.
+
+- This operator is useful for constructing geometric stiffness matrices  
+  (initial stress effects) in truss and structural stability problems.  
+"""
+struct TangentialGradOp <: AbstractOp end
+
+TangentialGrad(P) = OpApplied(P, TangentialGradOp())
 
 """
     op_outdim(::IdOp, P::Problem)
@@ -308,6 +357,14 @@ end
 function op_outdim(op::AdvOp, P::Problem)
     @assert P.pdim == 1  # scalar field
     return 1
+end
+
+function op_outdim(::AxialGradOp, P::Problem)
+    return 1
+end
+
+function op_outdim(::TangentialGradOp, P::Problem)
+    return P.dim
 end
 
 """
@@ -544,6 +601,74 @@ function build_B!(B::AbstractMatrix, op::AdvOp,
             val += b[d] * ∂h[d, (k-1)*numNodes+a]
         end
         B[1, a] = val
+    end
+
+    return B
+end
+
+function get_tangent(P::Problem, k::Int)
+    if P.dim == 2
+        return (1.0, 0.0)
+    elseif P.dim == 3
+        return (1.0, 0.0, 0.0)
+    else
+        return (1.0,)
+    end
+end
+
+function compute_tangent(Jk, dim)
+    t = Jk[1:dim,1]
+    return t / norm(t)
+end
+
+function build_B!(B::AbstractMatrix, ::AxialGradOp,
+                  P::Problem, k::Int, h, ∂h, numNodes::Integer, t)
+
+    fill!(B, 0.0)
+
+    pdim = P.pdim   # 2
+    dim  = P.dim    # 2
+
+    @inbounds for a in 1:numNodes
[…]
