+++ src/multifield.jl
@@ -345,9 +345,21 @@ function op_outdim(::CurlOp, P::Problem)
 end
 
 function op_outdim(::SymGradOp, P::Problem)
-    @assert P.pdim == P.dim "SymGradOp requires vector field with pdim == dim."
-    return (P.dim == 2) ? 3 : 6  # engineering strain components
+    @assert P.pdim == P.dim
+
+    if P.dim == 1
+        return 1
+    elseif P.dim == 2
+        return 3
+    else
+        return 6
+    end
 end
+
+#function op_outdim(::SymGradOp, P::Problem)
+#    @assert P.pdim == P.dim "SymGradOp requires vector field with pdim == dim."
+#    return (P.dim == 2) ? 3 : 6  # engineering strain components
+#end
 
 function op_outdim(::TensorDivOp, P::Problem)
     @assert P.pdim == P.dim^2 "TensorDivOp requires pdim == dim^2 (full 2nd-order tensor)."
@@ -526,7 +538,14 @@ function build_B!(B::AbstractMatrix, ::SymGradOp,
     pdim = P.pdim
     @assert pdim == dim
 
-    if dim == 2
+    if dim == 1
+        # ε = du/dx
+        @inbounds for a in 1:numNodes
+            col = a  # scalar field
+            B[1, col] = ∂h[1, (k-1)*numNodes + a]
+        end
+
+    elseif dim == 2
         # rows: [εxx, εyy, γxy]
         @inbounds for a in 1:numNodes
             colx = (a - 1) * pdim + 1
@@ -991,8 +1010,8 @@ function assemble_operator(
         @assert out_u == out_s
     
     elseif coefficient isa AbstractMatrix
-        @assert size(coefficient,1) == out_u
-        @assert size(coefficient,2) == out_s
+        @assert size(coefficient,1) == out_s
+        @assert size(coefficient,2) == out_u
     
     elseif coefficient isa AbstractVector
         if length(coefficient) == 1 &&
@@ -1007,8 +1026,8 @@ function assemble_operator(
             @assert A1 isa AbstractMatrix
             @assert An isa AbstractMatrix
     
-            @assert size(A1,1) == out_u
-            @assert size(An,2) == out_s
+            @assert size(A1,1) == out_s
+            @assert size(An,2) == out_u
     
             for i in 1:length(coefficient)-1
                 Ai = coefficient[i]
@@ -1204,14 +1223,7 @@ function assemble_operator(
                                 build_B!(Bs, op_s, Ps, k, h, ∂h, numNodes)
                             end
 
-                            #@disp size(Bu)
-                            #@disp size(Cgp)
-                            #@disp size(Bs)
-                            #tmp = similar(Bu, size(Cgp,1), size(Bu,2))
-                            #tmp = zeros(size(Bs,2), size(Cgp,1))
-                            #mul!(tmpBu, Cgp, Bu)
-                            #mul!(Ke, transpose(Bs), tmpBu, w, 1.0)
-                            mul!(tmp, transpose(Bs), transpose(Cgp))
+                            mul!(tmp, transpose(Bs), Cgp) # Here Cgp was transposed
                             mul!(Ke, tmp, Bu, w, 1.0)
 
                         else
@@ -1232,8 +1244,11 @@ function assemble_operator(
                             comp_b = mod(b_loc - 1, Pu.pdim) + 1
                             Jb_node = nnet[e, node_b]
                             Jb = (Jb_node - 1) * Pu.pdim + comp_b
-                
-                            if pos > length(I)
+                            
+                            #@assert pos >= 1
[…]
