### 2026-03-31 22012c3
Multiplication to solve energy
e.g.:
kene = v * M * v / 2
sene = u * K * u / 2
diss = v * C * v
