+++ src/linear.jl
@@ -5683,96 +5683,116 @@ end
     CDM(K::SystemMatrix, 
         M::SystemMatrix, 
         C::SystemMatrix, 
-        f::VectorField, 
+        f::Union{ScalarField,VectorField}, 
         bc::Vector{BoundaryCondition}, 
-        u0::VectorField, 
-        v0::VectorField, 
+        u0::Union{ScalarField,VectorField}, 
+        v0::Union{ScalarField,VectorField}, 
         n::Int, 
         Δt::Float64)
+
 aliases:
 
-    CDM(K, M, f, bc, u0, v0, n, Δt)
+    CDM(K, M, C, f, bc, u0, v0, n, Δt)
     CDM(K, M, C, f, u0, v0, n, Δt; support=BoundaryCondition[])
-    CDM(K, M, f, u0, v0, n, Δt; support=BoundaryCondition[])
 
-Solves a transient structural dynamic problem using the **central difference method (CDM)**,
-an explicit time integration scheme. (support ≡ bc)
+Solves a transient problem using an **explicit central difference–type scheme (CDM)**.
 
 The semi-discrete system
-```
 
+```
 M * ü(t) + C * u̇(t) + K * u(t) = f(t)
-
 ```
-is advanced in time using central finite differences.
-The method supports:
+
+is advanced in time using finite differences with a **lumped (diagonal) mass matrix**.
+
+The implementation supports:
+
 - time-independent or time-dependent load vectors `f`
 - time-independent or time-dependent displacement boundary conditions
-- consistent treatment of constraint-induced inertia and damping terms
-- vector-valued unknowns (`VectorField`)
+- scalar and vector unknowns (`ScalarField`, `VectorField`)
+- consistent incorporation of prescribed displacements into the right-hand side
 
 Boundary conditions are applied *solver-side*:
-on constrained DOFs the prescribed displacements override the initial conditions,
-while on free DOFs the initial displacement `u0` and velocity `v0` are used.
-Velocities on constrained DOFs are derived from the prescribed displacements.
 
-If `f.nsteps == 1`, the load is treated as time-independent.
-If `f.nsteps == n`, the load is applied time step–by–time step.
+- constrained DOFs are prescribed directly from `bc`
+- free DOFs evolve according to the dynamic equation
+- velocities on constrained DOFs are computed from the prescribed displacement history
+
+The effect of prescribed displacements is included through
+
+```
+- K_fc * u_c - C_fc * u̇_c
+```
+
+terms in the right-hand side.
+
+If `f.nsteps == 1`, the load is treated as time-independent.  
+If `f.nsteps == n`, the load is applied step-by-step.
 
 ---
 
 ### Arguments
+
 - `K::SystemMatrix`  
   Global stiffness matrix.
 - `M::SystemMatrix`  
-  Global mass matrix (assumed diagonal or lumped for efficiency).
+  Global mass matrix (must be diagonal / lumped).
 - `C::SystemMatrix`  
-  Global damping matrix. If omitted, zero damping is assumed.
-- `f::VectorField`  
+  Global damping matrix.
+- `f::Union{ScalarField,VectorField}`  
   External load vector (time-independent or time-dependent).
 - `bc::Vector{BoundaryCondition}`  
-  Displacement-type boundary conditions (possibly time-dependent).
-- `u0::VectorField`  
-  Initial displacement field. Overridden on constrained DOFs.
-- `v0::VectorField`  
[…]
