+++ src/multifield.jl
@@ -2839,40 +2839,108 @@ function split_multifield_solution(
 end
 
 """
-    FDM(K::SystemMatrix,
-        C::SystemMatrix,
-        q::SystemVector,
-        bc::Vector{BoundaryCondition},
-        X0::SystemVector,
-        n::Int,
-        Δt::Float64;
+    FDM(K::SystemMatrix, 
+        C::SystemMatrix, 
+        q::Union{ScalarField,VectorField,TensorField,SystemVector}, 
+        bc::Vector{BoundaryCondition}, 
+        X0::Union{ScalarField,VectorField,TensorField,SystemVector}, 
+        n::Int, 
+        Δt::Float64; 
         ϑ=0.5)
 
-Alias: FDM(K, C, q, X0, n, Δt; ϑ=0.5, support=Vector{BoundaryCondition}())
+alias:
 
-Time integration for a multifield first-order system
+    FDM(K, C, q, X0, n, Δt; ϑ=0.5, support=bc)
 
-    C * ẋ + K * x = q
+Solves a transient diffusion-type problem (e.g. heat conduction) using the
+finite difference method in time (ϑ-method).
 
-using the theta-method.
+The semi-discrete system
+```
 
-Arguments
-- `K`: system matrix
-- `C`: capacity / mass-like matrix
-- `q`: global RHS vector as `SystemVector`
-- `bc`: Dirichlet boundary conditions
-- `X0`: initial state as `SystemVector`
-- `n`: number of stored time steps
-- `Δt`: time step size
+C * Ẋ(t) + K * X(t) = q(t)
 
-Keyword arguments
-- `ϑ`: theta parameter
-    - `0.0` explicit Euler
-    - `0.5` Crank-Nicolson
-    - `1.0` implicit Euler
+```
+is integrated in time using the ϑ-scheme:
+- ϑ = 0   : Forward Euler (explicit)
+- ϑ = 1/2 : Crank–Nicolson
+- ϑ = 1   : Backward Euler (implicit)
+- 0 < ϑ < 1 : intermediate schemes
 
-Returns
-Tuple of fields in the block order of `K.problems`.
+The method supports:
+- time-independent or time-dependent load vectors `q`
+- time-independent or time-dependent Dirichlet boundary conditions
+- scalar, vector and tensor unknowns (`ScalarField`, `VectorField`, `TensorField`)
+- multi-field block systems (`SystemVector`)
+- consistent treatment of prescribed DOFs via matrix partitioning
+
+Boundary conditions are applied *solver-side*:
+- constrained DOFs are prescribed directly at each time step
+- free DOFs are solved from the reduced system
+- contributions of constrained DOFs enter the RHS through
+```
+
+* K_fc X_c - C_fc Ẋ_c
+
+```
+
+If `q.nsteps == 1`, the load is treated as time-independent.  
+If `q.nsteps == n`, a true ϑ-weighted load
+```
+
+q^{n+ϑ} = (1-ϑ) q^n + ϑ q^{n+1}
+
+```
+is used.
+
+For `ϑ = 0` and diagonal `C`, a fully explicit update is used.
+
+---
+
+### Arguments
+- `K::SystemMatrix`  
+  Diffusion (conductivity / stiffness) matrix.
+- `C::SystemMatrix`  
+  Capacity (mass / heat capacity) matrix.
+- `q`  
+  Load / source vector:
+  - single-field: `ScalarField`, `VectorField`, `TensorField`
+  - multi-field: `SystemVector`
+- `bc::Vector{BoundaryCondition}`  
+  Dirichlet boundary conditions (possibly time-dependent).
+- `X0`  
+  Initial condition (same type as `q`).  
[…]
