+++ src/linear.jl
@@ -2820,7 +2820,8 @@ function dampingMatrix(K, M, ωₘₐₓ; α=0.0, ξ=0.01, β=[2ξ[i]/(ωₘₐ�
     end
     dof, dof = size(M.A)
     dof2, dof2 = size(K.A)
-    if dof != nnz(M.A)
+    #if dof != nnz(M.A)
+    if !isdiag(M.A)
         error("dampingMatrix: M is not lumped!")
     end
     if dof != dof2
@@ -5826,7 +5827,7 @@ function CDM(K::SystemMatrix, M::SystemMatrix, C::SystemMatrix, f::Union{ScalarF
     vv = copy(uu)
     applyBoundaryConditions!(uu, bc)
     for i in 1:n-1
-        vv.a[fix,i] = (uu.a[fix,i+1] - uu.a[fix,i]) / Δt
+        vv.a[fix,i] .= (uu.a[fix,i+1] .- uu.a[fix,i]) ./ Δt
     end
     if n >= 3
         vv.a[fix,end] .= 2.0 .* vv.a[fix,end-1] .- vv.a[fix,end-2]
@@ -5837,8 +5838,8 @@ function CDM(K::SystemMatrix, M::SystemMatrix, C::SystemMatrix, f::Union{ScalarF
     uu.a[free, 1] .= DoFs(U0)[free, 1]
     vv.a[free, 1] .= DoFs(V0)[free, 1]
 
-    isdiag = size(M.A, 1) == nnz(M.A) ? true : false
-    if !isdiag
+    #isdiag = size(M.A, 1) == nnz(M.A) ? true : false
+    if !isdiag(M.A)
         error("CDM: only lumped M is supported.")
     end
 
@@ -5853,7 +5854,10 @@ function CDM(K::SystemMatrix, M::SystemMatrix, C::SystemMatrix, f::Union{ScalarF
 
     ff = f.a[free, :]
 
-    invMff = spdiagm(1.0 ./ diag(Mff))
+    d = diag(Mff)
+    all(abs.(d) .> 0) || error("CDM: zero diagonal entry detected in Mff on free DOFs.")
+    invMff = spdiagm(1.0 ./ d)
+    #invMff = spdiagm(1.0 ./ diag(Mff))
 
     nsteps = n
     
@@ -5861,12 +5865,27 @@ function CDM(K::SystemMatrix, M::SystemMatrix, C::SystemMatrix, f::Union{ScalarF
     v = vv.a
     t = ts
     
-    a0 = M.A \ (f.a[:,1] - K.A * u[:,1] - C.A * v[:,1])
-    u00 = u[:,1] - v[:,1] * Δt + a0 * Δt^2 / 2
-    u00[fix] .= 2.0 .* u[fix,1] .- u[fix,2]
+    u00 = zeros(eltype(u), size(u,1))
+
+    a0 = Mff \ (
+        f.a[free,1] .-
+        Kff * u[free,1] .-
+        Kfc * u[fix,1] .-
+        Cff * v[free,1] .-
+        Cfc * v[fix,1])
+
+    u00[free] .= u[free,1] .- v[free,1] .* Δt .+ 0.5 .* a0 .* Δt^2
+    u00[fix]  .= 2.0 .* u[fix,1] .- u[fix,2]
+
+    #a0 = M.A \ (f.a[:,1] - K.A * u[:,1] - C.A * v[:,1])
+    #u00 = u[:,1] - v[:,1] * Δt + a0 * Δt^2 / 2
+    #a0 = M.A[free,free] \ (f.a[free,1] - K.A[free,free] * u[free,1] - C.A[free,free] * v[free,1])
[…]
