### 2026-04-08 319203c
vector and tensor fields with mixed Number or ScalarField component in constructors
