### 2026-04-09 perebalazs/LowLevelFEM.jl#35 @cfarm6
Joss Review: Good Practices - Testing
The tests for the package only cover a single simulation. The tests should be developed further to cover more of the functionality of the package and have a finer resolution in the event that a change breaks functionality in a portion of the code.

[Review](https://github.com/openjournals/joss-reviews/issues/10096)
