+++ src/operators.jl
@@ -2318,6 +2318,7 @@ end
 """
     *(A::SystemMatrix, b::Number)
     *(b::Number, A::SystemMatrix)
+    /(A::SystemMatrix, b::Number)
 
 Scalar multiplication of a system matrix.
 """
@@ -2326,6 +2327,9 @@ function *(A::SystemMatrix, b::Number)
 end
 function *(b::Number, A::SystemMatrix)
     SystemMatrix(A.A * b, A.model, A.test_model, A.problems, A.offsets)
+end
+function /(A::SystemMatrix, b::Number)
+    SystemMatrix(A.A / b, A.model, A.test_model, A.problems, A.offsets)
 end
 
 
