+++ v1.13.10/siteinfo.js
@@ -1,0 +1,2 @@
+var DOCUMENTER_CURRENT_VERSION = "v1.13.10";
+var DOCUMENTER_IS_DEV_VERSION = false;
