### 2026-05-22 5e72b7d
bug fixes in operators.jl
