+++ src/operators.jl
@@ -54,8 +54,13 @@ function *(A::ScalarField, B::ScalarField)
         AAi = A.A[ia]
         BBi = B.A[ib]
 
-        # kisebbik csomópontszám a biztonság kedvéért
-        n = min(size(AAi, 1), size(BBi, 1))
+        nA = size(AAi, 1)
+        nB = size(BBi, 1)
+        nA == nB || error(
+            "*(ScalarField, ScalarField): incompatible local sizes on element $e: " *
+            "A rows=$nA, B rows=$nB, size(A)=$(size(AAi)), size(B)=$(size(BBi))")
+
+        n = nA
         @views C[ii] = AAi[1:n, :] .* BBi[1:n, :]
         num[ii] = e
     end
@@ -147,7 +152,13 @@ function /(AA::ScalarField, BB::ScalarField)
         AAi = A.A[ia]
         BBi = B.A[ib]
 
-        n = min(size(AAi, 1), size(BBi, 1))
+        nA = size(AAi, 1)
+        nB = size(BBi, 1)
+        nA == nB || error(
+            "/(ScalarField, ScalarField): incompatible local sizes on element $e: " *
+            "A rows=$nA, B rows=$nB, size(A)=$(size(AAi)), size(B)=$(size(BBi))")
+
+        n = nA
         @views C[ii] = AAi[1:n, :] ./ BBi[1:n, :]
         num[ii] = e
     end
@@ -736,15 +747,21 @@ function *(AA::ScalarField, BB::VectorField)
         AAi = A.A[ia]
         BBi = B.A[ib]
 
-        n_nodes = min(size(AAi, 1), size(BBi, 1) ÷ 3)
+        nA = size(AAi, 1)
+        nB = fld(size(BBi, 1), 3)
+        nA == nB || error(
+            "*(ScalarField, VectorField): incompatible local block counts on element $e: " *
+            "scalar rows=$nA, vector blocks=$nB, size(A)=$(size(AAi)), size(B)=$(size(BBi))")
+
+        n_nodes = nA
         rows = 3n_nodes
 
         result = Matrix{Float64}(undef, rows, nsteps)
 
         if n_nodes > 0
             @inbounds @views for j in 1:n_nodes
-                dest = result[3j-2:3j, :]
-                src = BBi[3j-2:3j, :]
+                dest = result[3*j-2:3*j, :]
+                src = BBi[3*j-2:3*j, :]
                 if a_has_single_step
                     dest .= src .* AAi[j, 1]
                 else
@@ -806,15 +823,21 @@ function /(BB::VectorField, AA::ScalarField)
         AAi = A.A[ia]
         BBi = B.A[ib]
 
-        n_nodes = min(size(AAi, 1), size(BBi, 1) ÷ 3)
+        nA = size(AAi, 1)
+        nB = fld(size(BBi, 1), 3)
+        nA == nB || error(
+            "/(VectorField, ScalarField): incompatible local block counts on element $e: " *
+            "scalar rows=$nA, vector blocks=$nB, size(A)=$(size(AAi)), size(B)=$(size(BBi))")
+
+        n_nodes = nA
         rows = 3n_nodes
 
         result = Matrix{Float64}(undef, rows, nsteps)
 
         if n_nodes > 0
             @inbounds @views for j in 1:n_nodes
-                dest = result[3j-2:3j, :]
-                src = BBi[3j-2:3j, :]
+                dest = result[3*j-2:3*j, :]
+                src = BBi[3*j-2:3*j, :]
                 if a_has_single_step
                     dest .= src ./ AAi[j, 1]
                 else
[…]
