+++ src/LowLevelFEM.jl
@@ -62,16 +62,18 @@ export probe_field
         tsteps = [1.0]
         sfA = ScalarField([reshape(collect(1.0:3.0), 3, 1)], [;;], tsteps, [1], 1, :scalar, prob)
         sfB = ScalarField([reshape(fill(2.0, 3), 3, 1)], [;;], tsteps, [1], 1, :scalar, prob)
-        vfA = VectorField([reshape(collect(1.0:3.0), 3, 1)], [;;], tsteps, [1], 1, :v3D, prob)
-        vfB = VectorField([reshape(collect(2.0:4.0), 3, 1)], [;;], tsteps, [1], 1, :v3D, prob)
+        vfA = VectorField([reshape(collect(1.0:9.0), 9, 1)], [;;], tsteps, [1], 1, :v3D, prob)
+        vfB = VectorField([reshape(collect(2.0:10.0), 9, 1)], [;;], tsteps, [1], 1, :v3D, prob)
         tf_block_I = reshape([1.0, 0.0, 0.0,
                               0.0, 1.0, 0.0,
                               0.0, 0.0, 1.0], 9, 1)
         tf_block_diag = reshape([2.0, 0.0, 0.0,
                                  0.0, 3.0, 0.0,
                                  0.0, 0.0, 4.0], 9, 1)
-        tfA = TensorField([tf_block_I], [;;], tsteps, [1], 1, :e, prob)
-        tfB = TensorField([tf_block_diag], [;;], tsteps, [1], 1, :e, prob)
+        tfA = TensorField([repeat(tf_block_I, 3, 1)], [;;], tsteps, [1], 1, :e, prob)
+        tfB = TensorField([repeat(tf_block_diag, 3, 1)], [;;], tsteps, [1], 1, :e, prob)
+        #tfA = TensorField([tf_block_I], [;;], tsteps, [1], 1, :e, prob)
+        #tfB = TensorField([tf_block_diag], [;;], tsteps, [1], 1, :e, prob)
 
         _ = sfA + sfB
         _ = sfA - sfB
