+++ src/nonlinear.jl
@@ -6,7 +6,8 @@ export nodePositionVector, ∇, curl, rot, div, grad
 export showDeformationResults
 export grad_xy
 export materialTangentMatrix, initialStressMatrix, externalTangentFollower, internalForceVector
-export IIPiolaKirchhoff
+export IIPiolaKirchhoffStress
+export energyMaterial, tensorToMatrix, tensorToVector, IIPiolaKirchhoff, tangentMatrix
 
 """
     nodePositionVector(problem)
@@ -3260,7 +3261,7 @@ function externalTangentFollower(
 end
 
 """
-    IIPiolaKirchhoff(energy, C::TensorField, params)
+    IIPiolaKirchhoffStress(energy, C::TensorField, params)
 
 Computes the nodal Second Piola–Kirchhoff (II PK) stress tensor field
 from a given nodal Right Cauchy–Green deformation tensor field.
@@ -3301,7 +3302,7 @@ Notes:
 - For use in weak forms or Newton iterations, Gauss-point evaluation
   should be used instead.
 """
-function IIPiolaKirchhoff(
+function IIPiolaKirchhoffStress(
     energy::Function,
     C::TensorField,
     params::NamedTuple)
@@ -3717,3 +3718,284 @@ function loadVectorNew(problem, loads; F=nothing)
     end
 end
 =#
+
+
+################################################################################
+### hyperelastic tangent matrix and II P-K stress for DSL ######################
+################################################################################
+
+const VOIGT3D = ((1, 1), (2, 2), (3, 3), (1, 2), (2, 3), (3, 1))
+
+@inline function _prepare_energy_params_elementwise(params::NamedTuple, C::TensorField)
+    names = keys(params)
+    vals = values(params)
+
+    newvals = map(vals) do v
+        if v isa Number
+            v
+        elseif v isa ScalarField
+            ve = isElementwise(v) ? v : nodesToElements(v)
+
+            if ve.nsteps != 1 && ve.nsteps != C.nsteps
+                error("energyMaterial: parameter nsteps must be 1 or equal to C.nsteps.")
+            end
+
+            return Dict(zip(ve.numElem, ve.A))
+        else
+            error("energyMaterial: parameters must be Number or ScalarField, got $(typeof(v)).")
+        end
+    end
+
+    return NamedTuple{names}(newvals)
+end
+
+
+@inline _param_value_at_element_node(v::Number, elem::Int, a::Int, it::Int) = v
+
+@inline function _param_value_at_element_node(v::Dict, elem::Int, a::Int, it::Int)
+    A = v[elem]
+    itt = size(A, 2) == 1 ? 1 : it
+    return A[a, itt]
+end
+
+
+@inline function _params_at_element_node(params::NamedTuple, elem::Int, a::Int, it::Int)
+    names = keys(params)
+    vals = values(params)
+
+    newvals = map(vals) do v
+        _param_value_at_element_node(v, elem, a, it)
+    end
+
+    return NamedTuple{names}(newvals)
+end
+
+"""
+    tensorToVector(T::TensorField)
+
+Convert a symmetric tensor field to Voigt vector form.
+
+# Arguments
+- `T`: symmetric tensor field
+
+# Returns
+A vector of `ScalarField`s in the ordering:
+
+    [xx, yy, zz, xy, yz, zx]
+
+# Examples
+```julia
+v = tensorToVector(S)
+```
+"""
+function tensorToVector(T::TensorField)
+    return [T[i, j] for (i, j) in VOIGT3D]
+end
+
+"""
+    tensorToMatrix(T::TensorField)
+
+Convert a tensor field to a matrix of scalar fields.
[…]
