+++ docs/src/reference/multifield.md
@@ -2,6 +2,395 @@
 
 Weak-form DSL and multifield assembly interface.
 
+## Operator size and component ordering
+
+This section summarizes the output size and component ordering of the basic
+differential operators used by the weak-form DSL.
+
+All operator vectors are written in the order used internally by LowLevelFEM.
+
+---
+
+## Scalar fields
+
+Let
+
+```math
+p = p(x,y,z)
+```
+
+be a scalar field.
+
+| Dimension | Operator  | Size | Component ordering |
+| ---------:| --------- | ----:| ------------------ |
+| 1D        | `Grad(P)` | `1`  | `[p,x]`            |
+| 2D        | `Grad(P)` | `2`  | `[p,x, p,y]`       |
+| 3D        | `Grad(P)` | `3`  | `[p,x, p,y, p,z]`  |
+| 1D/2D/3D  | `Id(P)`   | `1`  | `[p]`              |
+
+`Div(P)`, `Curl(P)` and `SymGrad(P)` are not defined for scalar fields.
+
+---
+
+## Vector fields
+
+Let
+
+```math
+u =
+\begin{bmatrix}
+u_x \\
+u_y \\
+u_z
+\end{bmatrix}
+```
+
+be a vector field. In 2D, only `ux` and `uy` are present.
+
+---
+
+## `Grad(Pu)` for vector fields
+
+`Grad(Pu)` returns the full displacement gradient in component-major ordering.
+
+### 2D vector field
+
+```math
+u =
+\begin{bmatrix}
+u_x \\
+u_y
+\end{bmatrix}
+```
+
+| Operator   | Size | Component ordering         |
+| ---------- | ----:| -------------------------- |
+| `Grad(Pu)` | `4`  | `[ux,x, ux,y, uy,x, uy,y]` |
+
+Equivalent matrix form:
+
+```math
+\nabla u =
+\begin{bmatrix}
+u_{x,x} & u_{x,y} \\
+u_{y,x} & u_{y,y}
+\end{bmatrix}
+```
+
+Flattened as:
+
+```math
+[u_{x,x}, u_{x,y}, u_{y,x}, u_{y,y}]
+```
+
+### 3D vector field
+
+| Operator   | Size | Component ordering                                       |
+| ---------- | ----:| -------------------------------------------------------- |
+| `Grad(Pu)` | `9`  | `[ux,x, ux,y, ux,z, uy,x, uy,y, uy,z, uz,x, uz,y, uz,z]` |
+
+Equivalent matrix form:
+
+```math
+\nabla u =
+\begin{bmatrix}
+u_{x,x} & u_{x,y} & u_{x,z} \\
+u_{y,x} & u_{y,y} & u_{y,z} \\
+u_{z,x} & u_{z,y} & u_{z,z}
+\end{bmatrix}
+```
+
+Flattened as:
+
+```math
+[u_{x,x}, u_{x,y}, u_{x,z},
+ u_{y,x}, u_{y,y}, u_{y,z},
+ u_{z,x}, u_{z,y}, u_{z,z}]
+```
+
+---
+
+## `SymGrad(Pu)` for vector fields
+
+`SymGrad(Pu)` returns the engineering strain vector.
+
+### 2D vector field
+
+| Operator      | Size | Component ordering |
+| ------------- | ----:| ------------------ |
+| `SymGrad(Pu)` | `3`  | `[εxx, εyy, γxy]`  |
+
+Explicitly:
+
+```math
+\operatorname{SymGrad}(u) =
+\begin{bmatrix}
+u_{x,x} \\
+u_{y,y} \\
+u_{x,y} + u_{y,x}
+\end{bmatrix}
+```
+
+### 3D vector field
+
[…]
