### 2026-05-26 621719e
mergeFields for elementwise fields
