### 2026-05-26 585fd41
Embedded Surface Operators and Membrane Mechanics
