+++ src/multifield.orig.jl
@@ -61,9 +61,6 @@ export TensorDiv
 export Adv
 export AxialGrad
 export TangentialGrad
-export SurfaceGrad
-export SurfaceDiv
-export SurfaceSymGrad
 
 export ∫
 export ∫Ω
@@ -267,8 +264,6 @@ struct AxialGradOp <: AbstractOp end
 
 #AxialGrad(P) = AxialGradOp()
 
-struct TangentialGradOp <: AbstractOp end
-
 """
     TangentialGrad(P)
 
@@ -290,19 +285,6 @@ For vector fields `u ∈ ℝᵈ`, the result is a vector field of dimension `d`.
 # Returns
 - `OpApplied`: Operator application object representing `(∇u) ⋅ t`.
 
-Directional derivative operator along the local element tangent direction.
-
-Computes derivatives projected onto the local tangent vector of the element.
-
-Typical applications include:
-
-- beam and rod formulations,
-- directional gradients,
-- streamline-like operators,
-- embedded directional mechanics.
-
-See also: `AxialGrad`
-
 # Example
 ```julia
 Kg = ∫(TangentialGrad(Pu) ⋅ N0 ⋅ TangentialGrad(Pu); Ω="truss")
@@ -316,30 +298,9 @@ Kg = ∫(TangentialGrad(Pu) ⋅ N0 ⋅ TangentialGrad(Pu); Ω="truss")
 - This operator is useful for constructing geometric stiffness matrices  
   (initial stress effects) in truss and structural stability problems.  
 """
+struct TangentialGradOp <: AbstractOp end
+
 TangentialGrad(P) = OpApplied(P, TangentialGradOp())
-
-"""
-Surface gradient operator.
-
-Computes tangential derivatives on embedded lower-dimensional entities,
-typically a 2D surface embedded in 3D.
-"""
-struct SurfaceGradOp <: AbstractOp end
-
-"""
-Surface divergence operator.
-
-Uses tangential derivatives on embedded lower-dimensional entities.
-"""
-struct SurfaceDivOp <: AbstractOp end
-
-"""
-Surface symmetric gradient operator.
-
-This is the membrane strain operator for a displacement field living on a
-surface embedded in the ambient space.
-"""
-struct SurfaceSymGradOp <: AbstractOp end
 
 """
     op_outdim(::IdOp, P::Problem)
@@ -417,22 +378,6 @@ end
 
 function op_outdim(::TangentialGradOp, P::Problem)
     return P.dim
-end
-
-function op_outdim(::SurfaceGradOp, P::Problem)
-    return op_outdim(GradOp(), P)
-end
-
-function op_outdim(::SurfaceDivOp, P::Problem)
-    return op_outdim(DivOp(), P)
-end
-
-#function op_outdim(::SurfaceSymGradOp, P::Problem)
-#    return op_outdim(SymGradOp(), P)
-#end
-function op_outdim(::SurfaceSymGradOp, P::Problem)
-    @assert P.pdim == 3 "SurfaceSymGradOp currently expects a 3D displacement field."
-    return 3
 end
 
 """
@@ -749,83 +694,6 @@ function build_B!(B::AbstractMatrix, ::TangentialGradOp,
     return B
 end
 
-function build_B!(B::AbstractMatrix, ::SurfaceGradOp,
-    P::Problem, k::Int, h, ∂h, numNodes::Integer)
-
-    return build_B!(B, GradOp(), P, k, h, ∂h, numNodes)
-end
-
-function build_B!(B::AbstractMatrix, ::SurfaceDivOp,
-    P::Problem, k::Int, h, ∂h, numNodes::Integer)
-
[…]
