### 2026-05-27 c53a6de
bug fix in assemble_operator
+ Timoshenko beam works :-) with DSL
