### 2026-05-28 de70ec9
better error message in assemble_operator
