### 2026-05-29 291cbf9
full and reduced integration
