### 2026-06-03 89b8cd2
bug fix in ScalarField constructor
Fixed a node tag vs. node index inconsistency in elementwise ScalarField construction. The issue appeared only on renumbered meshes (e.g. bandwidth=:Metis) and could lead to incorrect field initialization from spatial functions.
