### 2026-06-07 5573117
Add contributing guidelines and issue templates
