+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,32 @@
+---
+name: Feature request
+about: Suggest an improvement or new feature for LowLevelFEM.jl
+title: "[Feature]: "
+labels: enhancement
+assignees: ""
+---
+
+## Motivation
+
+What problem would this feature solve?
+
+## Proposed solution
+
+Describe the functionality you would like to see.
+
+## Example use case
+
+If possible, provide a short example of the intended workflow.
+
+```julia
+using LowLevelFEM
+# possible usage
+```
+
+## Alternatives considered
+
+Have you considered any workaround or alternative approach?
+
+## Additional context
+
+Add any other relevant information here.
