+++ CONTRIBUTING.md
@@ -1,0 +1,44 @@
+# Contributing to LowLevelFEM.jl
+
+Thank you for your interest in contributing to LowLevelFEM.jl.
+
+LowLevelFEM.jl is an open-source finite element toolbox written in Julia with a focus on education, research, and rapid prototyping of finite element formulations.
+
+## Reporting Issues
+
+If you encounter a bug, please open an issue on GitHub and include:
+
+* Julia version
+* LowLevelFEM version
+* Operating system
+* A minimal reproducible example
+* Full error message and stack trace
+
+## Feature Requests
+
+Suggestions for new functionality are welcome. When proposing a feature, please describe:
+
+* the engineering or scientific problem it addresses,
+* the expected workflow,
+* and, if possible, a simple example.
+
+## Pull Requests
+
+Pull requests are welcome.
+
+Please try to:
+
+* follow the existing coding style,
+* include documentation for new public functions,
+* add tests whenever practical,
+* ensure that existing tests continue to pass.
+
+## Documentation
+
+Improvements to documentation, tutorials, examples, and notebooks are highly appreciated.
+
+## Questions and Discussion
+
+For questions, ideas, or discussion about future development, please use GitHub Issues or Discussions.
+
+Thank you for helping improve LowLevelFEM.jl.
