+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,46 @@
+---
+name: Bug report
+about: Report a reproducible problem in LowLevelFEM.jl
+title: "[Bug]: "
+labels: bug
+assignees: ""
+
+---
+
+## Description
+
+Please describe the problem clearly.
+
+## Minimal reproducible example
+
+Please provide a small example that reproduces the issue.
+
+```julia
+using LowLevelFEM
+# your code here
+```
+
+## Expected behavior
+
+What did you expect to happen?
+
+## Actual behavior
+
+What happened instead?
+
+## Error message / stack trace
+
+```text
+Paste the full error message here.
+```
+
+## Environment
+
+* LowLevelFEM.jl version:
+* Julia version:
+* Operating system:
+* Gmsh version, if relevant:
+
+## Additional context
+
+Add any other relevant information here.
