### 2026-06-07 10ca7eb
Expand automated test suite
