### 2026-06-07 perebalazs/LowLevelFEM.jl#35 @perebalazs
Thank you again for the suggestion.

I have now expanded the automated test suite beyond the original single end-to-end simulation.

The current test suite includes:

- field algebra tests,
- differential operator tests,
- matrix assembly tests,
- analytical validation tests against known elasticity solutions,
- and end-to-end solver workflow tests.

The number of automated tests has increased from the original single simulation-based test to multiple test sets covering different layers of the package. This provides much finer resolution for detecting regressions and identifying which subsystem is affected by a change.

The test coverage is still being improved, but I believe the current structure addresses the main concern raised in this review comment.

Thank you for encouraging this improvement.
