### 2026-06-07 05a6580
New example + move old examples to 'old_examples'
