### 2026-06-07 37f78bd
New example + corrected README.md
