### 2026-06-07 1ec5f68
Revise SECURITY.md to simplify content
Removed supported versions and reporting guidelines sections.
