+++ SECURITY.md
@@ -1,0 +1,7 @@
+# Security Policy
+
+LowLevelFEM is a research and educational package.
+
+If you discover a security-related issue, please open a GitHub issue or contact the maintainer directly.
+
+Maintainer: Balázs Pere
