### 2026-06-07 8104d30
Update issue templates
