+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -3,7 +3,8 @@ name: Feature request
 about: Suggest an improvement or new feature for LowLevelFEM.jl
 title: "[Feature]: "
 labels: enhancement
-assignees: ""
+assignees: ''
+
 ---
 
 ## Motivation
