+++ .github/ISSUE_TEMPLATE/pull-request-template.md
@@ -1,0 +1,24 @@
+---
+name: Pull request template
+about: Describe the changes introduced by this pull request.
+title: ''
+labels: ''
+assignees: perebalazs
+
+---
+
+## Description
+
+Describe the changes introduced by this pull request.
+
+## Type of change
+
+- [ ] Bug fix
+- [ ] New feature
+- [ ] Documentation update
+- [ ] Refactoring
+
+## Checklist
+
+- [ ] Tests pass
+- [ ] Documentation updated if needed
