### 2026-06-22 f03b170
compound operator
