### 2026-06-23 6ac8d4d
compound operator in DSL
