+++ src/multifield-old.jl
@@ -1074,257 +1074,36 @@ function _eval_coefficient_at_gp(Cprep, elem, hgp)
     error("Unsupported prepared coefficient type $(typeof(Cprep))")
 end
 
-
-# ---------------------------------------------------------------------------
-# Small local matrix multiplication with optional structural sparsity
-# ---------------------------------------------------------------------------
-#
-# This layer is intentionally hidden below the weak-form DSL.  The user can keep
-# writing paper-like expressions such as
-#
-#     ∫( SymGrad(Pu) ⋅ C ⋅ SymGrad(Pu) )
-#
-# while the local Gauss-point products try to avoid multiplications by known
-# zero entries.  If a factor is not sparse enough, the code falls back to dense
-# BLAS-backed mul!.
-
-struct MatrixPattern
-    rows::Vector{Int}
-    cols::Vector{Int}
-    nnz::Int
-    m::Int
-    n::Int
-end
-
-MatrixPattern(m::Integer, n::Integer) = MatrixPattern(Int[], Int[], 0, Int(m), Int(n))
-
-@inline density(p::MatrixPattern) = p.nnz / (p.m * p.n)
-
-"""
-    detect_pattern(M; tol=0.0)
-
-Detect the nonzero pattern of a small local matrix.
-
-The default `tol=0.0` is deliberately strict: entries that are exactly zero are
-treated as structural zeros, but accidentally small nonzero entries are kept.
-For finite element operator matrices this is usually the safe choice.
-"""
-function detect_pattern(M::AbstractMatrix; tol::Real=0.0)
-    m, n = size(M)
-    rows = Int[]
-    cols = Int[]
-    sizehint!(rows, length(M))
-    sizehint!(cols, length(M))
-
-    @inbounds for j in 1:n
-        for i in 1:m
-            mij = M[i, j]
-            if tol == 0
-                if !iszero(mij)
-                    push!(rows, i)
-                    push!(cols, j)
-                end
-            else
-                if abs(mij) > tol
-                    push!(rows, i)
-                    push!(cols, j)
-                end
-            end
-        end
-    end
-
-    return MatrixPattern(rows, cols, length(rows), m, n)
-end
-
-@inline function sparse_enough(
-    p::MatrixPattern;
-    density_limit::Real = 0.40,
-    min_saved_entries::Integer = 2
-)
-    total = p.m * p.n
-    return (total - p.nnz) >= min_saved_entries && density(p) <= density_limit
-end
-
-@inline function _scale_or_zero!(R, beta)
-    if beta == 0
-        fill!(R, 0.0)
-    elseif beta != 1
-        R .*= beta
-    end
-    return R
-end
-
-"""
-    mul_dense_sparse!(R, A, B, patB, alpha=1.0, beta=0.0)
-
-Compute `R = alpha*A*B + beta*R`, using only the nonzero entries of `B`.
-"""
-function mul_dense_sparse!(
-    R::AbstractMatrix,
-    A::AbstractMatrix,
-    B::AbstractMatrix,
-    patB::MatrixPattern,
-    alpha::Number = 1.0,
-    beta::Number = 0.0
-)
-    _scale_or_zero!(R, beta)
-
-    @inbounds for q in 1:patB.nnz
-        k = patB.rows[q]
-        j = patB.cols[q]
-        bkj = B[k, j]
-        iszero(bkj) && continue
-
-        αb = alpha * bkj
-        for i in axes(A, 1)
-            R[i, j] += A[i, k] * αb
-        end
-    end
-
-    return R
-end
-
-"""
-    mul_sparse_dense!(R, A, B, patA, alpha=1.0, beta=0.0)
-
[…]
