### 2026-06-25 1cf6ccc
compound operator - final
