+++ src/multifield.jl
@@ -41,6 +41,17 @@ Matrix chain coefficients
 where `A,B,C` may be matrices containing `Number` or `ScalarField`
 entries.
 
+Compound operator sums
+
+    B = A ⋅ Grad(Pu) + G ⋅ Pu
+    K = ∫(B' ⋅ D ⋅ B; Ω="body", weight=2π*r)
+or
+    K = ∫(B' ⋅ D ⋅ B * (2π*r); Ω="body")
+
+The sum is expanded into ordinary bilinear terms and assembled using the
+standard bilinear assembler. Each term may carry its own quadrature rule via
+`full(...)` and `reduced(...)`.
+
 The implementation is designed to be
 
 - transparent
@@ -64,6 +75,7 @@ export TangentialGrad
 export SurfaceGrad
 export SurfaceDiv
 export SurfaceSymGrad
+export full, reduced
 
 export ∫
 export ∫Ω
@@ -1422,7 +1434,15 @@ which is evaluated at Gauss points as
 
 `weight`
 
-Optional weighting tensor applied to the operator value. (Deprecated: kept for backward compatibility.)
+Optional scalar quadrature weight multiplying the integrand at each Gauss
+point. It may be a `Number` or a `ScalarField`.
+
+This is useful for geometry-related weights such as the axisymmetric factor
+`2πr`:
+
+    ∫(B' ⋅ D ⋅ B; Ω="body", weight=2π*r)
+
+The weight is not treated as a matrix coefficient.
 
 `domain`
 
@@ -2647,7 +2667,19 @@ end
 # ------------------------------------------------------------
 # Bilinear term
 # ------------------------------------------------------------
+"""
+    BilinearTerm(a, coef, b)
 
+Internal representation of a bilinear weak-form term.
+
+Represents
+
+    ∫ (a(v))' * coef * b(u)
+
+where `a` is the test-side operator and `b` is the trial-side operator.
+`coef` may be a scalar, a matrix, or a matrix chain containing `Number`
+and `ScalarField` entries.
+"""
 struct BilinearTerm
     a::OpApplied
     coef
@@ -2960,7 +2992,7 @@ end
 # Expression building
 # ------------------------------------------------------------
 
-import Base: +, -, *
+import Base: +, -, *, adjoint, transpose
 
 +(a::WeakExpr, b::WeakExpr) = WeakSum(a, b)
 
@@ -3336,6 +3368,334 @@ KΓ = ∫Γ("loaded_boundary", Id(Pu) ⋅ Id(Pu))
 ∫Γ(name, expr) = ∫(expr; Γ=name)
 
 const ε = SymGrad
+
+###############################################################
+# Compound operators ##########################################
+###############################################################
+
+"""
+    ChainSumTerm(chain, gauss)
+
+Internal storage for one term of a compound operator expression.
+
+`chain` is usually an `OpApplied` or a `MatrixChain`.
+`gauss` stores the preferred integration rule for this term.
+"""
+struct ChainSumTerm
+    chain::Any
+    gauss::Any
+end
+
+"""
+    ChainSum(terms)
+
+Symbolic sum of operator chains.
+
+Example:
+
+```julia
+B = Grad(Pu) ⋅ A + Pu ⋅ G
+```
+
+stores two terms without assembling anything.
+"""
+struct ChainSum
+    terms::Vector{ChainSumTerm}
+end
+
+"""
+TransposedChain(chain)
+
+Symbolic transpose marker for compound expressions.
+
+This does not assemble or numerically transpose anything immediately.
+It is interpreted during expansion of the bilinear form.
+"""
[…]
