### 2026-06-25 ff28350
weight for linear term
