### 2026-06-29 3f3c4db
new examples + bug fix
