### 2026-06-30 c41a38e
new examples
