+++ docs/src/tutorials/index.md
@@ -5,9 +5,10 @@ Tutorials provide scenario-driven workflows for common LowLevelFEM tasks.
 ## Available
 
 - [Linear Elasticity (2D)](linear-elasticity-2d.md)
-- [Heat Conduction (Transient)](heat-conduction-transient.md)
-- [Modal and Buckling Analysis](modal-and-buckling.md)
-- [Nonlinear Large Deformation](nonlinear-large-deformation.md)
-- [Multifield Weak-Form DSL](multifield-weak-form-dsl.md)
-- [Poisson and Custom Operators](poisson-and-custom-operators.md)
+- [Linear Elasticity (2D axisymmetric)](linear-elasticity-2d-axi.md)
+- [Transient Elasticity (2D)](transient-elasticity-2d.md)
+- [Heat Conduction](heat-conduction.md)
+- [Multifield Weak-Form DSL - Navier-Stokes](multifield-weak-form-dsl.md)
+- [Displacement Recovery from Strain](strain-displacement-relation.md)
+- [Torsion of a prismatic bar using the Prandtl stress function](prandtl-stress-function.md)
 - [Legacy Example Gallery](legacy-examples.md)
