### 2026-06-30 affc4d1
new examples
