+++ docs/src/tutorials/multifield-weak-form-dsl.md
@@ -1,11 +1,12 @@
-# Multifield Weak-Form DSL
+# Incompressible Navier–Stokes Flow
 
-This tutorial will show the recommended modern assembly workflow using
-`Grad`, `Div`, `SymGrad`, `Id`, `TensorDiv`, `Adv`, and `∫( ... )`.
+This tutorial demonstrates the numerical solution of the incompressible Navier–Stokes equations in a two-dimensional domain. The example illustrates how viscous fluid flow can be simulated using the finite element method, including the coupling between velocity and pressure fields.
 
-## Status
+The tutorial covers the complete workflow, including mesh generation, material properties, boundary conditions, solving the governing equations, and visualization of the resulting velocity and pressure fields. 
 
-Placeholder page.
+## Example
+
+[Navier-Stokes.ipynb](https://github.com/perebalazs/LowLevelFEM.jl/blob/main/examples/LinkedIn/Navier-Stokes.ipynb)
 
 ## Related
 
