+++ docs/src/tutorials/transient-elasticity-2d.md
@@ -1,11 +1,14 @@
-# Linear Elasticity (2D)
+# Transient Linear Elasticity (2D)
 
-This tutorial will contain a full 2D cantilever workflow with mesh setup,
-material definition, constraints, loading, solving, and stress postprocessing.
 
-## Status
 
-Placeholder page. Detailed step-by-step content can be migrated from existing example notebooks.
+This tutorial demonstrates the transient analysis of a two-dimensional linear elastic solid. In contrast to static analysis, the inertia of the structure is taken into account, allowing the simulation of time-dependent phenomena such as vibrations, wave propagation, and dynamic loading.
+
+The example introduces the complete workflow, including mesh generation, material definition, boundary conditions, time-dependent loading, time integration, and post-processing of displacement and stress fields. 
+
+## Example
+
+[lin_elast_2D_transient.ipynb](https://github.com/perebalazs/LowLevelFEM.jl/blob/main/examples/lin_elast_2D_transient.ipynb)
 
 ## Related
 
