### 2026-06-30 29fa6fb
update paper.md
