+++ paper/paper.md
@@ -19,11 +19,11 @@ bibliography: paper.bib
 
 # Summary
 
-LowLevelFEM.jl is a finite element method (FEM) [@zienkiewicz2005] toolbox written entirely in the Julia programming language [@bezanson2017]. Its design philosophy emphasizes **simplicity, transparency, and performance**, making it suitable for both educational and research purposes in mechanics and engineering. Unlike large frameworks that rely on domain-specific languages or compiled backends, LowLevelFEM provides users with direct access to FEM building blocks in Julia, enabling full control over discretization, assembly, and solution steps.
+LowLevelFEM.jl is a finite element method (FEM) [@zienkiewicz2005] toolbox written entirely in the Julia programming language [@bezanson2017]. Its design philosophy emphasizes **simplicity, transparency, and operator-based weak-form prototyping**, making it suitable for both educational and research purposes in mechanics and engineering. Unlike frameworks that primarily rely on high-level domain-specific languages, LowLevelFEM provides direct access to the underlying FEM building blocks in Julia, enabling full control over discretization, assembly, and solution steps, while also supporting an operator-based DSL for expressing custom weak-form formulations.
 
 The package currently supports two- and three-dimensional solid mechanics problems including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimalistic design lowers the entry barrier for students while still offering enough flexibility for advanced users to **inspect and control algorithms step by step, process intermediate results during the solution procedure, and perform operations with scalar, vector, and tensor fields**. LowLevelFEM uses **Gmsh** [@geuzaine2009] as its pre- and post-processor, ensuring compatibility with a widely adopted meshing and visualization tool. The toolbox also supports geometrically nonlinear formulations based on a Total Lagrangian framework and energy-driven constitutive modeling, enabling research-level investigations beyond linear elasticity.
 
-Thanks to Julia’s just-in-time compilation and multiple dispatch, the code remains concise while achieving performance comparable to traditional FEM codes in C/C++ or Fortran. LowLevelFEM is released under the MIT license and distributed via the Julia General registry. Documentation, tutorials, and examples are available at [https://juliahub.com/ui/Packages/General/LowLevelFEM](https://juliahub.com/ui/Packages/General/LowLevelFEM) or [https://perebalazs.github.io/LowLevelFEM.jl/stable/](https://perebalazs.github.io/LowLevelFEM.jl/stable/).
+Thanks to Julia’s just-in-time compilation and multiple dispatch, the implementation remains concise while providing good computational performance for educational and research applications. LowLevelFEM is released under the MIT license and distributed via the Julia General registry. Documentation, including a growing collection of executable Jupyter notebook tutorials, is available at [https://juliahub.com/ui/Packages/General/LowLevelFEM](https://juliahub.com/ui/Packages/General/LowLevelFEM) or [https://perebalazs.github.io/LowLevelFEM.jl/stable/](https://perebalazs.github.io/LowLevelFEM.jl/stable/).
 
 ## Example
 
@@ -32,12 +32,10 @@ Below is a simple example illustrating a typical LowLevelFEM workflow using Gmsh
 ```julia
 using LowLevelFEM
 
-# `gmsh` is exported by LowLevelFEM
-gmsh.initialize()
-gmsh.open("model.geo")
+openGeometry("model.geo")
 
 mat  = Material("body", E=2e5, ν=0.3)
-prob = Problem([mat], type=:PlaneStress)  # :Solid, :PlaneStrain, :AxiSymmetric, :HeatConduction, ...
[…]
