### 2026-07-06 6f6fb3f
integration of elementwise scalar field
