+++ docs/src/tutorials/transient-elasticity-2d.md
@@ -1,7 +1,5 @@
 # Transient Linear Elasticity (2D)
 
-
-
 This tutorial demonstrates the transient analysis of a two-dimensional linear elastic solid. In contrast to static analysis, the inertia of the structure is taken into account, allowing the simulation of time-dependent phenomena such as vibrations, wave propagation, and dynamic loading.
 
 The example introduces the complete workflow, including mesh generation, material definition, boundary conditions, time-dependent loading, time integration, and post-processing of displacement and stress fields. 
