### 2026-07-07 5a2becf
improve nodestoElements function
