+++ docs/src/manual/postprocessing-and-visualization.md
@@ -1,10 +1,29 @@
 # Postprocessing and Visualization
 
-This section will cover `show*Results`, path plotting, probing, and field projection workflows.
+LowLevelFEM provides built-in postprocessing and visualization tools based on Gmsh. Results can be displayed interactively, exported, probed at arbitrary locations, or evaluated along user-defined paths.
 
-## Status
+Typical workflow:
 
-Placeholder page.
+```julia
+u = solveDisplacement(problem, load=[force], support=[bc])
+S = solveStress(u)
+
+showDoFResults(u)
+showStressResults(S)
+
+openPostProcessor()
+```
+
+The most commonly used visualization functions are:
+
+- `showDoFResults` – displacement and other nodal degrees of freedom
+- `showStressResults` – stress components and equivalent stress
+- `showStrainResults` – strain components
+- `showHeatFluxResults` – heat flux fields
+- `plotOnPath` – evaluate and plot results along a user-defined path
+- `probe` – evaluate field values at arbitrary locations
+
+For complete function documentation, see the API reference below.
 
 ## Related API
 
