+++ docs/src/manual/troubleshooting.md
@@ -1,12 +1,18 @@
 # Troubleshooting
 
-Common issues to cover:
+This page collects solutions to the most common problems encountered when using LowLevelFEM.
 
-- Missing or mismatched physical group names
-- Incompatible problem dimensions and field dimensions
-- Boundary/load key mismatches
-- Gmsh initialization/finalization lifecycle errors
+Typical issues include:
 
-## Status
+- missing or mismatched physical group names
+- incompatible problem types and field dimensions
+- incorrect boundary-condition or load definitions
+- Gmsh initialization or finalization errors
 
-Placeholder page.
+If a simulation does not behave as expected, first verify that:
+
+- all referenced physical groups exist in the Gmsh model,
+- the selected problem type matches the mesh and material definition,
+- boundary conditions and loads are applied to the intended physical groups.
+
+This section will be expanded as additional common issues are identified.
