+++ README.md
@@ -14,7 +14,7 @@ LowLevelFEM is a Julia package for finite element analysis with an engineering-f
 designed to assemble finite element operators explicitly at matrix and field level.  
 It exposes each phase of the workflow as simple functions (mesh → matrices → loads/BCs → solve → postprocess → visualize), so you can customize, combine, or inspect any step **at matrix and field level**.  
 Typical tasks such as strain energy or resultants are one-liners (for example, `U = q' * K * q / 2`).
-The package is suitable not only for classical structural mechanics problems, but also for assembling general linear PDEs expressed in weak form.
+The package is suitable not only for classical structural mechanics problems, but also for rapid assembly of general linear PDEs expressed in weak form.
 
 ## Requirements
 
@@ -30,7 +30,8 @@ The package is suitable not only for classical structural mechanics problems, bu
 - Matrices: stiffness `K`, mass `M` (lumped or consistent), proportional damping `C` (Rayleigh/Caughey),
   heat conduction/capacity, latent heat, convection matrices/vectors, and generic Poisson-type operators
   for scalar and vector fields.
-- **Matrix- and field-oriented FEM workflow: global stiffness, mass, damping and coupling matrices, together with scalar, vector and tensor fields, are explicit and user-accessible objects, enabling direct implementation, inspection and modification of PDE operators (including Poisson-type operators) and constitutive laws at matrix and field level.**
+- **Explicit matrices and fields:** unlike many high-level FEM packages, LowLevelFEM keeps global matrices and finite element fields accessible throughout the workflow, making every stage of the analysis transparent and customizable.
+- **Operator-level programming:** implement, inspect and modify PDE operators, constitutive laws and weak-form formulations directly at matrix and field level.
 - Loads and constraints: nodal and distributed loads on physical groups; function-based loads and temperature BCs; elastic supports; initial displacement/velocity/temperature.
 - Thermal–structural coupling: thermal expansion, thermal stresses, and heat generated by elastic deformation.
 - Solvers: static (direct and iterative solvers with arbitrary preconditioners) and transient dynamics (central difference and HHT-α from the Newmark family).
@@ -53,29 +54,26 @@ Pkg.add("LowLevelFEM")
 ```julia
 using LowLevelFEM
 
-# `gmsh` is exported by LowLevelFEM
-gmsh.initialize()
-gmsh.open("your_model.geo")
+structured_rect_mesh() # uses Gmsh to create mesh
 
 mat = Material("body", E=2e5, ν=0.3)
 prob = Problem([mat], type=:PlaneStrain)  # :Solid, :PlaneStress, :AxiSymmetric, :HeatConduction, ...
 
-bc    = displacementConstraint("supp", ux=0, uy=0)
-force = load("load", fy=-1)
+bc    = displacementConstraint("left", ux=0, uy=0)
+force = load("right", fy=-1)
 
 u = solveDisplacement(prob, load=[force], support=[bc])
 S = solveStress(u)
 
-showDoFResults(u)
+showDoFResults(u, visible=true)
 showDoFResults(u, :ux)
 showStressResults(S)
 showStressResults(S, :sxy, name="Shear stress")
 
 openPostProcessor()
-gmsh.finalize()
 ```
 
-Note: physical group names in your geometry (created in Gmsh) must match the strings used above (e.g., `"body"`, `"supp"`, `"load"`).
+Note: physical group names in your geometry (created in Gmsh) must match the strings used above (e.g., `"body"`, `"left"`, `"right"`).
 
 ### An alternative solution (instead of `u = ...`, `S = ...`)
 
[…]
