### 2026-07-17 2115b44
bug fix and power function for ScalarField
