+++ paper/paper.md
@@ -19,9 +19,9 @@ bibliography: paper.bib
 
 # Summary
 
-LowLevelFEM.jl is a finite element method (FEM) [@zienkiewicz2005] toolbox written entirely in the Julia programming language [@bezanson2017]. Its design philosophy emphasizes **simplicity, transparency, and operator-based weak-form prototyping**, making it suitable for both educational and research purposes in mechanics and engineering. Unlike frameworks that primarily rely on high-level domain-specific languages, LowLevelFEM provides direct access to the underlying FEM building blocks in Julia, enabling full control over discretization, assembly, and solution steps, while also supporting an operator-based DSL for expressing custom weak-form formulations.
+LowLevelFEM.jl is a finite element method (FEM) [@zienkiewicz2005] toolbox written entirely in the Julia programming language [@bezanson2017]. Its design philosophy emphasizes **simplicity, transparency, and operator-based weak-form prototyping**, making it suitable for both educational and research purposes in mechanics and engineering. Unlike frameworks that primarily rely on high-level domain-specific languages (DSLs), LowLevelFEM provides direct access to the underlying FEM building blocks in Julia, enabling full control over discretization, assembly, and solution steps, while also supporting an operator-based DSL for expressing custom weak-form formulations.
 
-The package currently supports two- and three-dimensional solid mechanics problems including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimalistic design lowers the entry barrier for students while still offering enough flexibility for advanced users to **inspect and control algorithms step by step, process intermediate results during the solution procedure, and perform operations with scalar, vector, and tensor fields**. LowLevelFEM uses **Gmsh** [@geuzaine2009] as its pre- and post-processor, ensuring compatibility with a widely adopted meshing and visualization tool. The toolbox also supports geometrically nonlinear formulations based on a Total Lagrangian framework and energy-driven constitutive modeling, enabling research-level investigations beyond linear elasticity.
+The package currently supports two- and three-dimensional solid mechanics problems including plane stress, plane strain, axisymmetric, and 3D solid analyses. Its minimal design lowers the entry barrier for students while still offering enough flexibility for advanced users to **inspect and control algorithms step by step, process intermediate results during the solution procedure, and perform operations on scalar, vector, and tensor fields**. LowLevelFEM uses **Gmsh** [@geuzaine2009] as its pre- and post-processor, ensuring compatibility with a widely adopted meshing and visualization tool. The toolbox also supports geometrically nonlinear formulations based on a Total Lagrangian framework and energy-based constitutive modeling, enabling research-level investigations beyond linear elasticity.
 
 Thanks to Julia’s just-in-time compilation and multiple dispatch, the implementation remains concise while providing good computational performance for educational and research applications. LowLevelFEM is released under the MIT license and distributed via the Julia General registry. Documentation, including a growing collection of executable Jupyter notebook tutorials, is available at [https://juliahub.com/ui/Packages/General/LowLevelFEM](https://juliahub.com/ui/Packages/General/LowLevelFEM) or [https://perebalazs.github.io/LowLevelFEM.jl/stable/](https://perebalazs.github.io/LowLevelFEM.jl/stable/).
 
@@ -86,7 +86,7 @@ LowLevelFEM addresses these challenges by offering a **lightweight Julia-only im
 * Research projects where step-by-step inspection of the solution process and manipulation of intermediate fields is required.
 * Development, testing, and validation of novel finite element formulations.
 
[…]
