+++ paper/paper.bib
@@ -1,9 +1,9 @@
-@article{Bezanson2017,
+@article{bezanson2017,
     volume = {59},
     publisher = {SIAM},
     doi = {10.1137/141000671},
     year = {2017},
-    title = {Julia: A fresh approach to numerical computing},
+    title = {{Julia}: A fresh approach to numerical computing},
     pages = {65--98},
     author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
     journal = {SIAM Review},
@@ -14,11 +14,11 @@
     year = {2015},
     volume = {3},
     journal = {Archive of Numerical Software},
-    title = {The FEniCS Project Version 1.5},
+    title = {The {FEniCS} Project Version 1.5},
     pages = {9--23},
     number = {100},
     doi = {10.11588/ans.2015.100.20553},
-    author = {Aln{e}s, Martin and Blechta, Jan and Hake, Johan and Johansson, Anders and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and others},
+    author = {Aln{\ae}s, Martin and Blechta, Jan and Hake, Johan and Johansson, Anders and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and others},
 }
 
 @article{Bangerth2007,
@@ -30,18 +30,7 @@
     doi = {10.1145/1268776.1268779},
     publisher = {ACM New York, NY, USA},
     journal = {ACM Transactions on Mathematical Software (TOMS)},
-    title = {deal.II--A general-purpose object-oriented finite element library},
-}
-
-@article{bezanson2017,
-    title = {Julia: A Fresh Approach to Numerical Computing},
-    doi = {10.1137/141000671},
-    volume = {59},
-    year = {2017},
-    number = {1},
-    author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
-    pages = {65--98},
-    journal = {SIAM Review},
+    title = {{deal.II}--A general-purpose object-oriented finite element library},
 }
 
 @article{geuzaine2009,
@@ -52,11 +41,11 @@
     journal = {International Journal for Numerical Methods in Engineering},
     year = {2009},
     doi = {10.1002/nme.2579},
-    title = {Gmsh: A three-dimensional finite element mesh generator with built-in pre- and post-processing facilities},
+    title = {{Gmsh}: A three-dimensional finite element mesh generator with built-in pre- and post-processing facilities},
 }
 
 @article{badia2020,
-    title = {Gridap: An extensible Finite Element toolbox in Julia},
+    title = {{Gridap}: An Extensible Finite Element Toolbox in {Julia}},
     author = {Badia, Santiago and Martín, Antonio and Verdugo, Francesc},
     volume = {5},
     number = {52},
@@ -67,7 +56,7 @@
 }
 
 @article{knyazev2017,
-    title = {Arpack.jl: A Julia interface to ARPACK},
+    title = {{Arpack.jl}: A {Julia} Interface to {ARPACK}},
     year = {2017},
     journal = {Journal of Open Source Software},
     doi = {10.21105/joss.00142},
@@ -83,13 +72,14 @@
     publisher = {Butterworth–Heinemann},
     title = {The Finite Element Method},
     year = {2005},
+    isbn = {9780080472775},
 }
 
 @book{logg2012,
     author = {Logg, Anders and Mardal, Kent-Andre and Wells, Garth},
     year = {2012},
     publisher = {Springer},
-    title = {Automated Solution of Differential Equations by the Finite Element Method: The FEniCS Book},
[…]
