### 2026-07-24 094ac07
update paper.md
