+++ paper/paper.bib
@@ -93,3 +93,17 @@
     doi = {10.11588/ans.2015.100.20553},
     year = {2015},
 }
+
+@misc{ferrite2021,
+  author = {Carlsson, Kristoffer and Ekre, Fredrik and Contributors},
+  title = {Ferrite.jl},
+  year = {2021},
+  howpublished = {\url{https://github.com/Ferrite-FEM/Ferrite.jl}}
+}
+
+@misc{juafem2018,
+  author = {Carlsson, Kristoffer},
+  title = {JuAFEM.jl},
+  year = {2018},
+  howpublished = {\url{https://github.com/KristofferC/JuAFEM.jl}}
+}
