### 2026-07-25 1c59d1b
update paper.md and paper.bib
