+++ paper/paper.bib
@@ -1,109 +1,109 @@
 @article{bezanson2017,
-    volume = {59},
+    pages = {65--98},
     publisher = {SIAM},
+    author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
     doi = {10.1137/141000671},
+    number = {1},
     year = {2017},
     title = {{Julia}: A fresh approach to numerical computing},
-    pages = {65--98},
-    author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
     journal = {SIAM Review},
-    number = {1},
+    volume = {59},
 }
 
 @article{Alnaes2015,
+    pages = {9--23},
     year = {2015},
-    volume = {3},
     journal = {Archive of Numerical Software},
     title = {The {FEniCS} Project Version 1.5},
-    pages = {9--23},
     number = {100},
     doi = {10.11588/ans.2015.100.20553},
     author = {Aln{\ae}s, Martin and Blechta, Jan and Hake, Johan and Johansson, Anders and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and others},
+    volume = {3},
 }
 
 @article{Bangerth2007,
     author = {Bangerth, Wolfgang and Hartmann, Ralf and Kanschat, Guido},
-    number = {4},
-    year = {2007},
-    volume = {33},
     pages = {24},
-    doi = {10.1145/1268776.1268779},
     publisher = {ACM New York, NY, USA},
-    journal = {ACM Transactions on Mathematical Software (TOMS)},
+    year = {2007},
     title = {{deal.II}--A general-purpose object-oriented finite element library},
+    volume = {33},
+    doi = {10.1145/1268776.1268779},
+    journal = {ACM Transactions on Mathematical Software (TOMS)},
+    number = {4},
 }
 
 @article{geuzaine2009,
-    volume = {79},
     author = {Geuzaine, Christophe and Remacle, Jean-François},
-    pages = {1309--1331},
     number = {11},
-    journal = {International Journal for Numerical Methods in Engineering},
-    year = {2009},
     doi = {10.1002/nme.2579},
     title = {{Gmsh}: A three-dimensional finite element mesh generator with built-in pre- and post-processing facilities},
+    pages = {1309--1331},
+    journal = {International Journal for Numerical Methods in Engineering},
+    volume = {79},
+    year = {2009},
 }
 
 @article{badia2020,
-    title = {{Gridap}: An Extensible Finite Element Toolbox in {Julia}},
-    author = {Badia, Santiago and Martín, Antonio and Verdugo, Francesc},
     volume = {5},
-    number = {52},
-    journal = {Journal of Open Source Software},
     pages = {2520},
+    title = {{Gridap}: An Extensible Finite Element Toolbox in {Julia}},
     year = {2020},
+    journal = {Journal of Open Source Software},
+    number = {52},
     doi = {10.21105/joss.02520},
+    author = {Badia, Santiago and Martín, Antonio and Verdugo, Francesc},
 }
 
 @article{knyazev2017,
-    title = {{Arpack.jl}: A {Julia} Interface to {ARPACK}},
     year = {2017},
     journal = {Journal of Open Source Software},
     doi = {10.21105/joss.00142},
-    volume = {2},
-    number = {12},
     pages = {142},
+    number = {12},
+    title = {{Arpack.jl}: A {Julia} Interface to {ARPACK}},
+    volume = {2},
     author = {Knyazev, Andrew},
 }
 
 @book{zienkiewicz2005,
+    isbn = {9780080472775},
[…]
