+++ paper/paper.bib
@@ -21,7 +21,7 @@
     volume = {3},
 }
 
-@article{Bangerth2007,
+@article{bangerth2007,
     author = {Bangerth, Wolfgang and Hartmann, Ralf and Kanschat, Guido},
     pages = {24},
     publisher = {ACM New York, NY, USA},
@@ -55,15 +55,11 @@
     author = {Badia, Santiago and Martín, Antonio and Verdugo, Francesc},
 }
 
-@article{knyazev2017,
-    year = {2017},
-    journal = {Journal of Open Source Software},
-    doi = {10.21105/joss.00142},
-    pages = {142},
-    number = {12},
-    title = {{Arpack.jl}: A {Julia} Interface to {ARPACK}},
-    volume = {2},
-    author = {Knyazev, Andrew},
+@misc{arpackjl,
+  author = {JuliaLinearAlgebra contributors},
+  title = {Arpack.jl},
+  year = {2026},
+  howpublished = {\url{https://github.com/JuliaLinearAlgebra/Arpack.jl}}
 }
 
 @book{zienkiewicz2005,
@@ -81,17 +77,6 @@
     author = {Logg, Anders and Mardal, Kent-Andre and Wells, Garth},
     title = {Automated Solution of Differential Equations by the Finite Element Method: The {FEniCS} Book},
     doi = {10.1007/978-3-642-23099-8},
-}
-
-@article{bangerth2015,
-    journal = {Archive of Numerical Software},
-    pages = {1--8},
-    title = {The {deal.II} library, version 8.2},
-    year = {2015},
-    author = {Bangerth, Wolfgang and Heister, Timo and Heltai, Luca and Kanschat, Guido and Kronbichler, Martin and Maier, Matthias and Turcksin, Bruno and Young, David},
-    doi = {10.11588/ans.2015.100.20553},
-    number = {100},
-    volume = {3},
 }
 
 @misc{ferrite2021,
