### 2026-07-25 82f8e3c
Update paper.bib
