+++ paper/paper.bib
@@ -56,7 +56,7 @@
 }
 
 @misc{arpackjl,
-  author = {JuliaLinearAlgebra contributors},
+  author = {JuliaLinearAlgebra},
   title = {Arpack.jl},
   year = {2026},
   howpublished = {\url{https://github.com/JuliaLinearAlgebra/Arpack.jl}}
