### 2026-07-25 fd07b82
build based on 82f8e3c
