### 2026-07-25 e0e875f
Update paper.bib
