+++ paper/paper.bib
@@ -21,7 +21,7 @@
     volume = {3},
 }
 
-@article{bangerth2007,
+@article{bangerth2015,
     author = {Bangerth, Wolfgang and Hartmann, Ralf and Kanschat, Guido},
     pages = {24},
     publisher = {ACM New York, NY, USA},
