### 2026-07-27 aea8d13
optimization of compound operators
