### 2026-07-28 06fe7da
optimization of assemble_operator and assemble_linear
speed up is about 40% :-)
