+++ src/multifield.jl
@@ -1648,6 +1648,23 @@ function assemble_operator(
                 _, dfun, _ = gmsh.model.mesh.getBasisFunctions(et, intPoints, "GradLagrange")
                 ∇h = reshape(dfun, :, numIntPoints)
 
+                # Batched Jacobian retrieval
+                elemTags_global, _ = gmsh.model.mesh.getElementsByType(et)
+                jac_all, det_all, _ = gmsh.model.mesh.getJacobians(et, intPoints)
+
+                numElementsGlobal = length(elemTags_global)
+                nip = length(det_all) ÷ numElementsGlobal
+
+                @assert nip == numIntPoints
+
+                TagType = eltype(elemTags_global)
+                idxmap = Dict{TagType,Int}()
+                sizehint!(idxmap, numElementsGlobal)
+
+                @inbounds for (gi, gtag) in enumerate(elemTags_global)
+                    idxmap[gtag] = gi - 1
+                end
+
                 # buffers
                 nel = length(elemTags[itype])
                 nnet = zeros(Int, nel, numNodes)
@@ -1680,8 +1697,16 @@ function assemble_operator(
                 @inbounds for e in 1:nel
                     elem = elemTags[itype][e]
 
-                    jac, jacDet, _ = gmsh.model.mesh.getJacobian(elem, intPoints)
-                    Jac = reshape(jac, 3, :)
+                    gidx = idxmap[elem]
+
+                    jac_slice = @view jac_all[gidx * 9 * nip + 1 : (gidx + 1) * 9 * nip]
+
+                    jacDet = @view det_all[gidx * nip + 1 : (gidx + 1) * nip]
+
+                    Jac = reshape(jac_slice, 3, :)
+
+                    #jac, jacDet, _ = gmsh.model.mesh.getJacobian(elem, intPoints)
+                    #Jac = reshape(jac, 3, :)
 
                     @inbounds for k in 1:numIntPoints
                         @views invJac[1:3, 3k-2:3k] .= inv(Jac[1:3, 3k-2:3k])'
@@ -2095,23 +2120,52 @@ function assemble_linear(
 
                 ∇h = reshape(dfun, :, numIntPoints)
 
+                # Get Jacobians for all elements of this type
+                elemTags_global, _ = gmsh.model.mesh.getElementsByType(et)
+
+                jac_all, det_all, _ = gmsh.model.mesh.getJacobians(et, intPoints)
+
+                numElementsGlobal = length(elemTags_global)
+                nip = length(det_all) ÷ numElementsGlobal
+
+                @assert nip == numIntPoints
+
+                # Map element tag to its position in jac_all and det_all
+                idxmap = Dict{eltype(elemTags_global),Int}()
+                sizehint!(idxmap, numElementsGlobal)
+
+                @inbounds for (gi, gtag) in enumerate(elemTags_global)
+                    idxmap[gtag] = gi - 1
+                end
+
                 ndofs_loc = P.pdim * numNodes
                 B = zeros(outdim, ndofs_loc)
                 fe = zeros(ndofs_loc)
 
                 nel = length(elemTags[itype])
 
+                invJac = zeros(3, 3 * numIntPoints)
+                ∂h = zeros(P.dim, numNodes * numIntPoints)
+
                 for e in 1:nel
                     fill!(fe, 0.0)
 
                     elem = elemTags[itype][e]
 
-                    jac, jacDet, _ =
-                        gmsh.model.mesh.getJacobian(elem, intPoints)
+                    gidx = idxmap[elem]
 
-                    Jac = reshape(jac, 3, :)
-                    invJac = zeros(3, 3 * numIntPoints)
-                    ∂h = zeros(P.dim, numNodes * numIntPoints)
[…]
