### 2026-07-29 69f70e5
optimization of integrate function
