+++ src/general.jl
@@ -5721,71 +5721,107 @@ function integrate(problem::Problem, phName::String, f::Union{Function,ScalarFie
     f2 = f isa ScalarField && !is_elementwise ? f : nothing
     DIM = problem.dim
     b = problem.thickness
-    ncoord2 = zeros(3 * problem.non)
-    sum0 = 0
+    #ncoord2 = zeros(3 * problem.non)
+    sum0 = 0.0
     dimTags = gmsh.model.getEntitiesForPhysicalName(phName)
     for i ∈ 1:length(dimTags)
         dimTag = dimTags[i]
         dim = dimTag[1]
         tag = dimTag[2]
         elementTypes, elementTags, elemNodeTags = gmsh.model.mesh.getElements(dim, tag)
-        nodeTags::Vector{Int64}, ncoord, parametricCoord = gmsh.model.mesh.getNodes(dim, tag, true, false)
+        #nodeTags::Vector{Int64}, ncoord, parametricCoord = gmsh.model.mesh.getNodes(dim, tag, true, false)
         #ncoord2[nodeTags*3 .- 2] = ncoord[1:3:length(ncoord)]
         #ncoord2[nodeTags*3 .- 1] = ncoord[2:3:length(ncoord)]
         #ncoord2[nodeTags*3 .- 0] = ncoord[3:3:length(ncoord)]
-        @inbounds for (local_idx, nd) in enumerate(nodeTags)
-            ci = 3*(nd-1)
-            cn = 3*(local_idx-1)
-            ncoord2[ci+1] = ncoord[cn+1]
-            ncoord2[ci+2] = ncoord[cn+2]
-            ncoord2[ci+3] = ncoord[cn+3]
-        end
+        #@inbounds for (local_idx, nd) in enumerate(nodeTags)
+        #    ci = 3*(nd-1)
+        #    cn = 3*(local_idx-1)
+        #    ncoord2[ci+1] = ncoord[cn+1]
+        #    ncoord2[ci+2] = ncoord[cn+2]
+        #    ncoord2[ci+3] = ncoord[cn+3]
+        #end
 
         for ii in 1:length(elementTypes)
+            et = elementTypes[ii]
             elementName, dim, order1, numNodes::Int64, localNodeCoord, numPrimaryNodes = gmsh.model.mesh.getElementProperties(elementTypes[ii])
             qorder = order <= 0 ? order1 : order
-            intPoints, intWeights = gmsh.model.mesh.getIntegrationPoints(elementTypes[ii], "Gauss" * string(qorder))
+            intPoints, intWeights = gmsh.model.mesh.getIntegrationPoints(et, "Gauss" * string(qorder))
             numIntPoints = length(intWeights)
-            comp, fun, ori = gmsh.model.mesh.getBasisFunctions(elementTypes[ii], intPoints, "Lagrange")
+            comp, fun, ori = gmsh.model.mesh.getBasisFunctions(et, intPoints, "Lagrange")
             h = reshape(fun, :, numIntPoints)
-            nnet = zeros(Int, length(elementTags[ii]), numNodes)
+            #nnet = zeros(Int, length(elementTags[ii]), numNodes)
+            # Batched geometric data for all elements of this type
+            elemTagsGlobal, _ = gmsh.model.mesh.getElementsByType(et)
+
+            jacAll, detAll, coordAll = gmsh.model.mesh.getJacobians(et, intPoints)
+
+            numElementsGlobal = length(elemTagsGlobal)
+            nip = length(detAll) ÷ numElementsGlobal
+
+            @assert nip == numIntPoints
+
+            idxmap = Dict{eltype(elemTagsGlobal),Int}()
+            sizehint!(idxmap, numElementsGlobal)
+
[…]
