### 2026-07-29 56783be
update README.md
3rd, low level example for 2D linear elasticity
