### 2026-07-29 e9aff3c
optimization of compound operators
