+++ src/multifield.jl
@@ -1467,8 +1467,21 @@ function assemble_operator(
     coefficient::Union{Number,ScalarField,AbstractMatrix,AbstractVector}=1.0,
     weight=nothing,
     domain=nothing,
-    gauss=:full)
+    gauss=:full,
+    assembly::Symbol=:matrix,
+    I=nothing,
+    J=nothing,
+    V=nothing)
 
+    assembly ∈ (:matrix, :triplets, :add) ||
+    error("""
+    assemble_operator: invalid assembly mode $assembly.
+
+    Valid modes are:
+        :matrix
+        :triplets
+        :add
+    """)
     @assert Pu.name == Ps.name "Both problems must refer to the same gmsh model/mesh."
     @assert Pu.dim == Ps.dim "Both problems must have the same spatial dimension."
     gmsh.model.setCurrent(Pu.name)
@@ -1576,10 +1589,50 @@ function assemble_operator(
     # ------------------------------------------------------------
 
     # estimate IJV length: crude but OK for notebook prototype
-    lengthOfIJV = LowLevelFEM.estimateLengthOfIJV(Pu) * max(1, Ps.pdim) * max(1, Pu.pdim)
-    I = Vector{Int}(undef, lengthOfIJV)
-    J = Vector{Int}(undef, lengthOfIJV)
-    V = Vector{Float64}(undef, lengthOfIJV)
+    #lengthOfIJV = LowLevelFEM.estimateLengthOfIJV(Pu) * max(1, Ps.pdim) * max(1, Pu.pdim)
+    #I = Vector{Int}(undef, lengthOfIJV)
+    #J = Vector{Int}(undef, lengthOfIJV)
+    #V = Vector{Float64}(undef, lengthOfIJV)
+    #pos = 1
+    # ------------------------------------------------------------
+    # Assembly storage
+    # ------------------------------------------------------------
+
+    build_pattern = assembly !== :add
+
+    if build_pattern
+        I === nothing ||
+            error("assemble_operator: I must be nothing for assembly=$assembly.")
+
+        J === nothing ||
+            error("assemble_operator: J must be nothing for assembly=$assembly.")
+
+        V === nothing ||
+            error("assemble_operator: V must be nothing for assembly=$assembly.")
+
+        lengthOfIJV =
+            LowLevelFEM.estimateLengthOfIJV(Pu) *
+            max(1, Ps.pdim) *
+            max(1, Pu.pdim)
+
+        I = Vector{Int}(undef, lengthOfIJV)
+        J = Vector{Int}(undef, lengthOfIJV)
+        V = Vector{Float64}(undef, lengthOfIJV)
+
+    else
+        I === nothing ||
+            error("assemble_operator: I must be nothing for assembly=:add.")
+
+        J === nothing ||
+            error("assemble_operator: J must be nothing for assembly=:add.")
+
+        V isa Vector{Float64} ||
+            error(
+                "assemble_operator: assembly=:add requires " *
+                "V::Vector{Float64}; got $(typeof(V))."
+            )
+    end
+
     pos = 1
 
     dim = Pu.dim
@@ -1904,16 +1957,53 @@ function assemble_operator(
                             #@assert pos >= 1
                             #@assert pos - 1 <= length(I)
                             #@assert all(isfinite, V[1:pos-1])
-                            if pos >= length(I)
-                                newlen = Int(ceil(1.5*length(I))) + 1000
-                                resize!(I, newlen)
-                                resize!(J, newlen)
-                                resize!(V, newlen)
-                            end
+                            
+                            #if pos >= length(I)
+                            #    newlen = Int(ceil(1.5*length(I))) + 1000
+                            #    resize!(I, newlen)
+                            #    resize!(J, newlen)
+                            #    resize!(V, newlen)
[…]
