### 2026-07-30 2d60715
stabilize detect_pattern!
