### 2026-08-01 7f70dd9
Optimized compound bilinear assembly
