+++ src/multifield.jl
@@ -1100,7 +1100,7 @@ end
 # zero entries.  If a factor is not sparse enough, the code falls back to dense
 # BLAS-backed mul!.
 
-struct MatrixPattern
+mutable struct MatrixPattern
     rows::Vector{Int}
     cols::Vector{Int}
     nnz::Int
@@ -1146,6 +1146,48 @@ function detect_pattern(M::AbstractMatrix; tol::Real=0.0)
     end
 
     return MatrixPattern(rows, cols, length(rows), m, n)
+end
+
+"""
+    detect_pattern!(pattern, M; tol=0.0)
+
+Update a reusable matrix pattern from the current numerical values of `M`.
+"""
+function detect_pattern!(
+    pattern::MatrixPattern,
+    M::AbstractMatrix;
+    tol::Real = 0.0
+)
+    m, n = size(M)
+
+    pattern.m == m && pattern.n == n ||
+        error(
+            "detect_pattern!: matrix size changed from " *
+            "($(pattern.m), $(pattern.n)) to ($m, $n)."
+        )
+
+    empty!(pattern.rows)
+    empty!(pattern.cols)
+
+    @inbounds for j in 1:n
+        for i in 1:m
+            mij = M[i, j]
+
+            if tol == 0
+                if !iszero(mij)
+                    push!(pattern.rows, i)
+                    push!(pattern.cols, j)
+                end
+            elseif abs(mij) > tol
+                push!(pattern.rows, i)
+                push!(pattern.cols, j)
+            end
+        end
+    end
+
+    pattern.nnz = length(pattern.rows)
+
+    return pattern
 end
 
 @inline function sparse_enough(
@@ -1745,6 +1787,21 @@ function assemble_operator(
                 #patBsT = nothing
                 #patCgp = nothing
                 #patTmp = nothing
+                patBu  = MatrixPattern(size(Bu, 1), size(Bu, 2))
+                patBsT = MatrixPattern(size(Bs, 2), size(Bs, 1))
+
+                sizehint!(patBu.rows, length(Bu))
+                sizehint!(patBu.cols, length(Bu))
+
+                sizehint!(patBsT.rows, length(Bs))
+                sizehint!(patBsT.cols, length(Bs))
+
+                patTmp = MatrixPattern(size(tmp, 1), size(tmp, 2))
+
+                sizehint!(patTmp.rows, length(tmp))
+                sizehint!(patTmp.cols, length(tmp))
+
+                patCgp = nothing
 
                 # element loop
                 @inbounds for e in 1:nel
@@ -1852,17 +1909,29 @@ function assemble_operator(
                                 build_B!(Bs, op_s, Ps, k, h, ∂h, numNodes)
                             end
 
-                            #if patBu === nothing
-                            #    patBu = detect_pattern(Bu)
-                            #end
-                            #if patBsT === nothing
-                            #    patBsT = detect_pattern(transpose(Bs))
-                            #end
-                            mul_opt!(Ke, transpose(Bs), Bu;
-                                #patA=patBsT,
-                                #patB=patBu,
-                                alpha=w,
-                                beta=1.0
+                            ##if patBu === nothing
+                            ##    patBu = detect_pattern(Bu)
+                            ##end
+                            ##if patBsT === nothing
+                            ##    patBsT = detect_pattern(transpose(Bs))
+                            ##end
+                            #mul_opt!(Ke, transpose(Bs), Bu;
+                            #    #patA=patBsT,
+                            #    #patB=patBu,
+                            #    alpha=w,
+                            #    beta=1.0
+                            #)
+                            detect_pattern!(patBsT, transpose(Bs))
+                            detect_pattern!(patBu, Bu)
+
+                            mul_opt!(
+                                Ke,
+                                transpose(Bs),
+                                Bu;
[…]
