### 2026-08-03 f9b89a0
Revise citation format in README.md
Updated citation format for JOSS paper in README.
