### 2026-08-03 76516e4
Update citation format in index.md
