+++ src/multifield.jl
@@ -1415,6 +1415,73 @@ end
     return t1, t2
 end
 
+mutable struct OperatorWorkspace
+    invJac::Matrix{Float64}
+    ∂h::Matrix{Float64}
+    Bu::Matrix{Float64}
+    Bs::Matrix{Float64}
+    Ke::Matrix{Float64}
+    tmp::Matrix{Float64}
+
+    patBu::MatrixPattern
+    patBsT::MatrixPattern
+    patTmp::MatrixPattern
+    patCgp::Union{Nothing,MatrixPattern}
+end
+
+"""
+    OperatorWorkspace(
+        dim,
+        num_nodes,
+        num_integration_points,
+        out_u,
+        out_s,
+        ndofs_u,
+        ndofs_s
+    )
+
+Allocate reusable element-level work arrays for operator assembly.
+"""
+function OperatorWorkspace(
+    dim::Int,
+    num_nodes::Int,
+    num_integration_points::Int,
+    out_u::Int,
+    out_s::Int,
+    ndofs_u::Int,
+    ndofs_s::Int
+)
+    Bu = zeros(out_u, ndofs_u)
+    Bs = zeros(out_s, ndofs_s)
+    tmp = zeros(ndofs_s, out_u)
+
+    patBu  = MatrixPattern(out_u, ndofs_u)
+    patBsT = MatrixPattern(ndofs_s, out_s)
+    patTmp = MatrixPattern(ndofs_s, out_u)
+
+    sizehint!(patBu.rows, length(Bu))
+    sizehint!(patBu.cols, length(Bu))
+
+    sizehint!(patBsT.rows, length(Bs))
+    sizehint!(patBsT.cols, length(Bs))
+
+    sizehint!(patTmp.rows, length(tmp))
+    sizehint!(patTmp.cols, length(tmp))
+
+    return OperatorWorkspace(
+        zeros(3, 3 * num_integration_points),
+        zeros(dim, num_nodes * num_integration_points),
+        Bu,
+        Bs,
+        zeros(ndofs_s, ndofs_u),
+        tmp,
+        patBu,
+        patBsT,
+        patTmp,
+        nothing
+    )
+end
+
 """
     assemble_operator(Pu::Problem, 
                       op_u::AbstractOp, 
@@ -1514,6 +1581,8 @@ function assemble_operator(
     I=nothing,
     J=nothing,
     V=nothing)
+
+    LinearAlgebra.BLAS.set_num_threads(1)
 
     assembly ∈ (:matrix, :triplets, :add) ||
     error("""
@@ -1763,16 +1832,9 @@ function assemble_operator(
                 # buffers
                 nel = length(elemTags[itype])
                 nnet = zeros(Int, nel, numNodes)
-                invJac = zeros(3, 3numIntPoints)
-                ∂h = zeros(dim, numNodes * numIntPoints)
 
                 ndofs_u_loc = Pu.pdim * numNodes
                 ndofs_s_loc = Ps.pdim * numNodes
-
-                Bu = zeros(out_u, ndofs_u_loc)
-                Bs = zeros(out_s, ndofs_s_loc)
-                Ke = zeros(ndofs_s_loc, ndofs_u_loc)
-                tmp = zeros(size(Bs,2), size(Bu,1))
 
                 # connectivity table
                 @inbounds for e in 1:nel
@@ -1781,334 +1843,276 @@ function assemble_operator(
                     end
                 end
 
-                #tmpBu = similar(Bu)
-    
-                #patBu = nothing
-                #patBsT = nothing
-                #patCgp = nothing
-                #patTmp = nothing
-                patBu  = MatrixPattern(size(Bu, 1), size(Bu, 2))
[…]
