### 2026-08-07 fa94990
multi-thread for bilinear term is ready
