+++ examples/dev/multi-thread.jl
@@ -56,3 +56,5 @@ GC.gc()
 print("Error: ")
 err = norm(K1.A - K2.A) / norm(K1.A)
 println(100err, "%")
+println()
+println("LinAlg szálak száma: ", LinearAlgebra.BLAS.get_num_threads())
