+++ src/multifield.jl
@@ -1483,6 +1483,224 @@ function OperatorWorkspace(
 end
 
 """
+    assemble_element_matrix!(ws, elem, gidx, ...)
+
+Compute the local element matrix and store it in `ws.Ke`.
+"""
+function assemble_element_matrix!(
+    ws::OperatorWorkspace,
+    elem::Integer,
+    gidx::Integer,
+    jac_all,
+    det_all,
+    nip::Integer,
+    numIntPoints::Integer,
+    dim::Integer,
+    edim::Integer,
+    numNodes::Integer,
+    ∇h,
+    h,
+    intWeights,
+    Cprep,
+    Wprep,
+    op_u::AbstractOp,
+    Pu::Problem,
+    op_s::AbstractOp,
+    Ps::Problem
+)
+    invJac = ws.invJac
+    ∂h     = ws.∂h
+    Bu     = ws.Bu
+    Bs     = ws.Bs
+    Ke     = ws.Ke
+    tmp    = ws.tmp
+
+    jac_slice =
+        @view jac_all[gidx * 9 * nip + 1 : (gidx + 1) * 9 * nip]
+
+    jacDet =
+        @view det_all[gidx * nip + 1 : (gidx + 1) * nip]
+
+    Jac = reshape(jac_slice, 3, :)
+
+    @inbounds for k in 1:numIntPoints
+        @views invJac[1:3, 3k-2:3k] .=
+            inv(Jac[1:3, 3k-2:3k])'
+    end
+
+    # Physical / tangential gradients
+    fill!(∂h, 0.0)
+
+    @inbounds for k in 1:numIntPoints
+        if edim == dim
+            @views begin
+                invJk = invJac[1:dim, 3k-2:3k-(3-dim)]
+
+                for a in 1:numNodes
+                    gha = ∇h[a*3-2:a*3-(3-dim), k]
+                    dha = ∂h[1:dim, (k-1)*numNodes+a]
+                    mul!(dha, invJk, gha)
+                end
+            end
+        else
+            Jk = @view Jac[1:dim, 3k-2:3k]
+            Jtan = Matrix(Jk[:, 1:edim])
+            Gtan = Jtan * inv(Jtan' * Jtan)
+
+            for a in 1:numNodes
+                gha = ∇h[a*3-2:a*3-3+edim, k]
+                ∂h[1:dim, (k-1)*numNodes+a] .= Gtan * gha
+            end
+        end
+    end
+
+    fill!(Ke, 0.0)
+
+    @inbounds for k in 1:numIntPoints
+        if Cprep isa AbstractVector
+            Cgp = _eval_coefficient_chain_at_gp(
+                Cprep, elem, view(h, :, k)
+            )
+        else
+            Cgp = _eval_coefficient_at_gp(
+                Cprep, elem, view(h, :, k)
+            )
+        end
+
+        wcoef =
+            Wprep === nothing ?
+            1.0 :
+            _eval_coefficient_at_gp(Wprep, elem, view(h, :, k))
+
+        if Cgp isa Number
+            w = jacDet[k] * intWeights[k] * Cgp * wcoef
+
+            if op_u isa SurfaceSymGradOp
+                t1, t2 = surface_basis_from_J(Jac, k)
+                build_B!(Bu, op_u, Pu, k, h, ∂h, numNodes, t1, t2)
+            elseif op_u isa AxialGradOp
+                invJk = invJac[1:Pu.dim, 3k-2:3k-(3-Pu.dim)]
+                Jk = inv(invJk')
+                t = Jk[:, 1]
+                t = t / norm(t)
+                build_B!(Bu, op_u, Pu, k, h, ∂h, numNodes, t)
[…]
